### 2024-04-24 #638 @sigmafelix
`chopin`: Computation for Climate and Health research On Parallelized INfrastructure
Submitting Author Name: Insang Song
Submitting Author Github Handle: @sigmafelix
Repository:  https://github.com/NIEHS/chopin
Version submitted: 0.6.2.20240423
Submission type: Standard

Version accepted: TBD

---

```
Package: chopin
Title: CHOPIN: Computation for Climate and Health research On Parallelized INfrastructure
Version: 0.6.2.20240423
    person("Insang", "Song", , "geoissong@gmail.com", role = c("aut", "cre"),
    person("Kyle", "Messier", role = c("aut", "ctb"),
  )
Description: It enables users with basic understanding on geospatial data
    and sf and terra functions to parallelize geospatial operations for
    geospatial exposure assessment modeling and covariate computation.
    Parallelization is done by dividing large datasets into sub-regions
    with regular grids and data's own hierarchy. A computation over the
    large number of raster files can be parallelized with a chopin
    function as well.
LazyDataCompression: xz
    anticlust,
    exactextractr (>= 0.8.2),
    future.apply,
    igraph,
    rlang (>= 0.4.9),
    sf (>= 1.0-10),
    stars (>= 0.6-0),
    terra (>= 1.7-18)
    callr,
    covr,
    DiagrammeR,
    doFuture,
    doParallel,
    future,
    future.batchtools,
    future.callr,
    spatstat.random,
    units,
    withr
```

: `chopin` supports parallel processing for functions in popular spatial data manipulation packages `sf` and `terra` on a high-level parallelization framework `future`. This feature fits to the geospatial data category.

: Our first target audience group is spatial epidemiologists and health geographers who want to calculating spatial covariates from spatial and spatiotemporal datasets including climate, transportation, demographics, topography, hydrography, and others. We expect that users are cognizant of basic geographic information system/science. The wider audience could take advantage of the flexibility of this package for expediting spatial operations.

: A selection of functions in `terra` (e.g., `*app` and `predict`) supports internal parallelization, where a single dataset is accepted. To the best of our knowledge, no packaged solution exists for parallelization of spatial operations where two datasets are involved. [sprawl](https://github.com/lbusett/sprawl) (GitHub-only; not maintained) partially overlaps this package's functionality in that it includes convenience functions connecting multiple basic spatial operations. Besides the R functions, a handful of teaching or demonstration materials briefly covered parallelization of spatial data applications ([[1]](https://rpubs.com/Vinit_Sehgal/lgar21), [[2]](https://posit.co/blog/geospatial-distributed-processing-with-furrr/), [[3]](https://kadyb.github.io/stars-parallel/Tutorial.html)).

: Not applicable

: Presubmission inquiry of the previous version of this package -- #630 commented by @ldecicco-USGS

: Coverage rate (99%) and installation size (25.7 MB; of which data (2.0 MB) and extdata (23.1 MB) exceeded recommended sizes) resulted in notes.

Thank you very much for your consideration!
