### 2024-04-24 f8c6599
Reflecting ROpenSci initial check feedback - Added CONTRIBUTING.md - Missing @returns in clip_ras_ext - Updated NEWS.md
