+++ NEWS.md
@@ -1,3 +1,12 @@
+# 0.6.2
+- FIX: align input-output classes in `par_merge_grid`
+- IMPROVE: grid_advanced mode supports the maximum of merged unit grids
+- Added the balanced number mode of splitting input points (`par_group_balanced`; thanks to comments of Dr. Michael Fessler)
+- Added function of the balanced mode for `par_grid` (`par_group_grid`)
+- All internal `sapply` is changed to `vapply`
+- Added a diagram for guiding users' choice for `par_*` functions for parallelization considering raster/vector data situations
+- litr R Markdown file is moved to the archive
+
 # 0.5.0
 - `par_grid` unifies the classes of grid inputs
 - `par_cut_coords`: a supplementary function effectively operating with `par_def_q` for `mode = "grid_quantile"` in `par_make_gridset`. X- and Y-coordinates (of centroids for polygon inputs) are split into quantile groups to balance the number of features in each grid; does not account for geometric complexity (i.e., number of vertices)
