+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   v00_good_practice_parallelization: v00_good_practice_parallelization.html
   v01_par_make_gridset: v01_par_make_gridset.html
   v02_climate_examples: v02_climate_examples.html
-last_built: 2024-04-24T13:14Z
+last_built: 2024-04-24T15:12Z
 urls:
   reference: https://kyle-messier.github.io/chopin/reference
   article: https://kyle-messier.github.io/chopin/articles
