### 2024-04-24 e0ce974
Lint - Trailing whitespace - Lines with 80+ characters
