+++ R/gridding.R
@@ -398,8 +398,8 @@ par_cut_coords <- function(x = NULL, y = NULL, quantiles) {
 #' @param grid_in `sf` or `SpatVector` object.
 #' The grid generated by [`par_make_grid`].
 #' @param grid_min_features integer(1). Threshold to merge adjacent grids.
-#' @param merge_max integer(1). Maximum number of grids to merge per merged set.
-#' Default is 4.
+#' @param merge_max integer(1).
+#' Maximum number of grids to merge per merged set. Default is 4.
 #' For example, if the number of grids to merge is 20 and `merge_max` is 10,
 #' the function will split the 20 grids into two sets of 10 grids.
 #' @returns A `sf` or `SpatVector` object of computation grids.
@@ -515,7 +515,7 @@ par_merge_grid <-
     identified_graph_member <- identified_graph$membership
     identified_graph_member2 <- identified_graph_member
 
-    # for assigning merged grid id (original)  
+    # for assigning merged grid id (original)
     merge_idx <- which(rownames(grid_pc) %in% names(identified_graph_member))
 
     # nolint start
@@ -565,7 +565,8 @@ par_merge_grid <-
     # nolint end
     # 10. Assign membership information
     # Here we use the modified membership
-    merge_member <- split(rownames(grid_pc)[merge_idx], identified_graph_member)
+    merge_member <-
+      split(rownames(grid_pc)[merge_idx], identified_graph_member)
     # 11. Label the merged grids
     merge_member_label <-
       unlist(lapply(merge_member, function(x) paste(x, collapse = "_")))
@@ -617,7 +618,7 @@ par_merge_grid <-
 #' @description For balancing computational loads, the function uses
 #' the `anticlust` package to cluster the input points. The number of clusters
 #' is determined by the `num_cluster` argument. Each cluster will have
-#' equal number of points. Grids will be generated based on the cluster extents.
+#' equal number of points. Grids will be generated based on the cluster extents
 #' @note This function is only for two-dimensional points.
 #' The results will be inexhaustive grids.
 #' @param points_in `sf` or `SpatVector` object. Target points of computation.
