### 2024-05-06 #638 @beatrizmilz
Hello @sigmafelix !
Thank you for your submission. I am the editor who will be working with you on this package.

---

- 1. Seems like the package was created in the organization called `Spatiotemporal-Exposures-and-Toxicology` and then was transfered to `NIEHS`, and some of the links in the documentation are still pointing to the old organization. Can you update them to point to the new organization, please? This might help: <https://github.com/search?q=repo%3ANIEHS%2Fchopin%20Spatiotemporal-Exposures-and-Toxicology&type=code>

- 2. The vignette [Extracting Weather/Climate Geospatial Data with `chopin`](https://niehs.github.io/chopin/articles/v02_climate_examples.html) is well detailed and informative. However, the vignette [Generate computational grids](https://niehs.github.io/chopin/articles/v01_par_make_gridset.html) can be more detailed. It would be great if you could improve that vignette with more text explaining the code and the steps.

- 3. I ran `devtools::test()` and all the tests passed. However, I noticed that there is a lot of messages from dplyr joins, because aparently there are some joins that are being done without declaring which are the key columns. This is not a problem, but it would be nice if you could add the `by` argument to the `dplyr::**_join()` functions in the code. Here is part of the messages that I got:

```
✔ |         41 | processing [8.1s]
⠏ |          0 | tests                                                         Joining with `by = join_by(from_id)`
Joining with `by = join_by(to_id)`
Joining with `by = join_by(from_id, to_id)`
⠴ |          6 | tests                                                         Joining with `by = join_by(from_id)`
Joining with `by = join_by(to_id)`
Joining with `by = join_by(from_id, to_id)`
Joining with `by = join_by(from_id)`
Joining with `by = join_by(to_id)`
Joining with `by = join_by(from_id, to_id)`
⠙ |         12 | tests                                                         Joining with `by = join_by(from_id)`
Joining with `by = join_by(to_id, FIPS)`
Joining with `by = join_by(from_id, to_id)`
✔ |         13 | tests
```

- 4. The package has >99% of code covered by tests. 🎉
![Screenshot 2024-05-06 at 11-54-27 Screenshot](https://github.com/ropensci/software-review/assets/42153618/3b586f78-e1f6-4f83-a9e2-4e59f4a4f0ab)

Can you please address these comments?
