+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -10,6 +10,8 @@ assignees: ''
 ### Error description
 
 ### Reproducible code
+- Please use `reprex::reprex()` to generate a reproducible example for error reporting
+- If possible, please share the data we can use to reproduce the errors. Using a file link would be a great option.
 - Should include the data you use (please be aware of seed number if you generated random numbers)
 
 ### Expected behavior
