+++ tools/vignettes-sources/v01_par_make_gridset_orig.Rmd
@@ -22,11 +22,10 @@ knitr::opts_knit$set(
 )
 
 knitr::opts_chunk$set(warning = FALSE, message = FALSE, out.width = "900px", fig.asp = 0.56)
-
+knitr::fig_path(".png", options = list(fig.path = "figures/v01-"))
 ```
 
 
-# Prepare input data
 ```{r}
 library(chopin)
 library(dplyr)
@@ -35,23 +34,65 @@ library(terra)
 options(sf_use_s2 = FALSE)
 ```
 
-```{r nc-load}
+
+
+
+
+## Computational grids
+- Computational grids refer to a set of polygons that cover the entire spatial domain of interest. These grids are used to split the spatial domain into smaller pieces for parallel processing.
+- `chopin` provides functions to generate computational grids for parallel processing by `par_grid()` function.
+- The standard sequence of running `par_grid()` includes the following steps:
+    1. Generate computational grids with `par_make_gridset()` or `par_group_balanced()`-`par_group_grid()` sequence. This step will give you a set of grid polygons, one of which is original splits and the other is padded in consideration of buffer radius in the subsequent spatial operations.
+    2. Run `par_grid()` with the input dataset and the grid polygons.
+
+## Types of computational grids and their generation
+
+- There are two approaches to generate computational grids. One is to use `par_make_gridset()` with one of three `mode`s and the other is to use `par_group_balanced()` with `par_group_grid()`. Thus, users have **four options** in total to generate computational grids.
+
+### `par_make_gridset()`: standard interface
+- `par_make_gridset()` generates regular grid polygons with padding for parallel processing. Padding is the distance of overlapping between grid polygons, essentially from the buffer radius of the points when buffer polygons are concerned.
+- `par_make_gridset()` supports three internal `mode`s:
+    - `mode = "grid"`: generates regular grid polygons with padding, `nx` and `ny` arguments determine the number of columns and rows in the grid, respectively.
+    - `mode = "grid_quantile"`: generates regular grid polygons with padding based on quantiles of the number of points in each grid. The grids will look irregular and the points per grid will be more balanced than the `grid` mode.
+    - `mode = "grid_advanced"`: generates regular grid polygons with padding based on the number of points in each grid and the number of points in the entire dataset. `nx` and `ny` arguments determine the number of columns and rows in the grid, then `merge_max` argument controls how many adjacent grids are merged into one grid. `grid_min_features` argument determines the minimum number of points in each grid, which means grids with fewer points than this value will be merged with adjacent grids. Adjusting these arguments can balance the computational load among the threads and reduce the overhead of parallelization.
+
+
+### `par_group_balanced()` and `par_group_grid()`: focusing on getting the balanced clusters
+- Expanding the exhaustive grid generation, `par_group_balanced()` generates rectangles that cover the same number of points per grid. Users can use the output of this function into `par_grid()` for parallel processing by converting it to the grid sets with `par_group_grid()`.
+
+
+
+### Random points in NC
[…]
