### 2024-05-15 db83aab
group_grid + vignette update - Illustrated vignette on generating gridset reflecting group_grid function development - Tests for par_group_grid - Bug fix: merge_max argument passing in par_make_gridset - Version up (0.6.3)
