+++ tools/vignettes-sources/v01_par_make_gridset.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Generate computational grids"
-date: "2024-05-14"
+date: "2024-05-15"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Generate computational grids}
@@ -18,6 +18,8 @@ library(chopin)
 library(dplyr)
 library(sf)
 library(terra)
+library(anticlust)
+library(igraph)
 options(sf_use_s2 = FALSE)
 ```
 
@@ -29,12 +31,12 @@ options(sf_use_s2 = FALSE)
 - Computational grids refer to a set of polygons that cover the entire spatial domain of interest. These grids are used to split the spatial domain into smaller pieces for parallel processing.
 - `chopin` provides functions to generate computational grids for parallel processing by `par_grid()` function.
 - The standard sequence of running `par_grid()` includes the following steps:
-    1. Generate computational grids with `par_make_gridset()` or `par_group_balanced()`-`par_group_grid()` sequence. This step will give you a set of grid polygons, one of which is original splits and the other is padded in consideration of buffer radius in the subsequent spatial operations.
+    1. Generate computational grids with `par_make_gridset()` or `par_group_grid()`. This step will give you a set of grid polygons, one of which is original splits and the other is padded in consideration of buffer radius in the subsequent spatial operations.
     2. Run `par_grid()` with the input dataset and the grid polygons.
 
 ## Types of computational grids and their generation
 
-- There are two approaches to generate computational grids. One is to use `par_make_gridset()` with one of three `mode`s and the other is to use `par_group_balanced()` with `par_group_grid()`. Thus, users have **four options** in total to generate computational grids.
+- There are two approaches to generate computational grids. One is to use `par_make_gridset()` with one of three `mode`s and the other is to use `par_group_grid()`. Thus, users have **four options** in total to generate computational grids.
 
 ### `par_make_gridset()`: standard interface
 - `par_make_gridset()` generates regular grid polygons with padding for parallel processing. Padding is the distance of overlapping between grid polygons, essentially from the buffer radius of the points when buffer polygons are concerned.
@@ -44,8 +46,8 @@ options(sf_use_s2 = FALSE)
     - `mode = "grid_advanced"`: generates regular grid polygons with padding based on the number of points in each grid and the number of points in the entire dataset. `nx` and `ny` arguments determine the number of columns and rows in the grid, then `merge_max` argument controls how many adjacent grids are merged into one grid. `grid_min_features` argument determines the minimum number of points in each grid, which means grids with fewer points than this value will be merged with adjacent grids. Adjusting these arguments can balance the computational load among the threads and reduce the overhead of parallelization.
 
 
-### `par_group_balanced()` and `par_group_grid()`: focusing on getting the balanced clusters
-- Expanding the exhaustive grid generation, `par_group_balanced()` generates rectangles that cover the same number of points per grid. Users can use the output of this function into `par_grid()` for parallel processing by converting it to the grid sets with `par_group_grid()`.
[…]
