+++ tests/testthat/test-par_fallback.R
@@ -29,5 +29,3 @@ testthat::test_that(
     )
   }
 )
-
-
