+++ tests/testthat/test-distribute_suites.R
@@ -427,4 +427,3 @@ testthat::test_that(
     )
   }
 )
-
