### 2024-05-16 1fd28fd
par_group_balanced example fix
