+++ man/par_group_balanced.Rd
@@ -32,7 +32,9 @@ ncp <- terra::vect(
   ncpoints, geom = c("X", "Y"),
   keepgeom = FALSE, crs = "EPSG:5070"
 )
-par_group_balanced(ncp, 10)
+# 2,304 points / 12 = 192 points per cluster
+ncpbal <- par_group_balanced(ncp, 12)
+ncpbal
 }
 \author{
 Insang Song
