+++ tools/vignettes-sources/v01_par_make_gridset.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Generate computational grids"
-date: "2024-05-15"
+date: "2024-05-16"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Generate computational grids}
@@ -13,7 +13,7 @@ vignette: >
 
 
 
-```r
+``` r
 library(chopin)
 library(dplyr)
 library(sf)
@@ -54,7 +54,7 @@ options(sf_use_s2 = FALSE)
 ### Random points in NC
 - For demonstration of `par_make_gridset()`, we use moderately clustered point locations generated inside the counties of North Carolina.
 
-```r
+``` r
 ncpoly <- system.file("shape/nc.shp", package = "sf")
 ncsf <- sf::read_sf(ncpoly)
 ncsf <- sf::st_transform(ncsf, "EPSG:5070")
@@ -66,7 +66,7 @@ plot(sf::st_geometry(ncsf))
 <p class="caption">plot of chunk nc-gen-points</p>
 </div>
 
-```r
+``` r
 set.seed(202404)
 library(spatstat.random)
 ncpoints <-
@@ -87,7 +87,7 @@ plot(sf::st_geometry(ncpoints))
 <p class="caption">plot of chunk nc-gen-points</p>
 </div>
 
-```r
+``` r
 # convert to terra SpatVector
 ncpoints_tr <- chopin::dep_switch(ncpoints)
 ```
@@ -100,7 +100,7 @@ ncpoints_tr <- chopin::dep_switch(ncpoints)
 ## Visualize computational grids
 
 
-```r
+``` r
 ncpoints_tr <- terra::vect(ncpoints)
 compregions <-
   chopin::par_make_gridset(
@@ -119,7 +119,7 @@ class(compregions)
 ## [1] "list"
 ```
 
-```r
+``` r
 # length of 2
 names(compregions)
 ```
@@ -128,7 +128,7 @@ names(compregions)
 ## [1] "original" "padded"
 ```
 
-```r
+``` r
 par(mfrow = c(2, 1))
 plot(compregions$original, main = "Original grids")
 plot(compregions$padded, main = "Padded grids")
@@ -142,7 +142,7 @@ plot(compregions$padded, main = "Padded grids")
 ### Generate regular grid computational regions
 - `chopin::par_make_gridset()` takes a spatial dataset to generate regular grid polygons with `nx` and `ny` arguments with padding. Users will have both overlapping (by the degree of `radius`) and non-overlapping grids, both of which will be utilized to split locations and target datasets into sub-datasets for efficient processing.
 
-```r
+``` r
 compregions <-
   chopin::par_make_gridset(
     ncpoints_tr,
@@ -156,7 +156,7 @@ compregions <-
 - The output of `par_make_gridset()` is a list object with two elements named `original` (non-overlapping grid polygons) and `padded` (overlapping by `padding`). The class of each element depends on the input dataset class. The figures below illustrate the grid polygons with and without overlaps.
 
 
-```r
+``` r
 names(compregions)
 ```
 
@@ -164,7 +164,7 @@ names(compregions)
 ## [1] "original" "padded"
 ```
 
-```r
+``` r
 oldpar <- par()
 par(mfrow = c(2, 1))
 terra::plot(compregions$original, main = "Original grids")
@@ -176,7 +176,7 @@ terra::plot(compregions$padded, main = "Padded grids")
 <p class="caption">plot of chunk compare-compregions</p>
 </div>
 
-```r
+``` r
 par(mfrow = c(1, 1))
 terra::plot(compregions$original, main = "Original grids")
[…]
