### 2024-05-22 4a5cbce
0.6.4 dev - character input support for par_make_grid and its functionality in par_grid - Leveraging proxy argument in terra::vect, converting it back and forth to sf class object to ensure the future compatibility - Rebuilt README.md - TODO: deprecate par_fallback - TODO: check edge cases points exactly on the grid boundary - TODO: streamline unit grid export
