+++ R/processing.R
@@ -181,6 +181,7 @@ clip_ras_ext <- function(
 #'   as the `points`.
 #' @param max_cells integer(1). Maximum number of cells in memory.
 #' See [`exactextractr::exact_extract`] for more details.
+#' @param ... Placeholder.
 #' @returns a data.frame object with mean value
 #' @author Insang Song \email{geoissong@@gmail.com}
 #' @examples
@@ -212,7 +213,8 @@ extract_at_buffer <- function(
   kernel = NULL,
   bandwidth = NULL,
   extent = NULL,
-  max_cells = 2e7
+  max_cells = 2e7,
+  ...
 ) {
   # type check
   if (!any(
@@ -230,9 +232,6 @@ extract_at_buffer <- function(
       stop("Check class of the input raster.\n")
     }
   }
-  if (dep_check(points) == "sf") {
-    points <- dep_switch(points)
-  }
   if (!is.numeric(radius)) {
     stop("Check class of the input radius.\n")
   }
@@ -244,6 +243,10 @@ extract_at_buffer <- function(
   }
   if (is.character(points)) {
     points <- try(terra::vect(points, extent = extent))
+  } else {
+    if (dep_check(points) == "sf") {
+      points <- dep_switch(points)
+    }
   }
 
   if (!is.null(kernel)) {
@@ -288,7 +291,8 @@ extract_at_buffer_flat <- function(
   func = "mean",
   kernel = NULL,
   bandwidth = NULL,
-  max_cells = 2e7
+  max_cells = 2e7,
+  ...
 ) {
   # generate buffers
   bufs <- terra::buffer(points, width = radius, quadsegs = qsegs)
@@ -329,7 +333,8 @@ extract_at_buffer_kernel <- function(
   func = stats::weighted.mean,
   kernel = NULL,
   bandwidth = NULL,
-  max_cells = 2e7
+  max_cells = 2e7,
+  ...
 ) {
   # generate buffers
   bufs <- terra::buffer(points, width = radius, quadsegs = qsegs)
@@ -426,6 +431,7 @@ extract_at_buffer_kernel <- function(
 #'  For example, `"mean"` or `\(x, w) weighted.mean(x, w, na.rm = TRUE)`
 #' @param max_cells integer(1). Maximum number of cells in memory.
 #' See [`exactextractr::exact_extract`] for more details.
+#' @param ... Placeholder.
 #' @returns a data.frame object with function value
 #' @author Insang Song \email{geoissong@@gmail.com}
 #' @examples
@@ -450,7 +456,8 @@ extract_at_poly <- function(
   surf = NULL,
   id = NULL,
   func = "mean",
-  max_cells = 2e7
+  max_cells = 2e7,
+  ...
 ) {
   # type check
   if (!methods::is(polys, "SpatVector")) {
@@ -503,7 +510,7 @@ extract_at_poly <- function(
 #' @param mode one of `"polygon"`
 #'  (generic polygons to extract raster values with) or
 #'  `"buffer"` (point with buffer radius)
-#' @param ... various. Passed to extract_at_buffer.
+#' @param ... Placeholder.
 #'  See \code{?extract_at_buffer} for details.
 #' @returns A data.frame object with summarized raster values with
 #'  respect to the mode (polygon or buffer) and the function.
@@ -733,18 +740,24 @@ The result may not be accurate.\n",
 
 #' Area weighted summary using two polygon sf or SpatVector objects
 #' @family Macros for calculation
-#' @param poly_in A sf/SpatVector object at weighted means will be calculated.
-#' @param poly_weight A sf/SpatVector object from
+#' @param poly_in A sf/SpatVector object or file path of polygons detectable
+#'   with GDAL driver at weighted means will be calculated.
[…]
