### 2024-05-23 a3a9db1
0.6.4 dev (05232024) - par_fallback returns with refactoring - bug fix: grid_target_id did not work as expected -> now it selects using grid_target_id well - par_* functions use par_fallback and append data.frame with error messages - par_grid, par_hierarchy, and par_multirasters: combine argument is dropped. They will always use collapse::rowbind.
