+++ tests/testthat/test-par_fallback.R
@@ -1,4 +1,3 @@
-# Generated from chopin_rmarkdown_litr.rmd: do not edit by hand
 
 testthat::test_that(
   "par_fallback works",
