+++ R/scale_process.R
@@ -3,31 +3,39 @@
 #' @param err Error status or message.
 #' @param fun function.
 #' @param debug logical(1). Print error messages (`TRUE`) or not (`FALSE`)
-#' @returns data.frame with one column
-#' @note This function assumes that the `fun` has an argument named
+#' @param inputid character(1). ID of the computational region.
+#'   For example, `par_make_gridset` output should have `"CGRIDID"`,
+#'   which will be included in the outcome data.frame.
+#'   If the function does not have an argument containing `"id"`,
+#'   the output will have a column named `"chopin_domain_id"`.
+#' @returns data.frame with domain id and error message.
+#' @note This function assumes that the `fun` has an argument containing
 #' `"id"`.
 #' @author Insang Song
 #' @examples
 #' err <- simpleError("No input.")
-#' par_fallback(err, extract_at, debug = TRUE)
+#' par_fallback(err, extract_at, debug = TRUE, 1)
 #' @export
 par_fallback <-
   function(
     err = NULL,
     fun = NULL,
-    debug = FALSE
+    debug = FALSE,
+    inputid = NULL
   ) {
     if (debug) {
       print(err)
     }
-    fallback <- data.frame(ID = NA)
+    fallback <- matrix(NA, nrow = length(inputid), ncol = 1)
+    fallback <- as.data.frame(fallback)
     fun_args <- formals(fun)
-    indx <- grepl("id", names(fun_args))
-    if (any(indx)) {
-      detected_id <- fun_args[indx]
-    } else {
-      detected_id <- "id"
+    indx <- grep("id", names(fun_args))
+    detected_id <- "CGRIDID"
+    if (length(indx) != 0) {
+      detected_id <- "chopin_domain_id"
     }
+    fallback$error_message <- paste(unlist(err), collapse = " ")
+    fallback[, 1] <- inputid
     colnames(fallback)[1] <- detected_id
     return(fallback)
   }
@@ -69,8 +77,6 @@ par_fallback <-
 #'  \code{c(unique(grid_id)[id_from], unique(grid_id)[id_to])}
 #' @param debug logical(1). Prints error messages
 #' if there were any errors during the calculation.
-#' @param combine function. A function to combine results.
-#' Default is \code{dplyr::bind_rows}.
 #' @param fun_dist `sf`, `terra` or `chopin` functions.
 #' @param ... Arguments passed to the argument \code{fun_dist}.
 #' The **second** place should get a vector or raster dataset from which
@@ -138,7 +144,6 @@ par_grid <-
     grids,
     grid_target_id = NULL,
     debug = FALSE,
-    combine = collapse::rowbind,
     fun_dist,
     ...
   ) {
@@ -159,30 +164,33 @@ par_grid <-
     }
 
     # grid id selection check
-    if (is.character(grid_target_id) && !grepl(":", grid_target_id)) {
-      stop("Character grid_target_id should be in a form of 'startid:endid'.\n")
-    }
+    grid_target_ids <- unlist(grids$original[["CGRIDID"]])
     if (is.numeric(grid_target_id)) {
       if (length(grid_target_id) != 2) {
         stop(
           "Numeric grid_target_id should be in a form of c(startid, endid).\n"
         )
       }
-      grid_target_ids <- unlist(grids$original[["CGRIDID"]])[grid_target_id]
-    }
-    # subset using grids and grid_id
-    if (is.null(grid_target_id)) {
-      grid_target_ids <- unlist(grids$original[["CGRIDID"]])
+      grid_target_ids <- seq(grid_target_id[1], grid_target_id[2], by = 1)
[…]
