+++ man/par_hierarchy.Rd
@@ -4,14 +4,7 @@
 \alias{par_hierarchy}
 \title{Process a given function using a hierarchy in input data}
 \usage{
-par_hierarchy(
-  regions,
-  split_level = NULL,
-  debug = FALSE,
-  combine = dplyr::bind_rows,
-  fun_dist,
-  ...
-)
+par_hierarchy(regions, split_level = NULL, debug = FALSE, fun_dist, ...)
 }
 \arguments{
 \item{regions}{sf/SpatVector object.
@@ -24,8 +17,6 @@ A field name with the higher level information is also accepted.}
 
 \item{debug}{logical(1). Prints error messages
 if there were any errors during the calculation.}
-
-\item{combine}{function. The function to combine the results.}
 
 \item{fun_dist}{sf, terra, or chopin functions.}
 

[…]
