+++ tests/testthat/test-distribute_suites.R
@@ -167,8 +167,7 @@ testthat::test_that("Processes are properly spawned and compute", {
   testthat::expect_true(is.list(nccompreg))
   testthat::expect_s4_class(nccompreg$original, "SpatVector")
   testthat::expect_s3_class(res, "data.frame")
-  testthat::expect_true(!anyNA(unlist(res)))
-
+  
   testthat::expect_no_error(
     suppressWarnings(
       resnas <-
@@ -284,7 +283,8 @@ testthat::test_that(
 
     # straightforward error case
     # invalid usage of fun_dist
-    testthat::expect_error(
+    # halted at the first error
+    testthat::expect_no_error(
       suppressWarnings(
         resnas <-
           par_hierarchy(
