### 2024-05-23 5f0be52
0.6.4 - FIX: par_grid_merge: length mismatch - Zero intersection grids will be removed when further processing is called (e.g., par_grid_merge) or it is passed to par_* functions (i.e., par_grid) - Rebuilt vignettes
