+++ vignettes/v01_par_make_gridset.Rmd
@@ -1,12 +1,17 @@
 ---
 title: "Generate computational grids"
-date: "2024-05-16"
+date: "2024-05-23"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Generate computational grids}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
+
+
+
+
+
 
 ``` r
 library(chopin)
@@ -17,6 +22,10 @@ library(anticlust)
 library(igraph)
 options(sf_use_s2 = FALSE)
 ```
+
+
+
+
 
 ## Computational grids
 - Computational grids refer to a set of polygons that cover the entire spatial domain of interest. These grids are used to split the spatial domain into smaller pieces for parallel processing.
@@ -53,7 +62,7 @@ plot(sf::st_geometry(ncsf))
 ```
 
 <div class="figure">
-<img src="figures/nc-gen-points-1.png" alt="plot of chunk nc-gen-points" width="600px" />
+<img src="vignettes/figures/nc-gen-points-1.png" alt="plot of chunk nc-gen-points" width="600px" />
 <p class="caption">plot of chunk nc-gen-points</p>
 </div>
 
@@ -69,24 +78,30 @@ ncpoints <-
     kappa = 1.25e-9
   )
 ncpoints <- sf::st_as_sf(ncpoints)
+ncpoints <- sf::st_set_crs(ncpoints, "EPSG:5070")
 ncpoints$pid <- sprintf("PID-%05d", seq(1, nrow(ncpoints)))
 plot(sf::st_geometry(ncpoints))
 ```
 
-![](figures/nc-gen-points-2.png)
+<div class="figure">
+<img src="vignettes/figures/nc-gen-points-2.png" alt="plot of chunk nc-gen-points" width="600px" />
+<p class="caption">plot of chunk nc-gen-points</p>
+</div>
 
 ``` r
 # convert to terra SpatVector
 ncpoints_tr <- chopin::dep_switch(ncpoints)
 ```
 
+
 <!--![](https://i.imgur.com/ImPfGXP.png) -->
+
+
 
 ## Visualize computational grids
 
 
 ``` r
-ncpoints_tr <- terra::vect(ncpoints)
 compregions <-
   chopin::par_make_gridset(
     ncpoints_tr,
@@ -119,7 +134,10 @@ plot(compregions$original, main = "Original grids")
 plot(compregions$padded, main = "Padded grids")
 ```
 
-![](figures/grid-plain-1.png)
+<div class="figure">
+<img src="vignettes/figures/grid-plain-1.png" alt="plot of chunk grid-plain" width="600px" />
+<p class="caption">plot of chunk grid-plain</p>
+</div>
 
 ### Generate regular grid computational regions
 - `chopin::par_make_gridset()` takes a spatial dataset to generate regular grid polygons with `nx` and `ny` arguments with padding. Users will have both overlapping (by the degree of `radius`) and non-overlapping grids, both of which will be utilized to split locations and target datasets into sub-datasets for efficient processing.
@@ -153,7 +171,10 @@ terra::plot(compregions$original, main = "Original grids")
 terra::plot(compregions$padded, main = "Padded grids")
 ```
 
-![](figures/compare-compregions-1.png)
+<div class="figure">
+<img src="vignettes/figures/compare-compregions-1.png" alt="plot of chunk compare-compregions" width="600px" />
+<p class="caption">plot of chunk compare-compregions</p>
+</div>
 
 ``` r
 par(mfrow = c(1, 1))
@@ -161,7 +182,10 @@ terra::plot(compregions$original, main = "Original grids")
 terra::plot(ncpoints_tr, add = TRUE, col = "red", cex = 0.4)
 ```
 
[…]
