+++ R/scale_process.R
@@ -2,7 +2,6 @@
 #' @family Parallelization
 #' @param err Error status or message.
 #' @param fun function.
-#' @param debug logical(1). Print error messages (`TRUE`) or not (`FALSE`)
 #' @param inputid character(1). ID of the computational region.
 #'   For example, `par_make_gridset` output should have `"CGRIDID"`,
 #'   which will be included in the outcome data.frame.
@@ -20,12 +19,8 @@ par_fallback <-
   function(
     err = NULL,
     fun = NULL,
-    debug = FALSE,
     inputid = NULL
   ) {
-    if (debug) {
-      print(err)
-    }
     fallback <- matrix(NA, nrow = length(inputid), ncol = 1)
     fallback <- as.data.frame(fallback)
     fun_args <- formals(fun)
@@ -75,8 +70,9 @@ par_fallback <-
 #'  Default is NULL. If NULL, all grid_ids are used.
 #'  \code{"id_from:id_to"} format or
 #'  \code{c(unique(grid_id)[id_from], unique(grid_id)[id_to])}
-#' @param debug logical(1). Prints error messages
-#' if there were any errors during the calculation.
+#' @param debug logical(1). Default is `FALSE`. Otherwise,
+#'   if a unit computation fails, the error message and the `CGRIDID`
+#'   value where the error occurred will be included in the output.
 #' @param fun_dist `sf`, `terra` or `chopin` functions.
 #' @param ... Arguments passed to the argument \code{fun_dist}.
 #' The **second** place should get a vector or raster dataset from which
@@ -172,27 +168,22 @@ par_grid <-
         )
       }
       grid_target_ids <- seq(grid_target_id[1], grid_target_id[2], by = 1)
-      #grid_target_ids <- unlist(grids$original[["CGRIDID"]])[grid_target_id]
     }
     if (is.character(grid_target_id)) {
       # subset using grids and grid_id
       if (grepl(":", grid_target_id)) {
         grid_id_parsed <- strsplit(grid_target_id, ":", fixed = TRUE)[[1]]
         grid_id_parsed <- as.numeric(grid_id_parsed)
-        # cgridids <- unlist(grids$original[["CGRIDID"]])
         grid_target_ids <-
           seq(grid_id_parsed[1], grid_id_parsed[2], by = 1)
-          # c(which(cgridids == grid_id_parsed[1]),
-          #   which(cgridids == grid_id_parsed[2]))
       } else {
         stop("grid_target_id should be formed 'startid:endid'.\n")
       }
     }
-    # cgridids <- unlist(grids$original[["CGRIDID"]])
-    grids_target <-
+    grids_target_in <-
       grids$original[grid_target_ids, ]
     grids_target_list <-
-      base::split(grids_target, unlist(grids_target[["CGRIDID"]]))
+      base::split(grids_target_in, unlist(grids_target_in[["CGRIDID"]]))
 
     results_distributed <-
       future.apply::future_lapply(
@@ -239,7 +230,11 @@ par_grid <-
             return(res)
           },
           error = function(e) {
-            par_fallback(e, fun_dist, debug, inputid = grid$CGRIDID)
+            if (debug) {
+              par_fallback(e, fun_dist, inputid = grid$CGRIDID)
+            } else {
+              return(NULL)
+            }
           })
 
           return(run_result)
@@ -286,12 +281,18 @@ par_grid <-
 #' are spatially sparsely distributed.
 #' @param regions sf/SpatVector object.
 #'  Computational regions. Only polygons are accepted.
-#' @param split_level character(nrow(regions)) or character(1).
+#' @param regions_id character(nrow(regions)) or character(1).
 #'  The regions will be split by the common level value.
 #'  The level should be higher than the original data level.
[…]
