+++ tests/testthat/test-par_fallback.R
@@ -16,7 +16,6 @@ testthat::test_that(
       par_fallback(
         err = foo1(dc, rdd, id = "site_id", mode = "buffer", radius = 5e4),
         fun = foo1,
-        debug = TRUE,
         1
       )
     )
@@ -24,7 +23,6 @@ testthat::test_that(
       par_fallback(
         err = foo1(id = "site"),
         fun = foo2,
-        debug = FALSE,
         1
       )
     )
