+++ tests/testthat/test-gridding.R
@@ -5,7 +5,7 @@ testthat::test_that("Balanced group tests", {
   withr::local_package("anticlust")
   withr::local_options(list(sf_use_s2 = FALSE))
 
-  rv <- terra::vect(matrix(rnorm(1000, 1e3, 350), ncol = 2))
+  rv <- terra::vect(matrix(rnorm(1000, 40, 10), ncol = 2), crs = "EPSG:4326")
   rs <- sf::st_as_sf(rv)
 
   testthat::expect_no_error(
@@ -217,7 +217,7 @@ testthat::test_that("Grid merge is well done.", {
                      mode = "grid",
                      nx = 20L, ny = 12L,
                      padding = 1e4L)
-  testthat::expect_message(
+  testthat::expect_no_error(
     gridmerged2 <- par_merge_grid(ncptr2, griddedtr2$original, 15L)
   )
   testthat::expect_s4_class(gridmerged2, "SpatVector")
