+++ tools/vignettes-sources/v02_climate_examples.Rmd
@@ -676,7 +676,7 @@ system.time(
   exhierarchy <-
     chopin::par_hierarchy(
       bgsf_main,
-      split_level = "STATEFP",
+      regions_id = "STATEFP",
       fun_dist = chopin::extract_at_poly,
       polys = bgsf_main,
       surf = bilssds,
