### 2024-05-27 #638 @beatrizmilz
Hi @sigmafelix !
There was one error on the R CMD check: https://github.com/ropensci/software-review/issues/638#issuecomment-2125717203 .
Could you please verify that, before we proceed to add the reviewers?

Thank you!
