### 2024-05-28 fa30c6e
0.7.0 dev - repetitive type check is separated to check_subject (internal) function -- reduced the codebase volume - Incorrect application of vapply FUN.VALUE fixed
