+++ R/check.R
@@ -398,3 +398,54 @@ any_class_args <- function(
   args_scanned <- vapply(args_scanned, FUN = any, FUN.VALUE = logical(1))
   return(args_scanned)
 }
+
+
+
+#' Check the subject object and perform necessary conversions if needed.
+#' @description
+#' This function checks the class of the input object and
+#'   performs necessary conversions if needed.
+#' @keywords internal
+#' @param subject The input object to be checked.
+#' @param extent The extent of the subject object.
+#' @param subject_id Optional. The ID of the subject object.
+#' @returns The checked and converted subject object.
+#' @examples
+#' # Check a SpatVector object
+#' ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
+#' ncsf <- sf::st_read(ncpath)
+#' check_subject(subject = ncsf, extent = extent, subject_id = "FIPS")
+#'
+#' # Check a character object
+#' check_subject(subject = ncpath, extent = extent, subject_id = "FIPS")
+#' @export
+check_subject <- function(
+  subject,
+  extent = NULL,
+  subject_id = NULL
+) {
+  # type check
+  if (!any(
+    vapply(
+      c("SpatVector", "sf", "character"),
+      FUN = methods::is,
+      FUN.VALUE = logical(1),
+      object = subject
+    )
+  )) {
+    stop("Check class of the input object.\n")
+  }
+  if (is.character(subject)) {
+    subject <- try(terra::vect(subject, extent = extent))
+  } else {
+    if (dep_check(subject) == "sf") {
+      subject <- dep_switch(subject)
+    }
+  }
+  if (!is.null(subject_id)) {
+    if (!subject_id %in% names(subject)) {
+      stop("id should exist in the input object\n")
+    }
+  }
+  return(subject)
+}
