### 2024-05-28 bcb1f55
Documentation & example fix - summarize_sedc - check_subject
