+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   v00_good_practice_parallelization: v00_good_practice_parallelization.html
   v01_par_make_gridset: v01_par_make_gridset.html
   v02_climate_examples: v02_climate_examples.html
-last_built: 2024-05-16T17:37Z
+last_built: 2024-05-28T20:30Z
 urls:
   reference: https://kyle-messier.github.io/chopin/reference
   article: https://kyle-messier.github.io/chopin/articles
