### 2024-05-30 #638 @beatrizmilz
> > Hi @sigmafelix ! There was one error on the R CMD check: [#638 (comment)](https://github.com/ropensci/software-review/issues/638#issuecomment-2125717203) . Could you please verify that, before we proceed to add the reviewers?
> > Thank you!
>
> @beatrizmilz Thank you for the update. In the most recent version of `chopin`, I think the R CMD CHECK error was fixed. Thank you!

Great!
