### 2024-06-20 #638 @robitalec
Note, I'm placing checkboxes below each bullet, to highlight the places where I think improvements would be beneficial. I won't checkbox all the things that are already completed to keep it tidier.

    - [ ] Consider listing system requirements
    - [ ] Consider linking to dependency eg. {sf} install instructions
    - [X] All vignettes run successfully locally
    - [ ] Consider revising, restructuring (details below)
    - [ ] Missing top-level documentation
    - [ ] Missing example for `extract_at`
    - [ ] There are no contributing guidelines in the README or CONTRIBUTING
    - [ ] Consider adding an .Rproj file for collaborators

    - [ ] Consider improving descriptions and commenting of tests
    - [ ] Consider replacing tests `expect_no_error` with more specific expectations
    - [ ] See suggestions below
    - [ ] See details in other sections

Thank you for inviting me to review @beatrizmilz and, as always, a huge thank you to the broader rOpenSci community for all of your incredible contributions, inclusion and support.

Thank you @sigmafelix and @kyle-messier for this interesting package with functionality I haven't seen elsewhere that should be a great tool for a wide userbase.

Before I start - I really link the name because it evokes chopping up spatial objects into chunks and -- music! Great choice.

---

My first main and most important comment is that, at the moment, I find {chopin} hard to get started using.

I think the pitch is currently too narrow (also mentioned [here](https://github.com/ropensci/software-review/issues/630
)), which might turn away folks who don't identify with "health" or "climate" research. I think there is potential for many users across disciplines who have large spatial data (everything is spatial!) to use {chopin} to perform analyses in more manageable chunks, allowing folks to improve the efficiency of their analyses and extend the potential of their machines.

When I open the README, one of the first things I see is notes on data restrictions and z dimensions and spherical geometry. I think this should come later in a caveats section or FAQ vignette.

Next, we get to "Basic design" which reads more like an overview of functions but also gets into which packages underpin {chopin} functionality, package versions, and registering parallel workers. I think this is the section that really needs to give users a concise and clear overview of the functions. This can and should be reused in other places to follow the recommended principle of *multiple points of entry*, where users should be anticipated to first encounter {chopin} across every source of information (README, documentation, vignettes) and given a clear introduction.

- "We suggest two flowcharts to help which function to use for parallel processing below" - note there is only one flowchart visible below and it isn't linked to in this sentence
- `extract_at_poly` missing description

It seems there are currently three main families of functions exported and number of helper or internal functions:
- `par_*`
    - `par_grid` and others for running functions in parallel
    - `par_make_*` for making grids
- `summarize_*`
- `extract_at_*`

These are the kinds of questions I'm left as an early user with after reading the overview
- Should `par_make_*` functions be split into a new `grid_*` family of functions?
- What is the difference to a user between `extract_at_buffer` with a kernel weight and `summary_aw` an area-weighted summary? What is the difference between `par_make_grid` and `par_make_gridset`?
- If `par_group_grid` is an extension of par_group_balanced for padded grids, should it be renamed something like `par_group_balanced_padded`? Or should `padded = TRUE` be added as an argument?
- Are regular R users going to find useful functions here or is {chopin} just for HPC users?

From the Description, current version

```
Package: chopin
Title: CHOPIN: Computation for Climate and Health research On Parallelized INfrastructure
Description: It enables users with basic understanding on geospatial data
    and sf and terra functions to parallelize geospatial operations for
    geospatial exposure assessment modeling and covariate computation.
    Parallelization is done by dividing large datasets into sub-regions
    with regular grids and data's own hierarchy. A computation over the
    large number of raster files can be parallelized with a chopin
    function as well.
```

Suggestions:

- I think there's a lot of useful information in the [flowchart](https://github.com/NIEHS/chopin/blob/main/man/figures/README-flowchart-mermaid-1.png), how can we incorporate more obviously?
- No need to repeat the package name, capitalized, in the title
- I think the acronym is neat, but the name stands on it's own. Broadening the scope would help users understand how they can use {chopin} and this starts at the title
- I don't think it's necessary to state the expectations of users in the description
- See more examples [here](https://cran.r-project.org/web/packages/available_packages_by_name.html)

```
Package: chopin
Title: Parallelize Geospatial Operations Over Sub-Regions, Regular Grids and Lists of Rasters
Description: It enables users to parallelize geospatial operations over
    sub-regions, regular grids and lists of rasters. Parallelization
    is done by dividing large datasets into sub-regions
    with regular grids and data's own hierarchy. A computation over the
    large number of raster files can be parallelized with a chopin
    function as well. (TODO: maybe mention {future}, {exactextractr}, {terra}, {sf})
```

---

Another reason that I might be finding {chopin} hard to get started using is that it seems to obscure lots of the regular terra/sf steps I would typically do myself as an R user. For example, the `extract_at_buffer` wraps the `exact_extract` function from {exactextractr} but before calling it, it runs `terra::buffer`, manages projections, crops, calculates extents, etc. This is also the case in functions like `reproject_std` and `vect_valid_repair` which reproject/repair objects using functions terra or sf based on the object type.

This also shows itself where the authors manually switch within functions depending on terra or sf input. For example, using `dep_check` and testing `if == 'terra' else if  == 'sf'`. This feels like an opportunity to take advantage of S3 methods. Read more in [Advanced R: S3](https://adv-r.hadley.nz/s3.html#s3-methods) and [R Packages: 11.9](https://r-pkgs.org/dependencies-in-practice.html#imports-and-exports-related-to-s3). An example applied is in {tmap}'s internal function [`tmapShape`](https://github.com/r-tmap/tmap/blob/master/R/tmapShape.R#L61-L174). Obviously this would require a fair amount of refactoring - I just wanted to highlight an alternative approach that might be more flexible to adding new supported data types in the future.

This is definitely an open ended comment - I just wonder where the point is for users where hiding away managing regular steps in a spatial analysis (like projections, extents, clipping, etc) actually results in more confusion than it helps.

---

My last general comment is - how do you envision {chopin} fitting into the broader [High Performance Computing](https://cran.r-project.org/web/views/HighPerformanceComputing.html) landscape in R? One thing I'm personally really interested is how it might interface with {targets}. Looks like maybe I'm saved from commenting too much on this, as my coreviewr @Aariq is one of the authors on the new [{geotargets}](https://github.com/njtierney/geotargets) package along with @njtierney and @brownag (thank you!). I have been eagerly following their progress as a regular {targets} user that has encountered the challenges of including spatial data in {targets} workflows. I wonder how these tools can be brought together or, alternatively, how the authors might communicate their differences and where a user might choose one over the other.

##### System requirements

Suggestions:

- Consider listing system requirements  (eg. netcdf) using recommendations [here](https://github.com/rstudio/r-system-requirements)
- See example from [{magick}](https://github.com/ropensci/magick/blob/master/DESCRIPTION#L20)

##### Vignettes

Issues:

An overview of the vignettes and README material:

- README
    - "Examples will navigate par_grid, par_hierarchy, and par_multirasters functions in chopin to parallelize geospatial operations."
        - par_grid: parallelize over artificial grid polygons
        - chopin::par_hierarchy: parallelize geospatial computations using intrinsic data hierarchy
        - par_multirasters: parallelize over multiple rasters
        - Parallelization of a generic geospatial operation
- Good practice of using `chopin` in HPC
    - Assumptions
    - Basic workflow
- Generate computational grids
    - Computational grids
    - Types of computational grids and their generation
    - Visualize computational grids
- Extracting Weather/Climate Geospatial Data with `chopin`
    - Introduction
    - TerraClimate
    - PRISM dataset
    - Scaled up examples
    - Discussion: which strategy is better? stacked vs file-based parallelization

It is unclear to a user which vignette is the introductory one. The rOpenSci Packaging Guide suggests "The general vignette should present a series of examples progressing in complexity from basic to advanced usage.". At the moment, this looks most like the "Extracting Weather/Climate Geospatial Data with `chopin`" vignette but the title is fairly specific and it is listed last on the website.

Consider splitting information about HPC or hardware into its own vignette. Also consider moving system.time benchmarking to a separate vignette? This could also potentially include tips for saving time and memory, and/or hardware recommendations.

Consider incorporating a section in the introductory vignette with a caveats section
- why is parallelization slower
- comparing single thread vs multi thread calculations
- why we turn off spherical geometry

Consider revising the vignettes and README to develop bullet points into paragraphs. Some shorter lists to compare options or list steps can be helpful but, generally, I find a full page of bullets hard to process.

Images on the website are giving 404 errors
- Generate computational grids

Package issues
- Extracting Weather/Climate Geospatial Data with `chopin`
    - `pkgs <-   c("chopin", "terra", "stars", "sf", "future", "doFuture", "dplyr", "parallelly", "tigris", "tictoc") invisible(sapply(pkgs, library, character.only = TRUE, quietly = TRUE))`
        - packages
        - consider loading packages in way that most users will be more familiar with eg. `library(chopin)`

Using :: or both :: and library in vignettes (users are likely more familiar with `library()`)
- Generate computational grids
- Extracting Weather/Climate Geospatial Data with `chopin`

Suggestions:

Missing link to external functions / packages
- Generate computational grids
    - {anticlust}
- Good practice of using `chopin` in HPC
    - {future.apply}
- Extracting Weather/Climate Geospatial Data with `chopin`
    - {tigris}
    - {terra}
    - ...

##### Function documentation

Issues:

Missing top-level documentation
- See [`usethis::use_package_doc`](https://usethis.r-lib.org/reference/use_package_doc.html)
- Consider using this space to translate the flowchart into a written overview of the main functions with inputs and outputs

Suggestions:

Missing links to {chopin} functions, external functions, other packages
- anticlust `par_group_balanced`
- sf (throughout)
- terra (throughout)
- `par_group_balanced` from `par_group_grid`, ...

`par_group_balanced`
- Users may benefit from a definition of an "inexhaustive grid"

Missing documentation of argument defaults
- `clip_ras_ext` nqsegs = 180L
- `extract_at_poly` func = mean
- ...

Unclear functions
- Given current documentation, it is unclear why/when/where a user use eg. the following function
    - https://niehs.github.io/chopin/reference/par_fallback.html

Use of ...
- Consider explicitly demonstrating and describing this functionality in documentation and vignette
    -  Eg. `par_grid`

##### Examples

Issues:

Missing example
- `extract_at` (reuse example see [roxygen2 vignette](https://roxygen2.r-lib.org/articles/reuse.html))

Suggestions:

Examples not run with `if (FALSE)` or dontrun. Consider a smaller toy example?
- `par_grid`
- `par_hierarchy`
- `par_merge_grid`
- `par_multirasters`
- `vect_valid_repair`

Using :: or both :: and library in examples
- `clip_vec_ext`
- `clip_ras_ext`
- `datamod`
- `dep_check`
- `dep_switch`
- ...

##### Community guidelines

Issues:

There are no contributing guidelines in the README or CONTRIBUTING

##### Automated tests

Issues:

There is good test coverage in the package but I am a bit concerned about the number of times the authors use the `expect_no_error` function.
When I think of effective package testing, I think of both clarity for the user reading the tests and in the authors setting up expectations for how the package may fail (anticipating your own errors, warnings or messages). What does the test `expect_no_error` communicate to the user or set up as expectations?

Here are some ideas for expanding or adjusting the tests:

- inline comments in the test scripts to explain to readers/our future selves as authors ([example](https://github.com/tidyverse/ggplot2/blob/main/tests/testthat/test-coord-flip.R))
- narrow the scope of each `test_that()` call and expand on the description
    - from [https://testthat.r-lib.org/reference/test_that.html], "A test encapsulates a series of expectations about a small, self-contained unit of functionality"
    - comparing in {chopin} eg. `test-gridding.R` ([L116](https://github.com/NIEHS/chopin/blob/main/tests/testthat/test-gridding.R#L116C1-L116C50)) to eg.{rnaturalearth} [`test_ne_states.R`](https://github.com/ropensci/rnaturalearth/blob/master/tests/testthat/test_ne_states.R)
        - we see {chopin} have one blanket description "grid split is well done" to contains all the tests
        - the {rnaturalearth} test has four smaller, more digestible test_that chunks that run a series of related tests
        - as a reader, I know what I'm looking at and find it easy to look for a specific test
        - as a potential contributor, I know which test might be related to my bug, or I have a good example of how to structure a potential new test in a pull request
- more examples: [dplyr test-join-cols.R](https://github.com/tidyverse/dplyr/blob/main/tests/testthat/test-join-cols.R), [purrr test-pluck.R](https://github.com/tidyverse/purrr/blob/main/tests/testthat/test-pluck.R)

```r
library(data.table)

paths <- dir('../chopin/tests/testthat', full.names = TRUE)
DT <- data.table(path = paths)

counts <- DT[, tstrsplit(grep('testthat::', readLines(path), value = TRUE), c('\\('), keep = 1), by = path]

counts[, .N, V1][order(-N)]
```
| Test function               |  N|
|:----------------------------|--:|
| testthat::expect_no_error   | 59|
| testthat::expect_error      | 47|
| testthat::test_that         | 30|
| testthat::expect_equal      | 20|
| testthat::expect_true       | 16|
| testthat::expect_s4_class   | 12|
| testthat::expect_s3_class   | 12|
| testthat::expect_no_error   | 12|
| testthat::expect_equal      | 10|
| testthat::expect_message    |  7|
| testthat::expect_error      |  6|
| testthat::expect_s3_class   |  5|
| testthat::expect_true       |  4|
| testthat::expect_condition  |  2|
| testthat::expect_warning    |  2|
| testthat::expect_warning    |  2|
| testthat::expect_no_warning |  1|

```r
counts[, .N, .(path, V1)][order(path, -N)]
```

| path                     |Test function               |  N|
|:-------------------------|:---------------------------|--:|
| test-any_class_args.R    |testthat::expect_true       |  5|
| test-any_class_args.R    |testthat::test_that         |  1|
| test-any_class_args.R    |testthat::expect_s4_class   |  1|
| test-check.R             |testthat::expect_equal      | 10|
| test-check.R             |testthat::test_that         |  9|
| test-check.R             |testthat::expect_no_error   |  8|
| test-check.R             |testthat::expect_error      |  8|
| test-check.R             |testthat::expect_equal      |  4|
| test-check.R             |testthat::expect_s3_class   |  2|
| test-check.R             |testthat::expect_s4_class   |  2|
| test-check.R             |testthat::expect_error      |  1|
| test-distribute_suites.R |testthat::expect_no_error   |  9|
| test-distribute_suites.R |testthat::expect_no_error   |  8|
| test-distribute_suites.R |testthat::expect_s3_class   |  5|
| test-distribute_suites.R |testthat::test_that         |  4|
| test-distribute_suites.R |testthat::expect_s3_class   |  4|
| test-distribute_suites.R |testthat::expect_true       |  4|
| test-distribute_suites.R |testthat::expect_error      |  2|
| test-distribute_suites.R |testthat::expect_true       |  2|
| test-distribute_suites.R |testthat::expect_error      |  2|
| test-distribute_suites.R |testthat::expect_equal      |  2|
| test-distribute_suites.R |testthat::expect_equal      |  2|
| test-distribute_suites.R |testthat::expect_condition  |  2|
| test-distribute_suites.R |testthat::expect_s4_class   |  1|
| test-gridding.R          |testthat::expect_error      | 18|
| test-gridding.R          |testthat::expect_no_error   | 17|
| test-gridding.R          |testthat::expect_message    |  7|
| test-gridding.R          |testthat::test_that         |  6|
| test-gridding.R          |testthat::expect_true       |  6|
| test-gridding.R          |testthat::expect_s4_class   |  6|
| test-gridding.R          |testthat::expect_equal      |  5|
| test-gridding.R          |testthat::expect_warning    |  2|
| test-gridding.R          |testthat::expect_warning    |  1|
| test-gridding.R          |testthat::expect_s3_class   |  1|
| test-gridding.R          |testthat::expect_no_warning |  1|
| test-par_fallback.R      |testthat::test_that         |  1|
| test-par_fallback.R      |testthat::expect_no_error   |  1|
| test-par_fallback.R      |testthat::expect_error      |  1|
| test-preprocessing.R     |testthat::expect_equal      |  4|
| test-preprocessing.R     |testthat::test_that         |  2|
| test-preprocessing.R     |testthat::expect_error      |  2|
| test-preprocessing.R     |testthat::expect_equal      |  2|
| test-preprocessing.R     |testthat::expect_s3_class   |  1|
| test-preprocessing.R     |testthat::expect_s4_class   |  1|
| test-processing.R        |testthat::expect_no_error   | 25|
| test-processing.R        |testthat::expect_error      | 19|
| test-processing.R        |testthat::test_that         |  7|
| test-processing.R        |testthat::expect_s3_class   |  4|
| test-processing.R        |testthat::expect_no_error   |  3|
| test-processing.R        |testthat::expect_true       |  3|
| test-processing.R        |testthat::expect_s4_class   |  1|
| test-processing.R        |testthat::expect_equal      |  1|
| test-processing.R        |testthat::expect_warning    |  1|

##### Packaging guidelines

Issues:

URLs
- Found some URLs moved or broken

```
> urlchecker::url_check()
! Warning: vignettes/v02_climate_examples.Rmd:135:107 Moved
Fourteen variables are available in TerraClimate data. The table below is from [the TerraClimate website](http://www.climatologylab.org/terraclimate-variables.html).
                                                                                                          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
                                                                                                          https://www.climatologylab.org/terraclimate-variables.html
✖ Error: man/par_make_gridset.Rd:45:11 Error: CRAN URL not in canonical form
[…]
