### 2024-06-20 #638 @sigmafelix
Thank you for the thorough review @robitalec! I am going to start from the points on which I can work immediately, and wait for the other review from @Aariq.

To your comment on the benefits/advantages of hiding necessary tasks for spatial data handling, I agree that this practice would confuse experienced users. In the beginning of the function design, we wanted to make parallelization easier for users who are less experienced with spatial data such that would have difficulty with obtaining variables from source data. Thus, most of intermediate steps between data input and export the final table were obscured by design. For details, in the current version of `chopin`, a vector input is reprojected to a raster input's coordinate system in favor of keeping raster cell values and layouts as they are. For experienced users, I think I could elaborate on the design choices in an introductory vignette for the package. Thank you.
