### 2024-06-24 #638 @Aariq
## Package Review

-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☒ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README
-   ☐ **Vignette(s):** demonstrating major functionality that runs successfully locally
    -   v02\_climate\_examples.Rmd is not reproducible locally for me
-   ☒ **Function Documentation:** for all exported functions
-   ☒ **Examples:** (that run successfully locally) for all exported functions
-   ☐ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).
    -   There’s a code of conduct, but no contribution guidelines

#### Functionality

-   ☒ **Installation:** Installation succeeds as documented.
-   ☒ **Functionality:** Any functional claims of the software have been confirmed.
-   ☒ **Performance:** Any performance claims of the software have been confirmed.
    -   Although, it would be good to see more examples where the overhead cost of parallelizaiton is worth it
-   ☒ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
-   ☒ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

Estimated hours spent reviewing: 3.5

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review Comments

The main functionality of this package is really nice. I’m especially impressed with the thought that went into *how* to break rasters into tiles for optimal parallelization. I’m looking forward to using this package (or at least adapting its ideas) in my own workflows. I think there is some overlap between `chopin` and `geotargets`, in that you can do parallel operations via “branching” in `targets` workflows, but I like that `chopin` doesn’t require knowing `targets` (which can be a bit of a paradigm shift in how one writes R code). I also think the functions for creating grids to iterate over could work well *with* branching in a `targets` workflow, and I’m looking forward to exploring that. A few things stand out to me as areas for potential improvement: First, it seems that a major limitation to the parallel processing workflow is that it will not work with `plan(multisession)` or with `plan(multicore)` on RStudio or Windows. When I try `plan(multisession)` on macOS with RStudio, I get the error `Error in eval(expr, p) : object 'ncpoints_tr' not found` or `subscript out of bounds`, I assume because the raster is not passed to the parallel processes or not `wrap()`ed and `unwrap()`ed if it is. That leaves the command line on macOS or linux as the only ways to run this workflow currently. I’ll be honest—that’s enough of a barrier for me that I’m unlikely to use this package. I’m curious if there’s a plan to get `plan(multisession)` working or if there’s a reason it can’t work. I discovered this limitation on my own running the examples in the readme, but it might be good to document this limitation in the readme and/or in the `@note` for `par_grid()`, `par_heirarchy()`, etc. Second, the majority (or at least very many) of the examples showing parallel workflows are actually *slower* than the sequential versions. I want to better understand when the parallel processing really helps—either through examples that actually show a time savings, or through more discussion in the documentation of when it is useful (ideally with specifics like raster extent and resolution, number of points, system specs, number of grids, etc.). Third, this package includes a lot of helper functions. Some of these seem like they maybe shouldn’t be exported because they obfuscate more than abstract. Others seem useful, but also like they open users to potentially writing confusing code. I’ve included some more detailed notes on this below.

#### README.md

GitHub can process and display mermaid charts in README.md, so rather than displaying an image of your mermaid graph, you could consider just including a mermaid chunk with \`\`\``mermaid`—that gives viewers zoom and pan controls. I’m not sure how `pkgdown` would handle it though.

I think the “To run the examples” section is probably not necessary.

The section on “why parallelization is slower” points the reader to the climate example vignette. However, most of the examples there also seem to be faster as single threaded operations (just based on outputs from `tictoc` or `system.time()`, haven’t been able to test myself).

#### Documentation

The `...` arg is described as “placeholder” in the documentation for `extract_at()`, but seems like it should say “additional arguments passed to `extract_at_buffer()` or `extract_at_poly()`”. The `...` argument for `extract_at_buffer()` and `extract_at_poly()` might be more accurately documented as “currently unused”.

Also regarding `...`, I’m not sure it’s good advice to recommend users use position based arguments with `...` as in `?par_grid`. It would be safer to recommend using named arguments—e.g. for `fun_dist = terra::extract` users should pass `x` and `y` arguments.

In `summarize_sedc()` I would love a little more detail about what exactly SEDC is in the description.

#### Vignettes

-   Images in the v01\_par\_make\_gridset vignette do not show up on the pkgdown website.
-   Consider [precomputing](https://ropensci.org/blog/2019/12/08/precompute-vignettes/) the climate example vignette rather than just using unexecuted code chunks—this may make it easier to keep the vignette up to date as the package evolves. The `lwgeom` package appears to be required for that vignette, the `saveRDS()` code fails if the `input/` directory doesn’t exist, and there doesn’t appear to be any code or instructions for downloading the TerraClimate data, so I can’t reproduce the code in this vignette easily.

#### Functions

Just a few notes on things that stand out to me:

-   I think a lot of the helper functions might be viewed as less helpful by some (which is fine, they don’t have to use them) because they are not explicit about what objects they take or return. While the package authors likely viewed this as a convenience, others may view these helpers as danger. E.g., let’s say you switch your code to start with a `sf` object instead of a `terra` object—now `dep_switch()` does something *different* than it did before even though that part of the code has **not** changed! A “safer” alternative would be `st_as_sf()` which does the same thing whether run on a SpatVector or an sf object. Similarly, I could imagine a wrapper around `terra::vect()`, that when given a SpatVector does nothing, would be a “safer” alternative when expecting a SpatVector output.

-   `any_class_args()` I don’t quite understand the problem this is attempting to solve, but it feels like something that should not be exported. Also, why not use `inherits()` here in place of `grepl(obj, class)`?

-   `dep_check()` Another place where `inherits()` might be cleaner. E.g. I think this does the same: `r     dep_check <- function(input) {       classes <- c("sf", "stars", "SpatVector", "SpatRaster", "SpatVectorProxy")       if (!inherits(input, classes)) stop("Input should be one of sf or Spat* object.\n")       class <- classes[inherits(input, classes, which = TRUE)]       ifelse(class %in% c("SpatVector", "SpatRaster", "SpatVectorProxy"), "terra", "sf")     }`

-   `dep_switch()` is a nice shortcut, but I feel it makes code a little harder to read since it’s not immediately clear whether it is translating from `sf`/`stars` to `terra` or the other way around.

-   `get_clip_ext()`: a little confusing/surprising that the description is “setting the clipping extent” (does it get or set?). This is another example of a helper that could potentially do more damage than good because of its assumption that the projection has units in meters (there don’t appear to be any checks on this).

-   `par_def_q()`: this seems like an internal function that shouldn’t be exported—it obscures what it does compared to just an example using `quantiles()` with `seq()`. It doesn’t appear to be used internally either, so maybe just removed?

#### Tests

I agree with @robitalec that test coverage is great, but there is maybe too much use of `expect_no_error()`. Possible improvments might include additional checks the output has the expected structure, checks that the expected object type is returned with `expect_s3_class()`, and/or using `expect_snapshot()` to at least check that the output of functions doesn’t unexpectedly change. It would be good to see tests with different `future::plan()`s running on different operating systems with CI. Some combination of `withr::local_options(list(future.plan = "multisession"))`, and `testthat::skip_on_os()` should make that possible.
