### 2024-06-25 #638 @sigmafelix
Thank you for the prompt and detailed review, @Aariq! There are a lot of functions/details I need to learn, but I am sure that these functions will help me to improve the package. `multicore` plan was applied in the most of use cases because it ran without serialization of `terra` objects. I am going to integrate serialization at the beginning of parallelization (like `geotargets`) or another workaround in the revision. I will try to submit a revision in a few weeks. Thanks!
