### 2024-06-25 #638 @kyle-messier
@robitalec  @Aariq  This is what peer-review is all about! Thank y'all so much for the thoughtful and detailed reviews. It has been a tremendous learning experience. We look forward to improving `chopin` and your comments will also make other packages and projects better too.
