### 2024-07-01 fcddaf3
0.8.0 dev - S3 experiment in summarize_aw: stalls at roxygen2 documentation (i.e., devtools::document()) - summarize_aw: first two arguments renamed x and y - mirai version, plain future from future.apply - backend setting prototype - DESCRIPTION   - Changed title   - Imports reorganized   - SystemRequirements
