+++ R/check.R
@@ -1,5 +1,6 @@
 #' Return the package the input object is based on
 #' @family Helper functions
+#' @keywords internal
 #' @description Detect whether the input object is sf or Spat* object.
 #' @author Insang Song
 #' @param input Spat* in terra or sf object.
@@ -14,22 +15,17 @@
 #' dep_check(nc_sf)
 #' nc_vect <- terra::vect(nc_sf)
 #' dep_check(nc_vect)
-#' ## END OF EXAMPLE
-#' @export
 dep_check <- function(input) {
-  if (!any(class(input) %in%
-  c("sf", "stars", "SpatVector", "SpatRaster", "SpatVectorProxy"))
+  if (
+    !inherits(
+      input,
+      c("sf", "stars", "SpatVector", "SpatRaster", "SpatVectorProxy")
+    )
   ) {
-    stop("Input should be one of sf or Spat* object.\n")
+    cli::cli_abort("Input should be one of sf or Spat* object.\n")
   }
   if (
-    any(
-      vapply(
-        c("SpatVector", "SpatRaster", "SpatVectorProxy"),
-        FUN = function(x) methods::is(input, x),
-        FUN.VALUE = logical(1)
-      )
-    )
+    inherits(input, c("SpatVector", "SpatRaster", "SpatVectorProxy"))
   ) {
     return("terra")
   }
@@ -39,6 +35,7 @@ dep_check <- function(input) {
 
 #' Return the input's GIS data model type
 #' @family Helper functions
+#' @keywords internal
 #' @description This function returns one of 'vector' or 'raster'
 #' depending on the input class.
 #' @param input Spat*/sf/stars object.
@@ -59,22 +56,26 @@ dep_check <- function(input) {
 #' ra_path <- system.file("ex/elev.tif", package = "terra")
 #' ra <- terra::rast(ra_path)
 #' datamod(ra)
-#' @importFrom methods is
-#' @export
 datamod <- function(input) {
-  if (!any(class(input) %in% c("sf", "stars", "SpatVector", "SpatRaster"))) {
+  if (
+    !inherits(
+      input,
+      c("sf", "stars", "SpatVector", "SpatRaster")
+    )
+  ) {
     stop("Input should be one of sf or Spat* object.\n")
   }
-  if (any(methods::is(input, "SpatVector"), methods::is(input, "sf"))) {
+  if (inherits(input, c("sf", "SpatVector"))) {
     return("vector")
   }
-  if (any(methods::is(input, "SpatRaster"), methods::is(input, "stars"))) {
+  if (inherits(input, c("stars", "SpatRaster"))) {
     return("raster")
   }
 }
 
 #' @title Check coordinate system then reproject
 #' @family Helper functions
+#' @keywords internal
 #' @description The input is checked whether its coordinate system is
 #'  present. If not, it is reprojected to the CRS specified in
 #' \code{crs_standard}.
@@ -132,6 +133,7 @@ reproject_std <-
 
 #' Validate and repair input vector data
 #' @family Helper functions
+#' @keywords internal
 #' @description It tries repairing input vector data.
 #' Vector validity violation usually appears in polygon data with
 #' self-crossing or
@@ -155,7 +157,6 @@ reproject_std <-
 #' }
 #' @importFrom terra makeValid
 #' @importFrom sf st_make_valid
-#' @export
 vect_valid_repair <- function(input_vector) {
   detected <- dep_check(input_vector)
 
@@ -170,6 +171,7 @@ vect_valid_repair <- function(input_vector) {
 
 #' Generate a rectangular polygon from extent
 #' @family Helper functions
+#' @keywords internal
 #' @param extent input extent.
 #'  A numeric vector with xmin/xmax/ymin/ymax,
 #'  [sf::st_bbox] or [terra::ext] outputs.
[…]
