+++ R/chopin-package.R
@@ -1,0 +1,9 @@
+#' @name chopin-package
+#' @aliases chopin
+#' @title Computation of spatial data by hierarchical and objective partitioning
+#'   of inputs for parallel processing
+#' @description
+#' The `chopin` package provides a set of functions to compute on divided
+#'  geospatial data.
+#' @keywords internal
+"_PACKAGE"
