+++ man/reproject_b2r.Rd
@@ -30,6 +30,7 @@ reproject_b2r(nc, elev)
 \seealso{
 Other Helper functions: 
 \code{\link{any_class_args}()},
+\code{\link{check_dist_incorrect}()},
 \code{\link{clip_ras_ext}()},
 \code{\link{clip_vec_ext}()},
 \code{\link{crs_check}()},

[…]
