+++ man/extract_at_poly.Rd
@@ -76,6 +76,7 @@ Other Macros for calculation:
 \code{\link{extract_at}()},
 \code{\link{extract_at_buffer}()},
 \code{\link{kernelfunction}()},
+\code{\link{summarize_aw_old}()},
 \code{\link{summarize_sedc}()}
 }
 \author{
