+++ NAMESPACE
@@ -25,7 +25,11 @@ export(par_multirasters)
 export(reproject_b2r)
 export(reproject_std)
 export(set_backend)
+export(summarize_aw)
 export(summarize_sedc)
+exportMethods(summarize_aw,SpatVector)
+exportMethods(summarize_aw,default)
+exportMethods(summarize_aw,sf)
 importFrom(alphahull,ahull)
 importFrom(anticlust,balanced_clustering)
 importFrom(collapse,rowbind)
