+++ man/extract_at_buffer.Rd
@@ -113,6 +113,7 @@ Other Macros for calculation:
 \code{\link{extract_at}()},
 \code{\link{extract_at_poly}()},
 \code{\link{kernelfunction}()},
+\code{\link{summarize_aw}()},
 \code{\link{summarize_aw_old}()},
 \code{\link{summarize_sedc}()}
 }
