+++ man/summarize_sedc.Rd
@@ -93,6 +93,7 @@ Other Macros for calculation:
 \code{\link{extract_at_buffer}()},
 \code{\link{extract_at_poly}()},
 \code{\link{kernelfunction}()},
+\code{\link{summarize_aw}()},
 \code{\link{summarize_aw_old}()}
 }
 \author{
