+++ man/extract_at.Rd
@@ -47,6 +47,7 @@ Other Macros for calculation:
 \code{\link{extract_at_buffer}()},
 \code{\link{extract_at_poly}()},
 \code{\link{kernelfunction}()},
+\code{\link{summarize_aw}()},
 \code{\link{summarize_aw_old}()},
 \code{\link{summarize_sedc}()}
 }
