+++ man/summarize_aw.Rd
@@ -1,0 +1,107 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/processing.R
+\name{summarize_aw}
+\alias{summarize_aw}
+\alias{summarize_aw,SpatVector,SpatVector-method}
+\alias{summarize_aw,sf,sf-method}
+\title{Area weighted summary using two polygon objects}
+\usage{
+summarize_aw(x, y, ...)
+
+\S4method{summarize_aw}{SpatVector,SpatVector}(
+  x = NULL,
+  y = NULL,
+  target_fields = NULL,
+  id_x = "ID",
+  fun = stats::weighted.mean,
+  extent = NULL
+)
+
+\S4method{summarize_aw}{sf,sf}(
+  x = NULL,
+  y = NULL,
+  target_fields = NULL,
+  id_x = "ID",
+  fun = NULL,
+  extent = NULL
+)
+}
+\arguments{
+\item{x}{A sf object or file path of polygons detectable
+with GDAL driver at weighted means will be calculated.}
+
+\item{y}{A sf object or file path of polygons from
+which weighted means will be calculated.}
+
+\item{target_fields}{character. Field names to calculate area-weighted.}
+
+\item{id_x}{character(1).
+The unique identifier of each polygon in \code{x}.
+Default is \code{"ID"}.}
+
+\item{fun}{function(1). The function to calculate the weighted summary.
+Default is \code{\link[stats:weighted.mean]{stats::weighted.mean}}. The function must have a \code{w} argument.}
+
+\item{extent}{numeric(4) or SpatExtent object. Extent of clipping \code{x}.
+It only works with \code{x} of character(1) file path.
+See \code{\link[terra:ext]{terra::ext}} for more details. Coordinate systems should match.}
+}
+\value{
+A data.frame with all numeric fields of area-weighted means.
+}
+\description{
+When \code{x} and \code{y} are different classes,
+\code{poly_weight} will be converted to the class of \code{x}.
+}
+\note{
+If \code{x} and \code{y} are characters, they will be
+read as \code{sf} objects.
+}
+\examples{
+# package
+library(sf)
+sf_use_s2(FALSE)
+nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))
+nc <- sf::st_transform(nc, 5070)
+pp <- sf::st_sample(nc, size = 300)
+pp <- sf::st_as_sf(pp)
+pp[["id"]] <- seq(1, nrow(pp))
+sf::st_crs(pp) <- "EPSG:5070"
+ppb <- sf::st_buffer(pp, nQuadSegs=180, dist = units::set_units(20, "km"))
+
+system.time(ppb_nc_aw <- summarize_aw(ppb, nc, c("BIR74", "BIR79"), "id"))
+summary(ppb_nc_aw)
+
+# terra examples
+library(terra)
+library(sf)
+options(sf_use_s2 = FALSE)
+ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
+elev <- system.file("ex/elev.tif", package = "terra")
+nc <- terra::vect(ncpath)
+elev <- terra::rast(elev)
+pp <- terra::spatSample(nc, size = 300)
+pp <- terra::project(pp, crs(elev))
+pp <- terra::as.points(pp)
+pp[["id"]] <- seq(1, nrow(pp))
+ppb <- terra::buffer(pp, 20000)
+
+system.time(
+  ppb_nc_aw <- summarize_aw(ppb, nc, c("BIR74", "BIR79"), "id")
+)
+summary(ppb_nc_aw)
+
+}
+\seealso{
+Other Macros for calculation: 
+\code{\link{extract_at}()},
[…]
