+++ man/summarize_aw_old.Rd
@@ -50,6 +50,7 @@ Other Macros for calculation:
 \code{\link{extract_at_buffer}()},
 \code{\link{extract_at_poly}()},
 \code{\link{kernelfunction}()},
+\code{\link{summarize_aw}()},
 \code{\link{summarize_sedc}()}
 }
 \author{
