+++ NAMESPACE
@@ -27,9 +27,7 @@ export(reproject_std)
 export(set_backend)
 export(summarize_aw)
 export(summarize_sedc)
-exportMethods(summarize_aw,SpatVector)
-exportMethods(summarize_aw,default)
-exportMethods(summarize_aw,sf)
+exportMethods(summarize_aw)
 importFrom(alphahull,ahull)
 importFrom(anticlust,balanced_clustering)
 importFrom(collapse,rowbind)
