+++ R/check.R
@@ -44,6 +44,7 @@ dep_check <- function(input) {
 #' it will be considered raster in this package.
 #' @author Insang Song
 #' @returns character(1). One of `"vector"` or `"raster"`.
+#' @importFrom cli cli_abort
 #' @examples
 #' library(sf)
 #' library(terra)
@@ -63,12 +64,22 @@ datamod <- function(input) {
       c("sf", "stars", "SpatVector", "SpatRaster")
     )
   ) {
-    stop("Input should be one of sf or Spat* object.\n")
+    cli::cli_abort("Input should be one of sf or Spat* object.\n")
   }
-  if (inherits(input, c("sf", "SpatVector"))) {
+  if (
+    inherits(
+      input,
+      c("sf", "SpatVector", "SpatVectorProxy", "SpatVectorCollection")
+    )
+  ) {
     return("vector")
   }
-  if (inherits(input, c("stars", "SpatRaster"))) {
+  if (
+    inherits(
+      input,
+      c("stars", "SpatRaster", "SpatRasterDataset", "SpatRasterCollection")
+    )
+  ) {
     return("raster")
   }
 }
@@ -185,23 +196,25 @@ vect_valid_repair <- function(input_vector) {
 #' library(terra)
 #' numext1 <- c(-100, -70, 30, 40)
 #' names(numext1) <- c("xmin", "xmax", "ymin", "ymax")
-#' ext2poly(numext1, "sf")
-#' ext2poly(numext1, "terra")
+#' ext_to_poly(numext1, "sf")
+#' ext_to_poly(numext1, "terra")
 #' @importFrom sf st_as_sf
 #' @importFrom sf st_bbox
 #' @importFrom sf st_set_crs
 #' @importFrom terra vect
 #' @importFrom terra ext
 #' @importFrom terra set.crs
-ext2poly <- function(
+ext_to_poly <- function(
     extent = NULL,
     output_class = c("sf", "terra"),
     crs = "EPSG:4326") {
   output_class <- match.arg(output_class)
-  if (methods::is(extent, "numeric")) {
+  if (is.numeric(extent)) {
     if (is.null(attr(extent, "names"))) {
-      stop("Your extent is an unnamed numeric vector.
-Please define names xmin/xmax/ymin/ymax explicitly.\n")
+      cli::cli_abort(
+        "Your extent is an unnamed numeric vector.",
+        "Please define names xmin/xmax/ymin/ymax explicitly."
+      )
     }
     extent <- switch(
       output_class,
@@ -250,7 +263,7 @@ Please define names xmin/xmax/ymin/ymax explicitly.\n")
 #'
 #' refextnum <- c(-100, -60, 20, 40)
 #' names(refextnum) <- c("xmin", "xmax", "ymin", "ymax")
-#' refext <- ext2poly(refextnum)
+#' refext <- ext_to_poly(refextnum)
 #' is_bbox_within_reference(nc, refext)
 #' @importFrom sf st_as_sfc st_crs st_bbox st_transform st_within
 is_bbox_within_reference <- function(
@@ -258,12 +271,16 @@ is_bbox_within_reference <- function(
   reference = NULL
 ) {
   reference <- sf::st_as_sfc(sf::st_bbox(reference))
-  print(sf::st_crs(reference))
+  cli::cli_inform(
+    "Full CRS:",
+    sf::st_crs(reference),
+    "--- CRS ---"
+  )
 
   data_query_bb <-
     sf::st_as_sfc(sf::st_bbox(data_query),
                   crs = sf::st_crs(data_query))
-  print(sf::st_crs(data_query_bb))
+
   query_matched <- sf::st_transform(data_query_bb, sf::st_crs(reference))
[…]
