+++ R/zzz.R
@@ -1,0 +1,7 @@
+#' @noRd
+.onLoad <- function(libname, pkgname) {
+  options(chopin.backend = "mirai")
+  if (interactive()) {
+    message(pkgname, " ", packageVersion(pkgname))
+  }
+}
