+++ R/processing.R
@@ -1,5 +1,6 @@
 #' Kernel functions
 #' @family Macros for calculation
+#' @keywords internal
 #' @param kernel Kernel type. One of
 #' `"uniform"`, `"quartic"`, `"triweight"`, and `"epanechnikov"`
 #' @param d Distance
@@ -18,7 +19,6 @@
 #' kernelfunction(v_dist, bw_dist2, "quartic")
 #' kernelfunction(v_dist, bw_dist2, "triweight")
 #' kernelfunction(v_dist, bw_dist2, "epanechnikov")
-#' @export
 kernelfunction <-
   function(
     d,
@@ -44,11 +44,12 @@ kernelfunction <-
 #' @family Helper functions
 #' @description Clip input vector by
 #'  the expected maximum extent of computation.
+#' @keywords internal
 #' @author Insang Song
-#' @param pnts `sf` or `SpatVector` object
+#' @param x `sf` or `SpatVector` object to be clipped
+#' @param y `sf` or `SpatVector` object
 #' @param radius `numeric(1)`. Circular buffer radius.
 #'  this value will be automatically multiplied by 1.1
-#' @param target_input `sf` or `SpatVector` object to be clipped
 #' @returns A clipped `sf` or `SpatVector` object.
 #' @examples
 #' library(sf)
@@ -64,34 +65,33 @@ kernelfunction <-
 #' clip_vec_ext(bcsd_rpntm, 1000, bcsd_rpnt)
 #' @importFrom sf st_intersection
 #' @importFrom terra intersect
-#' @export
 clip_vec_ext <- function(
-  pnts,
-  radius,
-  target_input
+  x,
+  y,
+  radius
 ) {
   if (any(
     vapply(
-      list(pnts, radius, target_input),
+      list(x, y, radius),
       FUN = is.null,
       FUN.VALUE = logical(1)
     )
   )) {
     cli::cli_abort("One or more required arguments are NULL. Please check.\n")
   }
-  detected_pnts <- dep_check(pnts)
-  detected_target <- dep_check(target_input)
+  detected_pnts <- dep_check(y)
+  detected_target <- dep_check(x)
 
   if (detected_pnts != detected_target) {
     cli::cli_warn("Inputs are not the same class.\n")
-    target_input <- dep_switch(target_input)
+    target_input <- dep_switch(x)
   }
 
-  ext_input <- get_clip_ext(pnts, radius)
+  ext_input <- get_clip_ext(y, radius)
   cli::cli_inform("Clip target features with the input feature extent...\n")
   if (detected_pnts == "sf") {
-    cae <- ext_input |>
-      sf::st_intersection(x = target_input)
+    cae <-
+      sf::st_intersection(x = target_input, y = ext_input)
   }
   if (detected_pnts == "terra") {
     cae <- terra::intersect(target_input, ext_input)
@@ -102,12 +102,13 @@ clip_vec_ext <- function(
 
 #' Clip input raster with a buffered vector extent.
 #' @family Helper functions
+#' @keywords internal
 #' @description Clip input raster by the expected maximum extent of
 #' computation.
-#' @param pnts `sf` or `SpatVector` object
+#' @param x `SpatRaster` object to be clipped
+#' @param y `sf` or `SpatVector` object
 #' @param radius numeric(1). buffer radius.
-#' This value will be automatically multiplied by 1.25
-#' @param ras `SpatRaster` object to be clipped
+#' This value will be automatically multiplied by 1.1
 #' @param nqsegs `integer(1)`. the number of points per a quarter circle
[…]
