+++ R/processing.R
@@ -148,184 +148,37 @@ clip_ras_ext <- function(
   return(cae)
 }
 
-#' Extract summarized values from raster with points and a buffer radius
-#' `r lifecycle::badge("superseded")`
-#' @family Macros for calculation
-#' @description For simplicity, it is assumed that the coordinate systems of
-#'  the points and the raster are the same.
-#' @note
-#' When `Sys.setenv("CHOPIN_FORCE_CROP" = "TRUE")` is set, the raster will be
-#' cropped to the extent of the polygons (with `snap` = `"out"`).
-#' To note, the function is designed to work with the `exactextractr` package.
-#' Arguments of `exactextractr::exact_extract` are set as below
-#' (default otherwise listed):
-#' * `force_df` = `TRUE`
-#' * `stack_apply` = `TRUE`
-#' * `max_cells_in_memory` = `2e8`
-#' * `progress` = `FALSE`
-#' @param points `sf`/`SpatVector` object.
-#' Coordinates where buffers will be generated.
-#' @param surf `SpatRaster` object or file path(s) with extensions
-#' that are GDAL-compatible. A raster from which a summary will be calculated
-#' @param radius numeric(1). Buffer radius.
-#'  Here we assume circular buffers only
-#' @param id character(1). Unique identifier of each point.
-#' @param qsegs integer(1). Number of vertices at a quarter of a circle.
-#'  Default is `90L`.
-#' @param func a function taking a numeric vector argument.
-#' @param kernel character(1). Name of a kernel function
-#' One of `"uniform"`, `"triweight"`, `"quartic"`, and `"epanechnikov"`
-#' @param bandwidth numeric(1). Kernel bandwidth.
-#' @param extent numeric(4) or SpatExtent. Extent of clipping vector.
-#'   It only works with `points` of character(1) file path.
-#'   When using numeric(4), it should be in the order of
-#'   `c(xmin, xmax, ymin, ymax)`. The coordinate system should be the same
-#'   as the `points`.
-#' @param max_cells integer(1). Maximum number of cells in memory.
-#' See [`exactextractr::exact_extract`] for more details.
-#' @param ... Placeholder.
-#' @returns a data.frame object with mean value
-#' @author Insang Song \email{geoissong@@gmail.com}
-#' @examples
-#' library(terra)
-#' rrast <- terra::rast(nrow = 100, ncol = 100)
-#' terra::crs(rrast) <- "EPSG:5070"
-#' terra::values(rrast) <- rgamma(1e4, 4, 2)
-#' rpnt <- terra::spatSample(rrast, 100L, as.points = TRUE)
-#' rpnt$pid <- sprintf("id_%03d", seq(1, 100))
-#' extract_at_buffer(rpnt, rrast, 4, "pid")
-#' @importFrom exactextractr exact_extract
-#' @importFrom methods is
-#' @importFrom terra ext
-#' @importFrom terra crop
-#' @importFrom terra buffer
-#' @importFrom dplyr group_by
-#' @importFrom dplyr summarize
-#' @importFrom dplyr all_of
-#' @importFrom dplyr across
-#' @importFrom dplyr ungroup
-#' @export
-extract_at_buffer <- function(
-  points = NULL,
-  surf = NULL,
-  radius = NULL,
-  id = NULL,
-  qsegs = 90L,
-  func = "mean",
-  kernel = NULL,
-  bandwidth = NULL,
-  extent = NULL,
-  max_cells = 2e7,
-  ...
-) {
-
-  if (!methods::is(surf, "SpatRaster")) {
-    surf <- try(terra::rast(surf))
[…]
