### 2024-07-03 67b23a8
0.7.4 - Name change: reproject_b2r -> reproject_to_raster, vect_valid_repair -> vect_validate - targets-friendly helper function: par_split_list - Vignette for par_split_list - Removed older summarize_aw from the source
