+++ R/preprocessing.R
@@ -1,62 +1,3 @@
-#' Switch spatial data class
-#' @family Helper functions
-#' @description Convert class between `sf`/`stars`-`terra`
-#' @author Insang Song
-#' @param input Spat* in terra or sf object.
-#' @returns Data converted to the other package class
-#' (if sf, terra; if terra, sf)
-#' @examples
-#' library(sf)
-#' library(stars)
-#' library(terra)
-#' options(sf_use_s2 = FALSE)
-#'
-#' ## generate a random raster
-#' ras_rand <- terra::rast(nrow = 30, ncol = 30)
-#' terra::values(ras_rand) <- runif(900)
-#' stars_rand <- dep_switch(ras_rand)
-#' stars_rand
-#' # should return stars object
-#'
-#' vec_rand <- terra::spatSample(ras_rand, size = 10L, as.points = TRUE)
-#' sf_rand <- dep_switch(vec_rand)
-#' sf_rand
-#' # should return sf object
-#' @importFrom terra vect rast
-#' @importFrom sf st_as_sf
-#' @importFrom stars st_as_stars
-#' @importFrom cli cli_abort cli_inform
-#' @keywords internal
-dep_switch <- function(input) {
-  if (!inherits(input, c("sf", "stars", "SpatVector", "SpatRaster"))) {
-    cli::cli_abort("Input should be one of sf or Spat* object.\n")
-  }
-  cls_input <- dep_check(input)
-  type_input <- datamod(input)
-  # search strings. can be expanded.
-  candidates <- c("sf", "terra")
-  cli::cli_inform(
-    sprintf(
-      "Switch %s class to %s...",
-      cls_input, setdiff(candidates, cls_input)
-    )
-  )
-
-  switched <-
-    switch(cls_input,
-      sf = switch(type_input,
-        vector = terra::vect(input),
-        raster = terra::rast(input)
-      ),
-      terra = switch(type_input,
-        vector = sf::st_as_sf(input),
-        raster = stars::st_as_stars(input)
-      )
-    )
-
-  return(switched)
-}
-
 #' Setting the clipping extent
 #' @family Helper functions
 #' @description Return clipping extent with buffer radius.
@@ -99,4 +40,113 @@ get_clip_ext <- function(
     ext_input <- sf::st_as_sfc(ext_input)
   }
   return(ext_input)
+}
+
+
+#' Clip to the buffered extent of input vector
+#' @family Helper functions
+#' @description Clip input vector by
+#'  the expected maximum extent of computation.
+#' @keywords internal
+#' @author Insang Song
+#' @param x `sf` or `SpatVector` object to be clipped
+#' @param y `sf` or `SpatVector` object
+#' @param radius `numeric(1)`. Circular buffer radius.
+#'  this value will be automatically multiplied by 1.1
+#' @returns A clipped `sf` or `SpatVector` object.
+#' @examples
+#' library(sf)
+#' library(stars)
+#' library(terra)
+#' options(sf_use_s2 = FALSE)
+#'
+#' bcsd_path <- system.file(package = "stars", "nc/bcsd_obs_1999.nc")
+#' bcsd <- stars::read_stars(bcsd_path)
+#' bcsd <- sf::st_as_sf(bcsd)
+#' bcsd_rpnt <- sf::st_as_sf(sf::st_sample(bcsd, 4L))
+#' bcsd_rpntm <- sf::st_as_sf(sf::st_sample(bcsd, 1000L))
+#' clip_vec_ext(bcsd_rpntm, 1000, bcsd_rpnt)
+#' @importFrom sf st_intersection
+#' @importFrom terra intersect
+clip_vec_ext <- function(
+  x,
+  y,
[…]
