### 2024-07-05 14a6eb3
0.7.5 - Preprocessing "dot-functions" -- universally usable functions are plain S3, whereas class-specific in S4 - Dropped   - Custom backend for future and mirai   - any_class_args, is_within_ref, par_fallback   - Arguments     - par_grid: grids_input_ids   - Data: SRTM (TIFF replaces RDS) - Added   - CITATION.cff   - Arguments     - par_*: pad_y to enable applying padded extent to x or y (e.g., y in raster-vector overlay and x otherwise) - Updated   - README.Rmd: split the combined one into two mermaid figures -> MD rebuilt   - Arguments       - par_*: debug to .debug
