+++ tools/vignettes-sources/v02_par_functions_spin_reprex.R
@@ -38,7 +38,7 @@ ncpoints$pid <- seq(1, nrow(ncpoints))
 
 
 ## ----load-srtm----------------------------------------------------------------
-srtm <- terra::unwrap(readRDS("../../tests/testdata/nc_srtm15_otm.rds"))
+srtm <- terra::unwrap(readRDS("../../tests/testdata/nc_srtm15_otm.tif"))
 srtm
 plot(srtm)
 terra::crs(srtm) <- "EPSG:5070"
@@ -159,7 +159,7 @@ system.time(
 ## ----multirasters-------------------------------------------------------------
 ncpath <- "../testdata/nc_hierarchy.gpkg"
 nccnty <- terra::vect(ncpath, layer = "county")
-ncelev <- terra::unwrap(readRDS("../testdata/nc_srtm15_otm.rds"))
+ncelev <- terra::unwrap(readRDS("../testdata/nc_srtm15_otm.tif"))
 terra::crs(ncelev) <- "EPSG:5070"
 names(ncelev) <- c("srtm15")
 tdir <- tempdir()
