+++ R/scale_process.R
@@ -36,6 +36,42 @@ par_fallback <-
   }
 
 
+#' Prescreen input data for parallelization
+#' 
+#' This function takes input object and type character to ingest
+#' the input object to return the object in the desired class.
+#' @param type character(1). "raster" or "vector".
+#' @param input object. Input object.
+#' @param input_id character(1). Default is NULL. If NULL, the function
+#'  will not check the object with an ID column.
+#' @param out_class character(1). Default is NULL, but should be one of
+#'   `c("sf", "terra")`.
+#' @param grid_ numeric(4)/SpatExtent/st_bbox object. Loading window.
+#' @keywords internal
+.par_screen <- function(
+  type,
+  input,
+  input_id = NULL,
+  out_class = NULL,
+  grid_ = NULL
+) {
+  # type check
+  match.arg(type, c("raster", "vector"))
+
+  if (type == "raster") {
+    scr <- .check_raster(input = input, extent = grid_)
+
+  } else {
+    scr <- .check_subject(input = input, input_id = input_id, extent = grid_,
+        out_class = out_class)
+  }
+  return(scr)
+
+}
+
+
+
+
 #' @title Process a given function in the entire or partial computational grids
 #' @family Parallelization
 #' @description
@@ -95,7 +131,7 @@ par_fallback <-
 #' ncpnts <- terra::project(ncpnts, "EPSG:5070")
 #' ncelev <-
 #'   terra::unwrap(
-#'     readRDS(system.file("extdata/nc_srtm15_otm.rds", package = "chopin"))
+#'     readRDS(system.file("extdata/nc_srtm15_otm.tif", package = "chopin"))
 #'   )
 #' terra::crs(ncelev) <- "EPSG:5070"
 #' names(ncelev) <- c("srtm15")
@@ -135,6 +171,7 @@ par_fallback <-
 #' @importFrom collapse rowbind
 #' @importFrom sf sf_use_s2
 #' @importFrom cli cli_abort cli_inform
+#' @importFrom methods getPackageName
 #' @export
 par_grid <-
   function(
@@ -185,36 +222,51 @@ par_grid <-
       grids$original[grid_target_ids, ]
     grids_target_list <-
       base::split(grids_target_in, unlist(grids_target_in[["CGRIDID"]]))
-
+    results <- vector("list", length = length(grids_target_list))
     # results_distributed <-
     #   future.apply::future_lapply(
     #     grids_target_list,
     #     function(grid) {
-          sf::sf_use_s2(FALSE)
-          grid <-
-            tryCatch(
-              terra::vect(grid),
-              error = function(e) grid
-            )
+    sf::sf_use_s2(FALSE)
 
-          args_input <- list(...)
-          run_result <- tryCatch({
+    foo_set <- methods::getPackageName(environment(fun_dist))
+
+    args_input <- list(...)
+    peek_x <- try(.check_character(args_input$x))
+    peek_y <- try(.check_character(args_input$y))
+
+    for (i in seq_along(results)) {
+      results[[i]] <-
+        .backend_worker({
+          tryCatch({
             ## Strongly assuming that
             # the first is "at", the second is "from"
-            if (is.character(args_input[[1]])) {
-              args_input$extent <- terra::ext(grid)
-            } else {
-              if (dep_check(grid) != dep_check(args_input[[1]])) {
-                grid <- dep_switch(grid)
-              }
-              grid <- reproject_std(grid, terra::crs(args_input[[1]]))
-              args_input[[1]] <-
-                args_input[[1]][grid, ]
+            gpad_in <- grids$padded[grids$padded$CGRIDID %in% grid$CGRIDID, ]
+            grid <- grids_target_list[[i]]
+            grid <-
+              tryCatch(
+                terra::vect(grid),
[…]

[…]
