+++ R/check.R
@@ -4,7 +4,7 @@
 #' @description Detect whether the input object is sf or Spat* object.
 #' @author Insang Song
 #' @param input Spat* in terra or sf object.
-#' @returns A character object; one of `"terra"` and `"sf"`
+#' @returns A character object; one of `"character"`, `"terra"` and `"sf"`
 #' @examples
 #' library(sf)
 #' library(terra)
@@ -19,10 +19,15 @@ dep_check <- function(input) {
   if (
     !inherits(
       input,
-      c("sf", "stars", "SpatVector", "SpatRaster", "SpatVectorProxy")
+      c("sf", "stars",
+        "SpatVector", "SpatRaster", "SpatVectorProxy",
+        "character")
     )
   ) {
     cli::cli_abort("Input should be one of sf or Spat* object.\n")
+  }
+  if (inherits(input, "character")) {
+    return("character")
   }
   if (
     inherits(input, c("SpatVector", "SpatRaster", "SpatVectorProxy"))
@@ -121,7 +126,8 @@ datamod <- function(input) {
   if (
     !inherits(
       input,
-      c("sf", "stars", "SpatVector", "SpatRaster")
+      c("sf", "stars", "SpatVector", "SpatRaster",
+        "SpatVectorProxy", "SpatRasterDataset", "SpatRasterCollection")
     )
   ) {
     cli::cli_abort("Input should be one of sf or Spat* object.\n")
@@ -169,11 +175,8 @@ datamod <- function(input) {
 #'
 #' nc_vect <- terra::vect(nc_sf)
 #' reproject_std(nc_vect, base_crs)
-#' @importFrom sf st_crs
-#' @importFrom sf st_transform
-#' @importFrom terra crs
-#' @importFrom terra project
-#' @export
+#' @importFrom sf st_crs st_transform
+#' @importFrom terra crs project
 reproject_std <-
   function(
     input,
@@ -191,21 +194,24 @@ reproject_std <-
       sf = sf::st_crs(crs_standard)$epsg,
       terra = terra::crs(crs_standard, describe = TRUE)$code
     )
-    if (input_crs == standard_crs) {
-      return(input)
+    if (input_crs != standard_crs) {
+      cli::cli_inform(
+        sprintf("Reprojecting input:...\n\n--- CRS ---\n%s\n", standard_crs)
+      )
+      input <- switch(
+        bound_package,
+        sf = sf::st_transform(input, sf::st_crs(crs_standard)),
+        terra = terra::project(x = input, y = crs_standard)
+      )
     }
-    input_transformed <- switch(
-      bound_package,
-      sf = sf::st_transform(input, sf::st_crs(crs_standard)),
-      terra = terra::project(x = input, y = crs_standard)
-    )
-    return(input_transformed)
+    return(input)
   }
 
 
 
 #' @title Align vector CRS to raster's
 #' @family Helper functions
+#' @keywords internal soft-deprecated
 #' @param vector `sf`/`stars`/`SpatVector`/`SpatRaster` object
 #' @param raster `SpatRaster` object
 #' @returns Reprojected object in the same class as \code{vector}
@@ -237,7 +243,7 @@ reproject_to_raster <-
 
 #' Validate and repair input vector data
 #' @family Helper functions
-#' @keywords internal
+#' @keywords internal soft-validated
 #' @description It tries repairing input vector data.
 #' Vector validity violation usually appears in polygon data with
 #' self-crossing or
@@ -275,7 +281,7 @@ vect_validate <- function(input_vector) {
 
 #' Generate a rectangular polygon from extent
 #' @family Helper functions
-#' @keywords internal
[…]
