+++ R/processing.R
@@ -133,6 +133,7 @@ kernelfunction <-
   kernel_func = stats::weighted.mean,
   bandwidth = NULL,
   max_cells = NULL,
+  .standalone = TRUE,
   ...
 ) {
   x <-
@@ -141,17 +142,19 @@ kernelfunction <-
       extent = extent
     )
 
-  y <-
-    .check_subject(
-      y,
-      extent = extent,
-      out_class = out_class,
-      subject_id = id
-    )
-  # reproject polygons to raster's crs
-  y <- reproject_to_raster(vector = y, raster = x)
-  if (dep_check(y) == "terra") {
-    y <- dep_switch(y)
+  if (.standalone) {
+    y <-
+      .check_vector(
+        y,
+        extent = extent,
+        out_class = out_class,
+        subject_id = id
+      )
+    # reproject polygons to raster's crs
+    y <- reproject_to_raster(vector = y, raster = x)
+    if (dep_check(y) == "terra") {
+      y <- dep_switch(y)
+    }
   }
 
   if (!is.null(radius)) {
@@ -292,6 +295,7 @@ setMethod(
     kernel_func = stats::weighted.mean,
     bandwidth = NULL,
     max_cells = 3e+07,
+    .standalone = TRUE,
     ...
   ) {
     .extract_at(
@@ -302,7 +306,8 @@ setMethod(
       kernel = kernel,
       kernel_func = kernel_func,
       bandwidth = bandwidth,
-      max_cells = max_cells
+      max_cells = max_cells,
+      .standalone = .standalone
     )
   }
 )
@@ -328,6 +333,7 @@ setMethod(
     kernel_func = stats::weighted.mean,
     bandwidth = NULL,
     max_cells = 3e+07,
+    .standalone = TRUE,
     ...
   ) {
     .extract_at(
@@ -338,7 +344,8 @@ setMethod(
       kernel = kernel,
       kernel_func = kernel_func,
       bandwidth = bandwidth,
-      max_cells = max_cells
+      max_cells = max_cells,
+      .standalone = .standalone
     )
   }
 )
@@ -364,6 +371,7 @@ setMethod(
     kernel_func = stats::weighted.mean,
     bandwidth = NULL,
     max_cells = 3e+07,
+    .standalone = TRUE,
     ...
   ) {
     .extract_at(
@@ -374,7 +382,8 @@ setMethod(
       kernel = kernel,
       kernel_func = kernel_func,
       bandwidth = bandwidth,
-      max_cells = max_cells
+      max_cells = max_cells,
+      .standalone = .standalone
     )
   }
 )
@@ -399,6 +408,7 @@ setMethod(
     kernel_func = stats::weighted.mean,
     bandwidth = NULL,
     max_cells = 3e+07,
+    .standalone = TRUE,
     ...
   ) {
     .extract_at(
@@ -409,7 +419,8 @@ setMethod(
       kernel = kernel,
       kernel_func = kernel_func,
       bandwidth = bandwidth,
-      max_cells = max_cells
+      max_cells = max_cells,
+      .standalone = TRUE
     )
   }
 )
@@ -435,6 +446,7 @@ setMethod(
     kernel_func = stats::weighted.mean,
     bandwidth = NULL,
     max_cells = 3e+07,
+    .standalone = TRUE,
     ...
   ) {
     .extract_at(
@@ -445,7 +457,8 @@ setMethod(
       kernel = kernel,
       kernel_func = kernel_func,
       bandwidth = bandwidth,
-      max_cells = max_cells
+      max_cells = max_cells,
+      .standalone = .standalone
     )
   }
 )
@@ -472,6 +485,7 @@ setMethod(
     kernel_func = stats::weighted.mean,
     bandwidth,
     max_cells = 3e+07,
+    .standalone = TRUE,
     ...
[…]
