+++ R/zzz.R
@@ -1,6 +1,5 @@
 #' @noRd
 .onLoad <- function(libname, pkgname) {
-  options(chopin.backend = "mirai")
   if (interactive()) {
     message(pkgname, " ", packageVersion(pkgname))
   }

[…]
