+++ R/gridding.R
@@ -34,33 +34,32 @@
 #' @description Using input points, the bounding box is split to
 #'  the predefined numbers of columns and rows.
 #'  Each grid will be buffered by the radius.
-#' @seealso [par_cut_coords], [par_merge_grid]
+#' @seealso [par_merge_grid]
 #' @author Insang Song
 #' @examples
 #' # data
 #' library(sf)
-#' sf_use_s2(FALSE)
+#' options(sf_use_s2 = FALSE)
 #' ncpath <- system.file("shape/nc.shp", package = "sf")
 #' nc <- read_sf(ncpath)
 #' nc <- st_transform(nc, "EPSG:5070")
 #'
 #' # run: nx and ny should strictly be integers
-#' # In the example below, nx is 12L, not 12.
 #' nc_comp_region <-
 #'   par_pad_grid(
 #'     nc,
 #'     mode = "grid",
-#'     nx = 12L, ny = 8L,
+#'     nx = 4L, ny = 2L,
 #'     padding = 10000)
 #' par(mfcol = c(1, 2))
-#' plot(nc_comp_region$original)
-#' plot(nc_comp_region$padded)
+#' plot(nc_comp_region$original$geometry)
+#' plot(nc_comp_region$padded$geometry)
 #'
 #' nc_comp_region_wkt <-
 #'   par_pad_grid(
 #'     nc,
 #'     mode = "grid",
-#'     nx = 12L, ny = 8L,
+#'     nx = 4L, ny = 2L,
 #'     padding = 10000,
 #'     return_wkt = TRUE)
 #' nc_comp_region_wkt$original
@@ -204,9 +203,12 @@ par_pad_grid <-
 #' @examples
 #' library(terra)
 #' library(sf)
+#' options(sf_use_s2 = FALSE)
+#'
 #' ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
 #' nc <- terra::vect(ncpath)
 #' nc_rp <- terra::spatSample(nc, 1000)
+#'
 #' nc_gr <- par_pad_balanced(nc_rp, 10L, 1000)
 #' nc_gr
 #' @importFrom terra as.polygons ext buffer
@@ -270,6 +272,7 @@ par_pad_balanced <-
 
 #' @title Generate grid polygons
 #' @family Parallelization
+#' @keywords internal
 #' @description Returns a sf object that includes x- and y- index
 #' by using two inputs ncutsx and ncutsy, which are x- and
 #' y-directional splits, respectively.
@@ -360,6 +363,7 @@ par_make_grid <-
 #' @importFrom terra vect distance
 #' @importFrom stats dist
 #' @importFrom cli cli_abort
+#' @keywords internal
 par_make_balanced <- function(
   points_in = NULL,
   n_clusters = NULL
@@ -401,6 +405,7 @@ par_def_q <- function(steps = 4L) {
 #' @title Partition coordinates into quantile polygons
 #' @note This function is only for two-dimensional points.
 #' @family Parallelization
+#' @keywords internal
 #' @param x numeric/sf/SpatVector. x-coordinates (if numeric).
 #' @param y numeric. y-coordinates.
 #' @param quantiles numeric vector. Quantiles.
@@ -409,7 +414,7 @@ par_def_q <- function(steps = 4L) {
 #' library(terra)
 #' random_points <-
 #'   data.frame(x = runif(1000, 0, 100), y = runif(1000, 0, 100))
-#' quantiles <- par_def_q(4L)
+#' quantiles <- seq(0, 1, length.out = 5L)
[…]
