+++ R/ncpoints.R
@@ -1,12 +1,0 @@
-#' Mildly clustered points in North Carolina, United States
-#' @family Dataset
-#' @format A data frame with 2,304 rows and two variables:
-#' \describe{
-#' \item{X}{X coordinate}
-#' \item{Y}{Y coordinate}
-#' }
-#' @note Coordinates are in EPSG:5070 (Conus Albers Equal Area)
-#' @source sf package data `nc`
-#' @examples
-#' data("ncpoints", package = "chopin")
-"ncpoints"
