+++ R/gridding.R
@@ -16,7 +16,7 @@
 #'  clip actual datasets. Depending on the length unit of the CRS of input.
 # nolint start
 #' @param unit character(1). The length unit for padding (optional).
-#'   units::set_units is used for padding when sf object is used.
+#'   `units::set_units` is used for padding when `sf` object is used.
 #'   See [link](https://cran.r-project.org/web/packages/units/vignettes/measurement_units_in_R.html)
 #'   for the list of acceptable unit forms.
 #' @param quantiles numeric. Quantiles for `grid_quantile` mode.
@@ -34,7 +34,7 @@
 #' @description Using input points, the bounding box is split to
 #'  the predefined numbers of columns and rows.
 #'  Each grid will be buffered by the radius.
-#' @seealso [par_merge_grid]
+#' @seealso [`par_merge_grid`]
 #' @author Insang Song
 #' @examples
 #' # data
@@ -522,7 +522,7 @@ par_cut_coords <- function(x = NULL, y = NULL, quantiles) {
 #' @author Insang Song
 #' @param points_in `sf` or `SpatVector` object. Target points of computation.
 #' @param grid_in `sf` or `SpatVector` object.
-#'   The grid generated by [`par_make_grid`].
+#'   The grid generated by the internal function `par_make_grid`.
 #' @param grid_min_features integer(1). Threshold to merge adjacent grids.
 #' @param merge_max integer(1).
 #'   Maximum number of grids to merge per merged set. Default is 4.
@@ -663,10 +663,12 @@ par_merge_grid <-
           which(graph_member_excess == unique(graph_member_excess)[i])
         graph_member_excess_repl <-
           graph_member_excess[graph_member_excess_this]
+        # TO SELF:
         # 1e6 is arbitrarily chosen; it should be large enough to avoid
-        # conflicts with the original membership
-        # I do believe this number will not be changed as 1e6+
-        # computation grids are not practical
+        # conflicts with the original membership.
+        # I do believe this number is sufficient as 1e6+
+        # computation grids are not practical.
+
         # split chunks length of merge_max
         graph_member_excess_split <-
           split(
