+++ man/reproject_to_raster.Rd
@@ -30,7 +30,6 @@ reproject_to_raster(nc, elev)
 }
 \seealso{
 Other Helper functions: 
-\code{\link{check_dist_incorrect}()},
 \code{\link{clip_ras_ext}()},
 \code{\link{clip_vec_ext}()},
 \code{\link{crs_check}()},
