+++ man/datamod.Rd
@@ -36,7 +36,6 @@ datamod(ra)
 }
 \seealso{
 Other Helper functions: 
-\code{\link{check_dist_incorrect}()},
 \code{\link{clip_ras_ext}()},
 \code{\link{clip_vec_ext}()},
 \code{\link{crs_check}()},

[…]
