+++ R/check.R
@@ -188,17 +188,17 @@ reproject_std <-
     bound_package <- dep_check(input)
     input_crs <- switch(
       bound_package,
-      sf = sf::st_crs(input)$epsg,
+      sf = sf::st_crs(input)$wkt,
       terra = terra::crs(input, describe = TRUE)$code
     )
     standard_crs <- switch(
       bound_package,
-      sf = sf::st_crs(crs_standard)$epsg,
+      sf = sf::st_crs(crs_standard)$wkt,
       terra = terra::crs(crs_standard, describe = TRUE)$code
     )
     if (input_crs != standard_crs) {
       cli::cli_inform(
-        sprintf("Reprojecting input:...\n\n--- CRS ---\n%s\n", standard_crs)
+        sprintf("Reprojecting input:\n\n--- CRS ---\n\n%s\n", standard_crs)
       )
       input <- switch(
         bound_package,
@@ -280,166 +280,6 @@ vect_validate <- function(input_vector) {
 }
 
 
-#' Generate a rectangular polygon from extent
-#' @family Helper functions
-#' @keywords internal soft-deprecated
-#' @param extent input extent.
-#'  A numeric vector with xmin/xmax/ymin/ymax,
-#'  [sf::st_bbox] or [terra::ext] outputs.
-#' @param output_class character(1).
-#'  Class of the output polygon. One of `"sf"` or `"terra"`
-#' @param crs character(1). Coordinate reference system definition.
-#' @returns `sf` or `SpatVector` object of a rectangular polygon.
-#' @author Insang Song
-#' @examples
-#' library(sf)
-#' library(terra)
-#' numext1 <- c(-100, -70, 30, 40)
-#' names(numext1) <- c("xmin", "xmax", "ymin", "ymax")
-#' ext_to_poly(numext1, "sf")
-#' ext_to_poly(numext1, "terra")
-#' @importFrom sf st_as_sf st_bbox st_set_crs
-#' @importFrom terra vect ext set.crs
-ext_to_poly <- function(
-    extent = NULL,
-    output_class = c("sf", "terra"),
-    crs = "EPSG:4326") {
-  output_class <- match.arg(output_class)
-  if (is.numeric(extent)) {
-    if (output_class == "sf") {
-      if (is.null(attr(extent, "names"))) {
-        cli::cli_abort(
-          paste0(
-            "Your extent is an unnamed numeric vector.\n",
-            "Please define names xmin/xmax/ymin/ymax explicitly.\n"
-          )
-        )
-      }
-    }
-    extent <- switch(
-      output_class,
-      sf = sf::st_bbox(extent),
-      terra = terra::ext(extent)
-    )
-  }
-
-  extent_polygon <- switch(
-    output_class,
-    sf = sf::st_as_sf(sf::st_as_sfc(extent)),
-    terra = terra::vect(extent)
-  )
-
-  extent_polygon <- switch(
-    output_class,
-    sf = sf::st_set_crs(extent_polygon, sf::st_crs(crs)),
-    terra = terra::set.crs(extent_polygon, terra::crs(crs))
-  )
-
-  return(extent_polygon)
-
-}
-
-
-#' Check if the data extent is inside the reference bounding box
-#' @family Helper functions
-#' @keywords internal soft-deprecated
-#' @description One of the most common errors in spatial computation is rooted
-#' in the entirely or partly incomparable spatial extents of input datasets.
-#' This function returns whether your data is inside the target computational
-#' extent.
-#' It is assumed that you know and have the exact computational region.
-#' This function will return `TRUE` if the reference region
[…]
