+++ tests/testthat/test-check.R
@@ -1,5 +1,5 @@
 
-testthat::test_that("What package does the input object belong?",
+testthat::test_that("Input object class is detected",
   {
     withr::local_package("stars")
     withr::local_package("terra")
@@ -13,11 +13,18 @@ testthat::test_that("What package does the input object belong?",
 
     testthat::expect_equal(packbound_stars, "sf")
     testthat::expect_equal(packbound_terra, "terra")
+
+    # proxy
+    ncpath <- system.file(package = "sf", "shape/nc.shp")
+    nc <- terra::vect(ncpath, proxy = TRUE)
+    packbound_proxy <- dep_check(nc)
+    testthat::expect_equal(packbound_proxy, "terra")
+
   }
 )
 
 
-testthat::test_that("What package does the input object belong?",
+testthat::test_that("Data model detection",
   {
     withr::local_package("stars")
     withr::local_package("terra")
@@ -52,11 +59,6 @@ testthat::test_that("CRS is transformed when it is not standard", {
   sf::st_crs(ncna) <- NA
   ncnatr <- terra::vect(ncna)
 
-  testthat::expect_no_error(reproject_std(nc, crs_standard = "EPSG:4326"))
-  testthat::expect_no_error(reproject_std(nc, crs_standard = "EPSG:5070"))
-  testthat::expect_no_error(reproject_std(nctr, crs_standard = "EPSG:4326"))
-  testthat::expect_no_error(reproject_std(nctr, crs_standard = "EPSG:5070"))
-
   nctr_align <- reproject_std(nctr, "EPSG:4326")
   nc_align <- reproject_std(nc, "EPSG:4326")
 
@@ -76,6 +78,28 @@ testthat::test_that("CRS is transformed when it is not standard", {
 })
 
 
+testthat::test_that("reproject to raster: sf", {
+  withr::local_package("sf")
+  withr::local_package("terra")
+  withr::local_options(list(sf_use_s2 = FALSE))
+
+  nc <- system.file(package = "sf", "shape/nc.shp")
+  nc <- sf::read_sf(nc)
+  nc <- sf::st_transform(nc, "EPSG:5070")
+  nctr <- terra::vect(nc)
+  terra::crs(nctr) <- "EPSG:5070"
+
+  # make random raster
+  rr <- terra::rast(matrix(runif(100), 10, 10), crs = "EPSG:4326")
+
+  ncr <- reproject_to_raster(nc, rr)
+  nctr <- reproject_to_raster(nctr, rr)
+
+  testthat::expect_s3_class(ncr, "sf")
+  testthat::expect_s4_class(nctr, "SpatVector")
+})
+
+
 testthat::test_that("vector validity check is cleared", {
   withr::local_package("sf")
   withr::local_package("terra")
@@ -91,94 +115,30 @@ testthat::test_that("vector validity check is cleared", {
 })
 
 
-testthat::test_that("input extent is converted to a polygon", {
+testthat::test_that(".check_id throws error with non-character id", {
   withr::local_package("sf")
-  withr::local_package("terra")
   withr::local_options(list(sf_use_s2 = FALSE))
+  input_char <- system.file("gpkg/nc.gpkg", package = "sf")
+  input_sf <- sf::st_read(input_char)
 
-  mainland_vec <- c(xmin = -128, xmax = -62, ymin = 25, ymax = 52)
-  mainland_box <- ext_to_poly(mainland_vec, output_class = "sf")
-  mainland_box_t <- ext_to_poly(mainland_vec, output_class = "terra")
-  mainland_vec_un <- unname(mainland_vec)
-
-  testthat::expect_s3_class(mainland_box, "sf")
[…]
