+++ NAMESPACE
@@ -58,7 +58,6 @@ importFrom(sf,st_read)
 importFrom(sf,st_relate)
 importFrom(sf,st_set_crs)
 importFrom(sf,st_transform)
-importFrom(sf,st_within)
 importFrom(sf,st_zm)
 importFrom(stars,st_as_stars)
 importFrom(stats,dist)
