### 2024-07-10 740749d
0.7.6 - `.check_par_spatraster`: SpatRaster input check in `par_*` main functions - Tests are being reorganized - Documentation revisions
