### 2024-07-11 #638 @beatrizmilz
> We are working on addressing the review comments. We will post our responses as soon as we finish the preparation, which is expected in 1-2 weeks. Thank you very much.

Hi @sigmafelix !  [This message](https://github.com/ropensci/software-review/issues/638#issuecomment-2217441336) was written by the Bot 🤖 . I know you are working on the changes! Thank you!
