### 2024-07-15 eae58a8
0.7.6 update - par_* function-data class alignment   - sf functions should use sf objects in x and y   - terra functions should use terra objects   - characters are tried inside parallel workers   - TODO: terra functions/objects should be converted/exported before putting into parallel workers; wrap-unwrap or abort? - Test revision in progress - TODO: add explicit loading of "chopin" in all par_* tests
