+++ tests/testthat/test-distribute_suites.R
@@ -1,13 +1,16 @@
+# testthat::skip_on_os()
 
-testthat::test_that("par_grid", {
+testthat::test_that("par_grid -- plain mode with raster path", {
   withr::local_package("terra")
   withr::local_package("sf")
   withr::local_package("future")
   withr::local_package("future.apply")
   withr::local_package("future.mirai")
   withr::local_package("dplyr")
-  withr::local_options(list(sf_use_s2 = FALSE))
-
+  withr::local_options(
+    list(sf_use_s2 = FALSE,
+         future.plan = "mirai_multisession")
+  )
   # Reading data
   ## NC counties polygon
   ncpath <- system.file("shape/nc.shp", package = "sf")
@@ -50,201 +53,331 @@ testthat::test_that("par_grid", {
       ny = 4L,
       padding = 3e4L
     )
-  res <-
-    suppressWarnings(
-      par_grid(
-        grids = nccompreg,
-        fun_dist = extract_at,
-        x = ncelev,
-        y = sf::st_as_sf(ncpnts),
-        qsegs = 90L,
-        radius = 5e3L,
-        id = "pid",
-        .debug = FALSE,
-        .standalone = FALSE
-      )
-    )
-
-
-  # check: sf <-> terra conversion changes coordinate precision?
-  # this result omits 2 points which are exactly on the boundary.
   testthat::expect_no_error({
-    resstr <-
+    res <-
       suppressWarnings(
         par_grid(
-          grids = NULL,
+          grids = nccompreg,
           fun_dist = extract_at,
-          points = test_fullpath,
-          surf = ncelev,
+          x = ncelevpath,
+          y = sf::st_as_sf(ncpnts),
           qsegs = 90L,
           radius = 5e3L,
-          id = "kid",
-          nx = 6L,
-          ny = 4L,
-          padding = 3e4L
+          id = "pid",
+          .debug = FALSE
         )
       )
   })
 
-  ncpntsf <- sf::st_as_sf(ncpnts)
-  testthat::expect_no_error(
-    resk <-
+  testthat::expect_no_error({
+    res <-
       suppressWarnings(
         par_grid(
           grids = nccompreg,
-          grid_target_id = NULL,
-          fun_dist = extract_at_buffer,
-          points = ncpntsf,
-          surf = ncelev,
+          fun_dist = extract_at,
+          x = ncelev,
+          y = sf::st_as_sf(ncpnts),
           qsegs = 90L,
           radius = 5e3L,
-          id = "pid"
+          id = "pid",
+          .debug = FALSE
         )
       )
+  })
+
+})
+
+
+testthat::test_that("par_grid -- grid_advanced mode", {
+  withr::local_package("terra")
+  withr::local_package("sf")
+  withr::local_package("future")
+  withr::local_package("future.apply")
+  withr::local_package("future.mirai")
+  withr::local_package("dplyr")
+  withr::local_options(
+    list(sf_use_s2 = FALSE,
+         future.plan = "mirai_multisession")
   )
+  # Reading data
+  ## NC counties polygon
+  ncpath <- system.file("shape/nc.shp", package = "sf")
+  ncpoly <- terra::vect(ncpath) %>%
+    terra::project("EPSG:5070")
+
+  ## Bundled random points in NC
+  ncpnts <-
+    readRDS(system.file("extdata/nc_random_point.rds", package = "chopin"))
+  ncpnts <- terra::vect(ncpnts)
+  ncpnts <- terra::project(ncpnts, "EPSG:5070")
+
+  ## Resampled SRTM data in NC
[…]
