### 2024-07-17 1752167
0.7.6 update - .check_vector signature fix - par_multirasters tests - par_multirasters logic fix in data inputs
