+++ tests/testthat/test-distribute_suites.R
@@ -1,4 +1,3 @@
-# testthat::skip_on_os()
 
 testthat::test_that("par_grid -- plain mode with raster path", {
   withr::local_package("terra")
@@ -7,6 +6,7 @@ testthat::test_that("par_grid -- plain mode with raster path", {
   withr::local_package("future.apply")
   withr::local_package("future.mirai")
   withr::local_package("dplyr")
+  withr::local_package("chopin")
   withr::local_options(
     list(sf_use_s2 = FALSE,
          future.plan = "mirai_multisession")
@@ -95,6 +95,7 @@ testthat::test_that("par_grid -- grid_advanced mode", {
   withr::local_package("future.apply")
   withr::local_package("future.mirai")
   withr::local_package("dplyr")
+  withr::local_package("chopin")
   withr::local_options(
     list(sf_use_s2 = FALSE,
          future.plan = "mirai_multisession")
@@ -551,6 +552,7 @@ testthat::test_that("generic function should be parallelized properly", {
   withr::local_package("future")
   withr::local_package("future.apply")
   withr::local_package("dplyr")
+  withr::local_package("chopin")
   withr::local_options(list(sf_use_s2 = FALSE))
 
   # main test
@@ -612,33 +614,80 @@ testthat::test_that("generic function should be parallelized properly", {
 
 
 testthat::test_that(
-  "Processes are properly spawned and compute over multirasters",
+  "par_multirasters -- character filenames, character y",
   {
     withr::local_package("terra")
     withr::local_package("sf")
     withr::local_package("future")
-    withr::local_package("mirai")
+    withr::local_package("future.mirai")
     withr::local_package("future.apply")
     withr::local_package("dplyr")
+    withr::local_package("chopin")
     withr::local_options(
       list(
         sf_use_s2 = FALSE,
-        future.plan = "multisession",
+        future.plan = "mirai_multisession",
         future.resolve.recursive = 2L
       )
     )
-
+    future::plan(future.mirai::mirai_multisession, workers = 2L)
     ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
     nccnty <- terra::vect(ncpath, layer = "county")
     ncelev <-
       system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
-    # terra::unwrap(
-    #   readRDS(
-    #     system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
-    #   )
-    # )
-    terra::crs(ncelev) <- "EPSG:5070"
-    names(ncelev) <- c("srtm15")
+    ncelev <- terra::rast(ncelev)
+    tdir <- tempdir(check = TRUE)
+    terra::writeRaster(ncelev, file.path(tdir, "test1.tif"), overwrite = TRUE)
+    terra::writeRaster(ncelev, file.path(tdir, "test2.tif"), overwrite = TRUE)
+    terra::writeRaster(ncelev, file.path(tdir, "test3.tif"), overwrite = TRUE)
+    terra::writeRaster(ncelev, file.path(tdir, "test4.tif"), overwrite = TRUE)
+    terra::writeRaster(ncelev, file.path(tdir, "test5.tif"), overwrite = TRUE)
+
+    testfiles <- list.files(tdir, pattern = "tif$", full.names = TRUE)
+    testthat::expect_no_error(
+      suppressWarnings(
+        res <- par_multirasters(
+          filenames = testfiles,
+          fun_dist = extract_at,
+          .debug = TRUE,
+          y = ncpath,
+          x = ncelev,
+          id = "GEOID",
+          func = "mean",
[…]
