### 2024-07-17 3a27b62
par_hierarchy bug fix - invalid object reference in reprojection part - Tests revised - S4 .check_vector edits for good practice
