+++ R/check.R
@@ -516,19 +516,47 @@ setMethod(
 NULL
 
 
+# nolint start
 setGeneric(
   ".check_vector",
-  function(input = NULL, input_id = NULL,
-           extent = NULL, out_class = NULL, ...) {
-    if (is.null(extent)) {
-      res <- .check_id(input, input_id)
-      return(res)
+  function(input, input_id = NULL, extent = NULL, out_class = character(1), ...) standardGeneric(".check_vector"),
+  signature = c("input", "input_id", "extent", "out_class")
+)
+# nolint end
+
+#' @keywords internal
+#' @noRd
+setMethod(
+  ".check_vector",
+  signature(input = "character", input_id = "ANY",
+            extent = "NULL", out_class = "character"),
+  function(input, input_id, extent, out_class, ...) {
+    if (!out_class %in% c("sf", "terra")) {
+      cli::cli_abort(c("out_class should be one of sf or terra.\n"))
+    }
+    input <- .check_id(input, input_id)
+
+    cli::cli_inform(
+      c("i" =
+          sprintf("Input is a character. Trying to read with %s\n.", out_class)
+      )
+    )
+    if (!is.null(extent)) {
+      cli::cli_alert_warning("Non-null extent is ignored.")
     }
 
-    extent <- .intersect_extent(extent, out_class)
-    input <- input[extent, ]
+    if (out_class == "sf") {
+      if (is.null(extent)) {
+        extent <- character(0)
+      }
+    }
 
-    .check_id(input, input_id)
+    input <- switch(
+      out_class,
+      terra = try(terra::vect(input, extent = extent), silent = TRUE),
+      sf = try(sf::st_read(input, wkt_filter = extent), silent = TRUE)
+    )
+    return(input)
   }
 )
 
@@ -537,11 +565,13 @@ setGeneric(
 setMethod(
   ".check_vector",
   signature(input = "character", input_id = "ANY",
-            extent = "ANY", out_class = "character"),
-  function(input, input_id = NULL, extent = NULL, out_class = "terra", ...) {
+            extent = "numeric", out_class = "character"),
+  function(input, input_id, extent, out_class, ...) {
     if (!out_class %in% c("sf", "terra")) {
       cli::cli_abort(c("out_class should be one of sf or terra.\n"))
     }
+    input <- .check_id(input, input_id)
+
     cli::cli_inform(
       c("i" =
           sprintf("Input is a character. Trying to read with %s\n.", out_class)
@@ -570,24 +600,30 @@ setMethod(
 #' @noRd
 setMethod(
   ".check_vector",
-  signature(input = "sf", input_id = "ANY",
-            extent = "numeric", out_class = "character"),
-  function(input, input_id = NULL, extent = NULL, out_class = "terra", ...) {
+  signature(input = "character", input_id = "ANY",
+            extent = "sf", out_class = "character"),
+  function(input, input_id, extent, out_class, ...) {
     if (!out_class %in% c("sf", "terra")) {
       cli::cli_abort(c("out_class should be one of sf or terra.\n"))
     }
-    if (out_class == "terra") {
-      input <- terra::vect(input)
-      if (!is.null(extent)) {
-        extent <- terra::ext(extent)
+    input <- .check_id(input, input_id)
+
+    cli::cli_inform(
+      c("i" =
+          sprintf("Input is a character. Trying to read with %s\n.", out_class)
+      )
+    )
+    if (out_class == "sf") {
+      input <- try(sf::st_read(input), silent = TRUE)
+      if (sf::st_crs(input) != sf::st_crs(extent)) {
[…]
