### 2024-07-22 552fd9e
test update - .check_vector - Removed deprecated function tests - par_hierarchy crs error debugging in progress
