+++ R/check.R
@@ -246,7 +246,7 @@ reproject_to_raster <-
 
 #' Validate and repair input vector data
 #' @family Helper functions
-#' @keywords internal soft-validated
+#' @keywords internal soft-deprecated
 #' @description It tries repairing input vector data.
 #' Vector validity violation usually appears in polygon data with
 #' self-crossing or
@@ -532,13 +532,12 @@ setMethod(
             extent = "NULL", out_class = "character"),
   function(input, input_id, extent, out_class, ...) {
     if (!out_class %in% c("sf", "terra")) {
-      cli::cli_abort(c("out_class should be one of sf or terra.\n"))
+      cli::cli_abort(c("out_class should be one of sf or terra."))
     }
-    input <- .check_id(input, input_id)
 
     cli::cli_inform(
       c("i" =
-          sprintf("Input is a character. Trying to read with %s\n.", out_class)
+          sprintf("Input is a character. Trying to read with %s.", out_class)
       )
     )
     if (!is.null(extent)) {
@@ -556,6 +555,8 @@ setMethod(
       terra = try(terra::vect(input, extent = extent), silent = TRUE),
       sf = try(sf::st_read(input, wkt_filter = extent), silent = TRUE)
     )
+    .check_id(input = input, input_id = input_id)
+
     return(input)
   }
 )
@@ -568,13 +569,12 @@ setMethod(
             extent = "numeric", out_class = "character"),
   function(input, input_id, extent, out_class, ...) {
     if (!out_class %in% c("sf", "terra")) {
-      cli::cli_abort(c("out_class should be one of sf or terra.\n"))
+      cli::cli_abort(c("out_class should be one of sf or terra."))
     }
-    input <- .check_id(input, input_id)
 
     cli::cli_inform(
       c("i" =
-          sprintf("Input is a character. Trying to read with %s\n.", out_class)
+          sprintf("Input is a character. Trying to read with %s.", out_class)
       )
     )
     if (out_class == "sf") {
@@ -591,6 +591,7 @@ setMethod(
       terra = try(terra::vect(input, extent = extent), silent = TRUE),
       sf = try(sf::st_read(input, wkt_filter = extent), silent = TRUE)
     )
+    .check_id(input = input, input_id = input_id)
     return(input)
   }
 )
@@ -606,7 +607,6 @@ setMethod(
     if (!out_class %in% c("sf", "terra")) {
       cli::cli_abort(c("out_class should be one of sf or terra.\n"))
     }
-    input <- .check_id(input, input_id)
 
     cli::cli_inform(
       c("i" =
@@ -626,6 +626,7 @@ setMethod(
         extent <- reproject_std(extent, terra::crs(input))
       }
     }
+    .check_id(input = input, input_id = input_id)
     input <- input[extent, ]
     return(input)
   }
@@ -640,13 +641,12 @@ setMethod(
             extent = "SpatVector", out_class = "character"),
   function(input, input_id, extent, out_class, ...) {
     if (!out_class %in% c("sf", "terra")) {
-      cli::cli_abort(c("out_class should be one of sf or terra.\n"))
+      cli::cli_abort(c("out_class should be one of sf or terra."))
     }
-    input <- .check_id(input, input_id)
 
     cli::cli_inform(
       c("i" =
-          sprintf("Input is a character. Trying to read with %s\n.", out_class)
+          sprintf("Input is a character. Trying to read with %s.", out_class)
       )
     )
     if (out_class == "sf") {
@@ -662,6 +662,8 @@ setMethod(
         extent <- reproject_std(extent, terra::crs(input))
       }
     }
+    .check_id(input = input, input_id = input_id)
+
[…]
