+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: chopin
 Title: Computation of spatial data by hierarchical and objective partitioning of inputs for parallel processing
-Version: 0.7.6.20240716
+Version: 0.7.6.20240721
 Authors@R: c(
     person("Insang", "Song", , "geoissong@gmail.com", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0001-8732-3256")),

[…]
