### 2024-07-22 c6fc92f
0.7.6 update (07222024) - .check_vector logic fix - par_hierarchy object reference fix - typo fix
