### 2024-07-23 090db4d
0.7.7 - par_hierarchy refactoring: two subregion filtering strategies (substring/intersect)   - with relevant messaging - tests are fixed where old arguments are included
