+++ tests/testthat/test-distribute_suites.R
@@ -1,3 +1,4 @@
+### .par_screen tests ####
 testthat::test_that(".par_screen -- vector", {
   withr::local_package("terra")
   withr::local_package("sf")
@@ -173,7 +174,7 @@ testthat::test_that(".check_character to .par_screen: try-error inputs", {
 
 
 
-
+### par_grid tests ####
 testthat::test_that("par_grid -- plain mode with raster path", {
   withr::local_package("terra")
   withr::local_package("sf")
@@ -537,6 +538,8 @@ testthat::test_that("par_grid -- par_pad_balanced", {
 })
 
 
+
+### par_hierarchy tests ####
 testthat::test_that(
   "par_hierarchy -- raster path input and spatraster fallback ", {
     withr::local_package("terra")
@@ -580,7 +583,9 @@ testthat::test_that(
           y = nctrct,
           x = ncelevpath,
           id = "GEOID",
-          func = "mean"
+          func = "mean",
+          pad_y = F,
+          .standalone = FALSE
         )
     )
     testthat::expect_true(is.data.frame(residb))
@@ -693,7 +698,8 @@ testthat::test_that("par_hierarchy: multicore-generic function dispatch", {
 
   ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
   nccnty <- terra::vect(ncpath, layer = "county")
-  nctrct <- terra::vect(ncpath, layer = "tracts")
+  nctrct <- sf::st_read(ncpath, layer = "tracts")
+  nctrct <- terra::vect(nctrct)
   ncelevpath <- system.file("extdata/nc_srtm15_otm.tif", package = "chopin")
   ncelev <- terra::rast(ncelevpath)
   ncroad <- system.file("extdata/ncroads_first.gpkg", package = "chopin")
@@ -710,32 +716,45 @@ testthat::test_that("par_hierarchy: multicore-generic function dispatch", {
   nctrctc <- terra::centroids(nctrct)
   ncroadv <- terra::vect(ncroad)
 
-  # straightforward error case
-  # invalid usage of fun_dist
-  # halted at the first error
-  testthat::expect_error(
-    suppressWarnings(
-      resnas <-
-        par_hierarchy(
-          regions = nccnty,
-          regions_id = "GEOID",
-          input_id = "GEOID",
-          pad_y = TRUE,
-          .debug = TRUE,
-          fun_dist = nearest,
-          x = nctrctc,
-          y = ncroadv
-        )
+  # no errors since 100km buffer is enough to capture
+  # nearest road for coastal tracts
+  resnas0 <-
+    par_hierarchy(
+      regions = nccnty,
+      regions_id = "GEOID",
+      pad_y = TRUE,
+      pad = 50000,
+      .debug = TRUE,
+      fun_dist = nearest,
+      x = nctrctc,
+      y = ncroadv
     )
-  )
 
+
+  # no errors since 100km buffer is enough to capture
+  # nearest road for coastal tracts
+  resnas <-
+    par_hierarchy(
+      regions = nccnty,
+      regions_id = "GEOID",
+      pad_y = TRUE,
+      pad = 100000,
+      .debug = TRUE,
+      fun_dist = nearest,
+      x = nctrctc,
+      y = ncroadv
+    )
+
+  # resnas0 and resnas must have different #rows
+  testthat::expect_true(nrow(resnas) > nrow(resnas0))
+
+  # regions are sf object
   nctrcc <- terra::centroids(nctrct)
   testthat::expect_no_error(
     resnasx <-
       par_hierarchy(
         regions = sf::st_as_sf(nccnty),
         regions_id = "GEOID",
-        input_id = "GEOID",
         pad_y = FALSE,
         fun_dist = extract_at,
         x = ncelev,
[…]
