+++ tests/testthat/test-distribute_suites.R
@@ -119,60 +119,6 @@ testthat::test_that(".par_screen -- raster", {
 })
 
 
-testthat::test_that(".check_character to .par_screen: try-error inputs", {
-  withr::local_package("terra")
-  withr::local_package("sf")
-  withr::local_package("dplyr")
-  withr::local_package("chopin")
-  withr::local_options(
-    list(sf_use_s2 = FALSE)
-  )
-  # Reading data
-  bundlevec <- system.file("gpkg/nc.gpkg", package = "sf")
-  bundleras <- system.file("ex/elev.tif", package = "terra")
-
-  # read file into terra objects
-  bundlevect <- terra::vect(bundlevec)
-  bundlerast <- terra::rast(bundleras)
-
-  chk_bundlevect <- .check_character(bundlevect)
-  chk_bundlerast <- .check_character(bundlerast)
-
-
-  # sf (as is)
-  scr_sfsf <-
-    .par_screen(
-      type = "raster",
-      input = scr_sf,
-      input_id = "FIPS",
-      out_class = "sf"
-    )
-  testthat::expect_s3_class(scr_sfsf, "sf")
-  testthat::expect_s4_class(scr_sfterra, "SpatVector")
-
-  # terra into sf
-  scr_terrasf <-
-    .par_screen(
-      type = "raster",
-      input = scr_terra,
-      input_id = "FIPS",
-      out_class = "sf"
-    )
-  # terra (as is)
-  scr_terraterra <-
-    .par_screen(
-      type = "raster",
-      input = scr_terra,
-      input_id = "FIPS",
-      out_class = "terra"
-    )
-
-  testthat::expect_s3_class(scr_terrasf, "sf")
-  testthat::expect_s4_class(scr_terraterra, "SpatVector")
-
-})
-
-
 
 ### par_grid tests ####
 testthat::test_that("par_grid -- plain mode with raster path", {
@@ -276,6 +222,7 @@ testthat::test_that("par_grid -- grid_advanced mode", {
     list(sf_use_s2 = FALSE,
          future.plan = "mirai_multisession")
   )
+  withr::local_seed(202407)
   # Reading data
   ## NC counties polygon
   ncpath <- system.file("shape/nc.shp", package = "sf")
@@ -386,6 +333,8 @@ testthat::test_that("par_grid -- grid_quantile mode", {
     list(sf_use_s2 = FALSE,
          future.plan = "mirai_multisession")
   )
+  withr::local_seed(202407)
+
   # Reading data
   ## NC counties polygon
   ncpath <- system.file("shape/nc.shp", package = "sf")
@@ -460,6 +409,7 @@ testthat::test_that("par_grid -- par_pad_balanced", {
          future.plan = "mirai_multisession",
          rlib_message_verbosity = "warning")
   )
+  withr::local_seed(202407)
 
   # Reading data
   ## NC counties polygon
@@ -556,6 +506,7 @@ testthat::test_that(
         future.plan = "mirai_multisession"
       )
     )
+    withr::local_seed(202407)
 
     ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
     nccnty <- sf::st_read(ncpath, layer = "county")
@@ -629,6 +580,7 @@ testthat::test_that("par_hierarchy: multicore-SpatRaster input", {
       future.resolve.recursive = 2L
     )
   )
+  withr::local_seed(202407)
   future::plan(future::multicore, workers = 2L)
 
   ncpath <- system.file("extdata/nc_hierarchy.gpkg", package = "chopin")
[…]
