### 2024-07-26 d306a48
Local CMD check
