+++ R/preprocessing.R
@@ -123,19 +123,6 @@ clip_vec_ext <- function(
 #' @param extrusion numeric(1). Extrusion factor for the extent. Default is 1.1
 #' @returns A clipped `SpatRaster` object.
 #' @author Insang Song
-#' @examples
-#' library(terra)
-#'
-#' ras_rand <- terra::rast(nrow = 20, ncol = 20)
-#' terra::values(ras_rand) <- runif(400L)
-#' vec_rand_p <-
-#'   data.frame(
-#'     x = c(3, 5, 3.2, 8),
-#'     y = c(12, 10, 15, 12),
-#'     z = c(0, 1, 2, 3)
-#'   )
-#' ras_rand_p <- terra::vect(vec_rand_p, geom = c("x", "y"))
-#' clip_ras_ext(x = ras_rand, y = vec_rand_p, radius = 1.5)
 #' @importFrom terra vect crop
 clip_ras_ext <- function(
   x = NULL,
