+++ DESCRIPTION
@@ -37,11 +37,9 @@ Imports:
     dplyr (>= 1.1.0),
     exactextractr (>= 0.8.2),
     future,
-    future.mirai,
     future.apply,
     igraph,
     methods,
-    mirai,
     rlang,
     sf (>= 1.0-10),
     stars (>= 0.6-0),
@@ -50,7 +48,10 @@ Imports:
     lifecycle
 Suggests: 
     covr,
+    targets,
     DiagrammeR,
+    future.mirai,
+    mirai,
     knitr,
     rmarkdown,
     spatstat.random,
