+++ NAMESPACE
@@ -10,7 +10,6 @@ export(par_pad_balanced)
 export(par_pad_grid)
 export(par_split_list)
 export(summarize_sedc)
-exportMethods("[")
 exportMethods(extract_at)
 exportMethods(summarize_aw)
 importFrom(anticlust,balanced_clustering)
