### 2024-07-27 9d5194c
check (cont'd) - Rd file structure - local R CMD check
