+++ R/check.R
@@ -6,6 +6,7 @@
 #' @param input Spat* in terra or sf object.
 #' @returns A character object; one of `"character"`, `"terra"` and `"sf"`
 #' @examples
+#' \dontrun{
 #' library(sf)
 #' library(terra)
 #' options(sf_use_s2 = FALSE)
@@ -15,6 +16,7 @@
 #' dep_check(nc_sf)
 #' nc_vect <- terra::vect(nc_sf)
 #' dep_check(nc_vect)
+#' }
 dep_check <- function(input) {
   if (
     !inherits(
@@ -113,6 +115,7 @@ dep_switch <- function(input) {
 #' @returns character(1). One of `"vector"` or `"raster"`.
 #' @importFrom cli cli_abort
 #' @examples
+#' \dontrun{
 #' library(sf)
 #' library(terra)
 #' options(sf_use_s2 = FALSE)
@@ -124,6 +127,7 @@ dep_switch <- function(input) {
 #' ra_path <- system.file("ex/elev.tif", package = "terra")
 #' ra <- terra::rast(ra_path)
 #' datamod(ra)
+#' }
 datamod <- function(input) {
   if (
     !inherits(
@@ -166,6 +170,7 @@ datamod <- function(input) {
 #' @returns A (reprojected) `sf` or `SpatVector` object.
 #' @author Insang Song
 #' @examples
+#' \dontrun{
 #' library(sf)
 #' library(terra)
 #' options(sf_use_s2 = FALSE)
@@ -177,6 +182,7 @@ datamod <- function(input) {
 #'
 #' nc_vect <- terra::vect(nc_sf)
 #' reproject_std(nc_vect, base_crs)
+#' }
 #' @importFrom sf st_crs st_transform
 #' @importFrom terra crs project
 reproject_std <-
@@ -219,6 +225,7 @@ reproject_std <-
 #' @returns Reprojected object in the same class as \code{vector}
 #' @author Insang Song
 #' @examples
+#' \dontrun{
 #' library(terra)
 #' library(sf)
 #' options(sf_use_s2 = FALSE)
@@ -228,6 +235,7 @@ reproject_std <-
 #' nc <- terra::vect(ncpath)
 #' elev <- terra::rast(elev)
 #' reproject_to_raster(nc, elev)
+#' }
 #' @importFrom sf st_transform
 #' @importFrom terra project
 #' @importFrom terra crs
@@ -260,12 +268,14 @@ reproject_to_raster <-
 #' the class of input_vector.
 #' @note This function works with GEOS (>=3.8).
 #' @examples
+#' \dontrun{
 #' library(terra)
 #' library(sf)
 #' ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
 #' nc <- terra::vect(ncpath)
 #'
 #' nc_valid <- vect_validate(nc)
+#' }
 #' @importFrom terra makeValid
 #' @importFrom sf st_make_valid
 vect_validate <- function(input_vector) {
@@ -281,17 +291,19 @@ vect_validate <- function(input_vector) {
 
 
 # ## .intersect_extent ####
+#' Get intersection extent
+#' @param input sf/SpatExtent/SpatVector/numeric
+#' @param out_class character(1). "sf" or "terra"
+#' @param ... other arguments. Placeholder.
+#' @name .intersect_extent
+#' @rdname .intersect_extent
 setGeneric(
   ".intersect_extent",
-  function(input = NULL, out_class = NULL, ...) {
-    if (is.null(input)) {
-      return(input)
-    }
-  }
[…]
