### 2024-07-30 9a07ee8
0.7.7 update - R CMD check (cont'd) - Documentation
