### 2024-07-31 568384d
0.7.7 (07312024) - par_grid & par_hierarchy: id argument check removed - README.md rebuilt - LICENSE.md is added to .Rbuildignore - LICENSE in DESRIPTION is changed to MIT - Minor changes in vignettes
