+++ README.md
@@ -1,5 +1,5 @@
 
-# `chopin`: Computation for Climate and Health research On Parallelized INfrastructure
+# `chopin`: Computation for Climate and Health research On Parallelized INfrastructure <img src="man/figures/chopin-logo.png" align="right" height="144" alt="overlapping irregular grid polygons filled with orange, green, and teal" /></a>
 
 <!-- badges: start -->
 
@@ -16,7 +16,7 @@ experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](h
 
 ### Objective
 
--   This package automates
+  - This package automates
     [parallelization](https://en.wikipedia.org/wiki/Parallel_computing)
     in spatial operations with `chopin` functions as well as
     [sf](https://github.com/r-spatial/sf)/[terra](https://github.com/rspatial/terra)
@@ -27,35 +27,35 @@ experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](h
 
 ### For whom `chopin` is useful
 
--   All who need to perform geospatial operations with large datasets
+  - All who need to perform geospatial operations with large datasets
     may find this package useful to accelerate the covariate calculation
     process for further analysis and modeling.
--   We assume that users–
-    -   Have basic knowledge of [geographic information system data
+  - We assume that users–
+      - Have basic knowledge of [geographic information system data
         models](https://r.geocompx.org/spatial-class), [coordinate
         systems and
-        transformations](https://gistbok.ucgis.org/bok-topics/coordinate-transformations),
+        transformations](https://doi.org/10.22224/gistbok/2023.1.2),
         [spatial operations](https://r.geocompx.org/spatial-operations),
         and [raster-vector
         overlay](https://r.geocompx.org/raster-vector);
-    -   Understood and planned **what they want to calculate**; and
-    -   Collected **datasets they need**
+      - Understood and planned **what they want to calculate**; and
+      - Collected **datasets they need**
 
 ## Overview
 
--   Processing functions accept
+  - Processing functions accept
     [terra](https://github.com/rspatial/terra)/[sf](https://github.com/r-spatial/sf)
     classes for spatial data. Raster-vector overlay is done with
     `exactextractr`.
--   From version 0.8.0, this package supports three basic functions that
+  - From version 0.8.0, this package supports three basic functions that
     are readily parallelized over multithread environments:
-    -   `extract_at`: extract raster values with point buffers or
+      - `extract_at`: extract raster values with point buffers or
         polygons with or without kernel weights
-    -   `summarize_sedc`: calculate sums of [exponentially decaying
+      - `summarize_sedc`: calculate sums of [exponentially decaying
         contributions](https://mserre.sph.unc.edu/BMElab_web/SEDCtutorial/index.html)
-    -   `summarize_aw`: area-weighted covariates based on target and
+      - `summarize_aw`: area-weighted covariates based on target and
         reference polygons
--   When processing points/polygons in parallel, the entire study area
+  - When processing points/polygons in parallel, the entire study area
[…]
