### 2024-08-07 f93fc1e
0.7.7 update - documentation (S4 documentation fix) - example error fix
