+++ man/par_merge_grid.Rd
@@ -55,7 +55,7 @@ scale = 15000, type = "Thomas")
 sf::st_crs(dg_sample) <- sf::st_crs(dg)
 dg_merged <- par_merge_grid(sf::st_as_sf(dg_sample), dgs, 100)
 
-plot(dg_merged$geometry)
+plot(sf::st_geometry(dg_merged))
 }
 \references{
 \itemize{
