+++ NAMESPACE
@@ -12,6 +12,7 @@ export(par_pad_grid)
 export(par_split_list)
 export(summarize_aw)
 export(summarize_sedc)
+exportMethods("[")
 exportMethods(extract_at)
 exportMethods(summarize_aw)
 importFrom(anticlust,balanced_clustering)
