### 2024-08-07 8695a66
remove /docs - test-cov yaml to upload /tmp for inspection
