+++ docs/_pkgdown.yml
@@ -1,10 +1,0 @@
-destination: ../../docs/
-
-url: ~
-template:
-  bootstrap: 5
-
-navbar:
-  structure:
-    left:  [reference, articles, search]
-    right: [github]
