### 2024-08-07 8ff99b5
test failure fix - future.plan in options() for multicore seems not to work properly: explicit plan() definition - _pkgdown.yml: missing function indices
