### 2024-08-07 cbc9401
Windows checks removed - Local Windows vignette builds did not reproduce the errors from GitHub action runners - Always upload artifacts in test-coverage-local
