+++ .github/workflows/test-coverage-local.yaml
@@ -80,6 +80,7 @@ jobs:
           ref: ${{ github.ref }} 
 
       - name: Upload artifacts
+        if: always()
         uses: actions/upload-artifact@v4
         with:
           name: test-outputs
