+++ .github/workflows/test-coverage-local.yaml
@@ -39,7 +39,7 @@ jobs:
 
       - name: Test coverage
         run: |
-          Rscript -e "covpack<-covr::package_coverage(install_path='${{ github.workspace}}'); \
+          Rscript -e "covpack<-covr::package_coverage(install_path='${{ github.workspace }}/cov', clean=FALSE); \
           covd<-covr::coverage_to_list(covpack)$totalcoverage; \
           write.table(covd[length(covd)], file = '${{ github.workspace }}/local_cov.Rout', row.names = F, col.names = F)"
         shell: bash
@@ -86,4 +86,4 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: test-outputs
-          path: ${{ github.workspace }}
+          path: ${{ github.workspace }}/cov
