+++ DESCRIPTION
@@ -48,6 +48,7 @@ Imports:
     lifecycle
 Suggests: 
     covr,
+    devtools,
     targets,
     DiagrammeR,
     future.mirai,
