### 2024-08-09 7512be6
alternative installation for tests - added skip_on_ci() in internal tests - explicit temporary installation for tests
