+++ tests/testthat.R
@@ -8,6 +8,13 @@
 
 library(testthat)
 library(devtools)
+temp_lib <- tempdir()
+devtools::install(pkg = "..", lib = temp_lib)
+
+# Set the library path for the session
+.libPaths(temp_lib)
+
+
 library(chopin)
 
 test_check("chopin")
