+++ tests/testthat.R
@@ -8,12 +8,11 @@
 
 library(testthat)
 library(devtools)
-temp_lib <- tempdir()
-devtools::install(pkg = ".", lib = temp_lib)
+# temp_lib <- tempdir()
+# devtools::install(pkg = basename("."), lib = temp_lib)
 
-# Set the library path for the session
-.libPaths(temp_lib)
-
+# # Set the library path for the session
+# .libPaths(temp_lib)
 
 library(chopin)
 
