### 2024-08-09 964b180
self local_package removed
