+++ tests/testthat/test-check.R
@@ -4,7 +4,7 @@ testthat::test_that("Input object class is detected",
     testthat::skip_on_ci()
     withr::local_package("stars")
     withr::local_package("terra")
-    withr::local_package("chopin")
+    # withr::local_package("chopin")
     withr::local_options(list(sf_use_s2 = FALSE))
     bcsd_path <- system.file(package = "stars", "nc/bcsd_obs_1999.nc")
     bcsd_stars <- stars::read_stars(bcsd_path)
@@ -31,7 +31,7 @@ testthat::test_that("Data model detection",
     testthat::skip_on_ci()
     withr::local_package("stars")
     withr::local_package("terra")
-    withr::local_package("chopin")
+    # withr::local_package("chopin")
     withr::local_options(list(sf_use_s2 = FALSE))
     bcsd_path <- system.file(package = "stars", "nc/bcsd_obs_1999.nc")
     bcsd_stars <- stars::read_stars(bcsd_path)
@@ -53,7 +53,7 @@ testthat::test_that("CRS is transformed when it is not standard", {
   testthat::skip_on_ci()
   withr::local_package("sf")
   withr::local_package("terra")
-  withr::local_package("chopin")
+  # withr::local_package("chopin")
   withr::local_options(list(sf_use_s2 = FALSE))
 
   nc <- system.file(package = "sf", "shape/nc.shp")
@@ -88,7 +88,7 @@ testthat::test_that("reproject to raster: sf", {
   testthat::skip_on_ci()
   withr::local_package("sf")
   withr::local_package("terra")
-  withr::local_package("chopin")
+  # withr::local_package("chopin")
   withr::local_options(list(sf_use_s2 = FALSE))
 
   nc <- system.file(package = "sf", "shape/nc.shp")
@@ -112,7 +112,7 @@ testthat::test_that("vector validity check is cleared", {
   testthat::skip_on_ci()
   withr::local_package("sf")
   withr::local_package("terra")
-  withr::local_package("chopin")
+  # withr::local_package("chopin")
   withr::with_dir(testthat::test_path("../.."), devtools::load_all())
   withr::local_options(list(sf_use_s2 = FALSE))
 
