### 2024-08-09 daa8bce
modified .check_package - testthat.R will explicitly load terra and sf
