+++ R/check.R
@@ -1124,24 +1124,39 @@ setMethod(
 #' Check the parent package of a function
 #' @param fun character(1). Function name
 #' @returns character(1). Package name. Only one of "sf", "terra", "chopin"
-#' @importFrom utils find
+# @importFrom utils find
 #' @noRd
-.check_package <-
-  function(fun) {
-    funname <- find(fun)
-    pkgname <- gsub("package:", "", funname)
-    pkgname <- grep("terra|sf|chopin", pkgname, value = TRUE)
-    if (length(pkgname) == 0) {
-      cli::cli_abort("No parent package is found.\n")
-    }
-    if (length(pkgname) > 1) {
-      cli::cli_abort("There are multiple parent packages matched.\n")
-    }
-    if (!pkgname %in% c("sf", "terra", "chopin")) {
-      cli::cli_abort("Function should be one from sf, terra, or chopin.\n")
-    }
-    return(pkgname)
+.check_package <- function(fun) {
+  fun_env <- environment(fun)
+  if (is.null(fun_env)) {
+    cli::cli_abort("Function environment not found.\n")
   }
+
+  pkgname <- environmentName(fun_env)
+  message("Detected package: ", pkgname)  # Debugging line
+
+  if (!pkgname %in% c("sf", "terra", "chopin")) {
+    cli::cli_abort("Function should be one from sf, terra, or chopin.\n")
+  }
+  return(pkgname)
+}
+
+# .check_package <-
+#   function(fun) {
+#     funname <- find(fun)
+#     pkgname <- gsub("package:", "", funname)
+#     pkgname <- grep("terra|sf|chopin", pkgname, value = TRUE)
+#     if (length(pkgname) == 0) {
+#       cli::cli_abort("No parent package is found.\n")
+#     }
+#     if (length(pkgname) > 1) {
+#       cli::cli_abort("There are multiple parent packages matched.\n")
+#     }
+#     if (!pkgname %in% c("sf", "terra", "chopin")) {
+#       cli::cli_abort("Function should be one from sf, terra, or chopin.\n")
+#     }
+#     return(pkgname)
+#   }
 
 
 #' Check the alignment of a function and the input objects
