+++ tests/testthat.R
@@ -13,7 +13,9 @@ library(devtools)
 
 # # Set the library path for the session
 # .libPaths(temp_lib)
-
+library(terra)
+library(sf)
 library(chopin)
+options(sf_use_s2 = FALSE)
 
 test_check("chopin")
