+++ NAMESPACE
@@ -88,4 +88,5 @@ importFrom(terra,rast)
 importFrom(terra,set.crs)
 importFrom(terra,vect)
 importFrom(utils,combn)
+importFrom(utils,find)
 importFrom(utils,packageVersion)
