### 2024-08-09 47510dc
explicit package loading in .check_package - another try
