### 2024-08-09 774e706
refactor .check_package - No more loaded package search - Still seeing package_coverage() failures in local - par_* main functions is reflown with if-else statements for package names
