+++ tests/testthat/test-check.R
@@ -2,6 +2,7 @@
 testthat::test_that("Input object class is detected",
   {
     testthat::skip_on_ci()
+    testthat::skip_on_covr()
     withr::local_package("stars")
     withr::local_package("terra")
     # withr::local_package("chopin")
@@ -9,7 +10,7 @@ testthat::test_that("Input object class is detected",
     bcsd_path <- system.file(package = "stars", "nc/bcsd_obs_1999.nc")
     bcsd_stars <- stars::read_stars(bcsd_path)
 
-    packbound_stars <- dep_check(bcsd_stars)
+    packbound_stars <- chopin:::dep_check(bcsd_stars)
     sprast_bcsd <- terra::rast(bcsd_path)
     packbound_terra <- dep_check(sprast_bcsd)
 
@@ -19,7 +20,7 @@ testthat::test_that("Input object class is detected",
     # proxy
     ncpath <- system.file(package = "sf", "shape/nc.shp")
     nc <- terra::vect(ncpath, proxy = TRUE)
-    packbound_proxy <- dep_check(nc)
+    packbound_proxy <- chopin:::dep_check(nc)
     testthat::expect_equal(packbound_proxy, "terra")
 
   }
@@ -29,6 +30,7 @@ testthat::test_that("Input object class is detected",
 testthat::test_that("Data model detection",
   {
     testthat::skip_on_ci()
+    testthat::skip_on_covr()
     withr::local_package("stars")
     withr::local_package("terra")
     # withr::local_package("chopin")
@@ -51,6 +53,7 @@ testthat::test_that("Data model detection",
 
 testthat::test_that("CRS is transformed when it is not standard", {
   testthat::skip_on_ci()
+  testthat::skip_on_covr()
   withr::local_package("sf")
   withr::local_package("terra")
   # withr::local_package("chopin")
@@ -86,6 +89,7 @@ testthat::test_that("CRS is transformed when it is not standard", {
 
 testthat::test_that("reproject to raster: sf", {
   testthat::skip_on_ci()
+  testthat::skip_on_covr()
   withr::local_package("sf")
   withr::local_package("terra")
   # withr::local_package("chopin")
@@ -110,6 +114,7 @@ testthat::test_that("reproject to raster: sf", {
 
 testthat::test_that("vector validity check is cleared", {
   testthat::skip_on_ci()
+  testthat::skip_on_covr()
   withr::local_package("sf")
   withr::local_package("terra")
   # withr::local_package("chopin")
@@ -128,6 +133,7 @@ testthat::test_that("vector validity check is cleared", {
 
 testthat::test_that(".check_id throws error with non-character id", {
   testthat::skip_on_ci()
+  testthat::skip_on_covr()
   withr::local_package("sf")
   withr::local_options(list(sf_use_s2 = FALSE))
   withr::with_dir(testthat::test_path("../.."), devtools::load_all())
@@ -142,6 +148,7 @@ testthat::test_that(".check_id throws error with non-character id", {
 
 testthat::test_that(".check_id throws error with nonexistent id", {
   testthat::skip_on_ci()
+  testthat::skip_on_covr()
   withr::local_package("sf")
   withr::with_dir(testthat::test_path("../.."), devtools::load_all())
   withr::local_options(list(sf_use_s2 = FALSE))
@@ -157,6 +164,7 @@ testthat::test_that(".check_id throws error with nonexistent id", {
 
 testthat::test_that(".check_vector with sf", {
   testthat::skip_on_ci()
+  testthat::skip_on_covr()
   withr::with_dir(testthat::test_path("../.."), devtools::load_all())
[…]
