+++ tools/test_init.r
@@ -5,8 +5,6 @@ library(devtools)
 library(future.mirai)
 plan(mirai_multisession, workers = 2L)
 
-load_all()
-
 
 ## reflect changes
 install(quick=T)
