+++ NAMESPACE
@@ -37,6 +37,7 @@ importFrom(igraph,components)
 importFrom(igraph,graph_from_edgelist)
 importFrom(igraph,mst)
 importFrom(lifecycle,deprecated)
+importFrom(methods,findFunction)
 importFrom(methods,getPackageName)
 importFrom(rlang,"!!!")
 importFrom(rlang,inject)

[…]
