+++ pkgdown.yml
@@ -1,12 +1,13 @@
 pandoc: 3.1.11
-pkgdown: 2.0.9
+pkgdown: 2.1.0
 pkgdown_sha: ~
 articles:
-  v00_good_practice_parallelization: v00_good_practice_parallelization.html
-  v01_par_make_gridset: v01_par_make_gridset.html
-  v02_climate_examples: v02_climate_examples.html
-last_built: 2024-05-30T19:46Z
+  v01_start: v01_start.html
+  v02_good_practice: v02_good_practice.html
+  v03_par_pad_grid: v03_par_pad_grid.html
+  v04_climate_examples: v04_climate_examples.html
+  v05_targets: v05_targets.html
+last_built: 2024-08-09T19:49Z
 urls:
-  reference: https://kyle-messier.github.io/chopin/reference
-  article: https://kyle-messier.github.io/chopin/articles
-
+  reference: https://NIEHS.github.io/chopin/reference
+  article: https://NIEHS.github.io/chopin/articles
