### 2024-08-09 8520688
0.7.8 - Late reflection of .check_package refactoring - Logo file replaced - Test coverage YAML is set to upload artifacts with covr::report() output - README rebuilt
