+++ tests/testthat/test-processing.R
@@ -121,7 +121,64 @@ testthat::test_that("Character input works", {
 
 })
 
-testthat::test_that("summarize_aw works as expected.", {
+
+
+testthat::test_that("summarize_aw works -- character-character", {
+  withr::local_package("sf")
+  withr::local_package("terra")
+  withr::local_package("units")
+  withr::local_package("dplyr")
+  withr::local_package("testthat")
+  withr::local_options(list(sf_use_s2 = FALSE))
+
+  ncpath <- system.file("shape/nc.shp", package = "sf")
+  nc <- sf::st_read(ncpath)
+  nc <- sf::st_transform(nc, 5070)
+  pp <- sf::st_sample(nc, size = 300)
+  pp <- sf::st_as_sf(pp)
+  pp[["id"]] <- seq(1, nrow(pp))
+  sf::st_crs(pp) <- "EPSG:5070"
+  flds <- c("BIR74", "SID74", "BIR79", "SID79")
+  ppb <- sf::st_buffer(pp, nQuadSegs = 180, dist = units::set_units(20, "km"))
+  ppbpath <- file.path(tempdir(), "ppb.shp")
+  sf::st_write(ppb, ppbpath, append = FALSE)
+
+  system.time({
+    ppb_nc_aw <- summarize_aw(ppbpath, ncpath, target_fields = flds, "id")
+  })
+  testthat::expect_true(inherits(ppb_nc_aw, "data.frame"))
+
+})
+
+
+testthat::test_that("summarize_aw works -- sf-sf", {
+  withr::local_package("sf")
+  withr::local_package("terra")
+  withr::local_package("units")
+  withr::local_package("dplyr")
+  withr::local_package("testthat")
+  withr::local_options(list(sf_use_s2 = FALSE))
+
+  nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"))
+  nc <- sf::st_transform(nc, 5070)
+  pp <- sf::st_sample(nc, size = 300)
+  pp <- sf::st_as_sf(pp)
+  pp[["id"]] <- seq(1, nrow(pp))
+  sf::st_crs(pp) <- "EPSG:5070"
+  flds <- c("BIR74", "SID74", "BIR79", "SID79")
+  ppb <- sf::st_buffer(pp, nQuadSegs = 180, dist = units::set_units(20, "km"))
+
+  testthat::expect_warning(
+    system.time({
+      ppb_nc_aw <- summarize_aw(ppb, nc, target_fields = flds, "id")
+    }),
+    "st_interpolate_aw assumes attributes are constant or uniform over areas of x"
+  )
+  testthat::expect_true(inherits(ppb_nc_aw, "data.frame"))
+
+})
+
+testthat::test_that("summarize_aw works -- terra-terra", {
   withr::local_package("sf")
   withr::local_package("terra")
   withr::local_package("units")
@@ -156,10 +213,30 @@ testthat::test_that("summarize_aw works as expected.", {
   )
   testthat::expect_s3_class(ppb_nc_aw, "data.frame")
 
+})
+
+testthat::test_that("summarize_aw works -- error cases", {
+  withr::local_package("sf")
+  withr::local_package("terra")
+  withr::local_package("units")
+  withr::local_package("dplyr")
+  withr::local_package("testthat")
+  withr::local_options(list(sf_use_s2 = FALSE))
+
+  nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"))
+  nc <- sf::st_transform(nc, 5070)
+  pp <- sf::st_sample(nc, size = 300)
+  pp <- sf::st_as_sf(pp)
[…]
