+++ .github/workflows/test-coverage-local.yaml
@@ -40,7 +40,7 @@ jobs:
       - name: Test coverage
         run: |
           Rscript -e "covpack<-covr::package_coverage(install_path='${{ github.workspace }}/cov', clean=FALSE); \
-          covr::report(covpack, file = file.path('${{ github.workspace }}', 'chopin-coverage-report.html')) \
+          covr::report(covpack, file = file.path('${{ github.workspace }}', 'chopin-coverage-report.html')); \
           covd<-covr::coverage_to_list(covpack)$totalcoverage; \
           write.table(covd[length(covd)], file = '${{ github.workspace }}/local_cov.Rout', row.names = F, col.names = F)"
         shell: bash
