### 2024-08-14 b1c7dd4
0.7.8 (08142024) - cov-report upload yaml - README figures - par_map_args test
