+++ tests/testthat/test-check.R
@@ -1,8 +1,7 @@
-
+# dep_check tests ####
 testthat::test_that("Input object class is detected",
   {
-    testthat::skip_on_ci()
-    testthat::skip_on_covr()
+    # testthat::skip_on_ci()
     withr::local_package("stars")
     withr::local_package("terra")
     # withr::local_package("chopin")
@@ -26,11 +25,11 @@ testthat::test_that("Input object class is detected",
   }
 )
 
-
+# datamod tests ####
 testthat::test_that("Data model detection",
   {
-    testthat::skip_on_ci()
-    testthat::skip_on_covr()
+    # testthat::skip_on_ci()
+    # testthat::skip_on_covr()
     withr::local_package("stars")
     withr::local_package("terra")
     # withr::local_package("chopin")
@@ -41,8 +40,8 @@ testthat::test_that("Data model detection",
     nc <- system.file(package = "sf", "shape/nc.shp")
     nc <- sf::read_sf(nc)
 
-    datatype_stars <- datamod(bcsd_stars)
-    datatype_sf <- datamod(nc)
+    datatype_stars <- chopin:::datamod(bcsd_stars)
+    datatype_sf <- chopin:::datamod(nc)
 
     testthat::expect_equal(datatype_stars, "raster")
     testthat::expect_equal(datatype_sf, "vector")
@@ -51,9 +50,10 @@ testthat::test_that("Data model detection",
   }
 )
 
+# reproject_std tests ####
 testthat::test_that("CRS is transformed when it is not standard", {
-  testthat::skip_on_ci()
-  testthat::skip_on_covr()
+  # testthat::skip_on_ci()
+  # testthat::skip_on_covr()
   withr::local_package("sf")
   withr::local_package("terra")
   # withr::local_package("chopin")
@@ -68,8 +68,8 @@ testthat::test_that("CRS is transformed when it is not standard", {
   sf::st_crs(ncna) <- NA
   ncnatr <- terra::vect(ncna)
 
-  nctr_align <- reproject_std(nctr, "EPSG:4326")
-  nc_align <- reproject_std(nc, "EPSG:4326")
+  nctr_align <- chopin:::reproject_std(nctr, "EPSG:4326")
+  nc_align <- chopin:::reproject_std(nc, "EPSG:4326")
 
   testthat::expect_s3_class(nc_align, "sf")
   testthat::expect_s4_class(nctr_align, "SpatVector")
@@ -82,14 +82,15 @@ testthat::test_that("CRS is transformed when it is not standard", {
 
   terra::crs(ncnatr) <- NULL
   # error case
-  testthat::expect_error(reproject_std(ncnatr, "EPSG:4326"))
+  testthat::expect_error(chopin:::reproject_std(ncnatr, "EPSG:4326"))
 
 })
 
 
+# reproject_to_raster tests ####
 testthat::test_that("reproject to raster: sf", {
-  testthat::skip_on_ci()
-  testthat::skip_on_covr()
+  # testthat::skip_on_ci()
+  # testthat::skip_on_covr()
   withr::local_package("sf")
   withr::local_package("terra")
   # withr::local_package("chopin")
@@ -104,17 +105,17 @@ testthat::test_that("reproject to raster: sf", {
   # make random raster
   rr <- terra::rast(matrix(runif(100), 10, 10), crs = "EPSG:4326")
 
-  ncr <- reproject_to_raster(nc, rr)
-  nctr <- reproject_to_raster(nctr, rr)
+  ncr <- chopin:::reproject_to_raster(nc, rr)
+  nctr <- chopin:::reproject_to_raster(nctr, rr)
 
   testthat::expect_s3_class(ncr, "sf")
   testthat::expect_s4_class(nctr, "SpatVector")
 })
 
-
+# vect_validate tests ####
[…]
