### 2024-08-16 e1b5583
test modified - ::: calls - yaml: Upload artifacts upon failure - mermaid static plot size adjusted
