+++ tests/testthat/test-check.R
@@ -151,7 +151,6 @@ testthat::test_that(".check_id throws error with nonexistent id", {
   # testthat::skip_on_ci()
   # testthat::skip_on_covr()
   withr::local_package("sf")
-  withr::with_dir(testthat::test_path("../.."), devtools::load_all())
   withr::local_options(list(sf_use_s2 = FALSE))
   input_char <- system.file("gpkg/nc.gpkg", package = "sf")
   input_sf <- sf::st_read(input_char)
@@ -167,7 +166,6 @@ testthat::test_that(".check_id throws error with nonexistent id", {
 testthat::test_that(".check_vector with sf", {
   # testthat::skip_on_ci()
   # testthat::skip_on_covr()
-  withr::with_dir(testthat::test_path("../.."), devtools::load_all())
   input_char <- system.file("gpkg/nc.gpkg", package = "sf")
   input_sf <- sf::st_read(input_char)
   checked_sf <-
@@ -183,7 +181,6 @@ testthat::test_that(".check_vector with sf", {
 testthat::test_that(".check_vector with terra", {
   # testthat::skip_on_ci()
   # testthat::skip_on_covr()
-  withr::with_dir(testthat::test_path("../.."), devtools::load_all())
   input_char <- system.file("gpkg/nc.gpkg", package = "sf")
   input_vect <- terra::vect(input_char)
   checked_vect <-
@@ -199,7 +196,6 @@ testthat::test_that(".check_vector with terra", {
 testthat::test_that(".check_vector with file path and extent", {
   # testthat::skip_on_ci()
   # testthat::skip_on_covr()
-  withr::with_dir(testthat::test_path("../.."), devtools::load_all())
   input_char <- system.file("gpkg/nc.gpkg", package = "sf")
   input_vect <- terra::vect(input_char)
 
@@ -305,18 +301,17 @@ testthat::test_that(".intersect_extent returns the intersection extent", {
 testthat::test_that(".check_character with non-character inputs",{
   # testthat::skip_on_ci()
   # testthat::skip_on_covr()
-  withr::with_dir(testthat::test_path("../.."), devtools::load_all())
   # test for non-character input
   testthat::expect_message(
-    testthat::expect_error(.check_character(3L)),
+    testthat::expect_error(chopin:::.check_character(3L)),
     "Input is not a character."
   )
   testthat::expect_message(
-    testthat::expect_error(.check_character(11.11)),
+    testthat::expect_error(chopin:::.check_character(11.11)),
     "Input is not a character."
   )
   testthat::expect_message(
-    testthat::expect_error(.check_character(NA)),
+    testthat::expect_error(chopin:::.check_character(NA)),
     "Input is not a character."
   )
 })
@@ -325,7 +320,6 @@ testthat::test_that(".check_character with non-character inputs",{
 testthat::test_that(".check_character with character inputs",{
   # testthat::skip_on_ci()
   # testthat::skip_on_covr()
-  withr::with_dir(testthat::test_path("../.."), devtools::load_all())
   withr::local_package("sf")
   withr::local_package("terra")
   withr::local_options(list(sf_use_s2 = FALSE))
@@ -335,14 +329,14 @@ testthat::test_that(".check_character with character inputs",{
   )
   # vector file
   ncfile <- system.file(package = "sf", "shape/nc.shp")
-  vec_file <- .check_character(ncfile)
+  vec_file <- chopin:::.check_character(ncfile)
   # remove attributes for comparison
[…]
