### 2024-08-17 fcff38b
drop with_dir() in tests
