### 2024-08-17 3e6cc46
crs() on stars object error fix
