+++ README.md
@@ -82,7 +82,36 @@ future::plan(future.mirai::mirai_multisession, workers = 4L)
 # the number of workers are up to users' choice
 ```
 
+### Function selection guide
+
+We provide two flowcharts to help users choose the right function for
+parallel processing. The raster-oriented flowchart is for users who want
+to start with raster data, and the vector-oriented flowchart is for
+users with large vector data.
+
+In **raster-oriented selection**, we suggest four factors to consider: -
+Number of raster files: for multiple files, `par_multirasters` is
+recommended. When there are multiple rasters that share the same extent
+and resolution, consider stacking the rasters into multilayer SpatRaster
+object by calling `terra::rast(filenames)`. - Raster resolution: We
+suggest 100 meters as a threshold. Rasters with resolution coarser than
+100 meters and a few layers would be better for the direct call of
+`exactextractr::exact_extract()`. - Raster extent: Using `SpatRaster` in
+`exactextractr::exact_extract()` is often minimally affected by the
+raster extent. - Memory size: `max_cells_in_memory` argument value of
+`exactextractr::exact_extract()`, raster resolution, and the number of
+layers in `SpatRaster` are multiplicatively related to the memory usage.
+
 <img src="man/figures/README-flowchart-mermaid-raster-1.png" width="100%" />
+
+For **vector-oriented selection**, we suggest three factors to
+consider: - Number of features: When the number of features is over
+100,000, consider using `par_grid` or `par_hierarchy` to split the data
+into smaller chunks. - Hierarchical structure: If the data has a
+hierarchical structure, consider using `par_hierarchy` to parallelize
+the operation. - Data grouping: If the data needs to be grouped in
+similar sizes, consider using `par_pad_balanced` or `par_pad_grid` with
+`mode = "grid_quantile"`.
 
 <img src="man/figures/README-flowchart-mermaid-vector-1.png" width="100%" />
 
@@ -207,7 +236,7 @@ system.time(
 )
 #> Input is a character. Attempt to read it with terra::rast...
 #>    user  system elapsed 
-#>   7.205   0.206   7.669
+#>   7.115   0.227   7.537
 ```
 
 #### Generate regular grid computational regions
@@ -298,7 +327,7 @@ system.time(
 #> Input is a character. Attempt to read it with terra::rast...
 #> ℹ Task at CGRIDID: 4 is successfully dispatched.
 #>    user  system elapsed 
-#>   0.329   0.021   8.337
+#>   0.360   0.010   8.089
 
 ncpoints_srtm <-
   extract_at(
@@ -358,7 +387,7 @@ path_nchrchy <- file.path(wdir, "nc_hierarchy.gpkg")
 nc_data <- path_nchrchy
 nc_county <- sf::st_read(nc_data, layer = "county")
 #> Reading layer `county' from data source 
-#>   `/tmp/RtmpkigVUL/temp_libpath3cdb0b6b55ee11/chopin/extdata/nc_hierarchy.gpkg' 
[…]
