+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   v03_par_pad_grid: v03_par_pad_grid.html
   v04_climate_examples: v04_climate_examples.html
   v05_targets: v05_targets.html
-last_built: 2024-08-17T03:39Z
+last_built: 2024-08-18T18:15Z
 urls:
   reference: https://NIEHS.github.io/chopin/reference
   article: https://NIEHS.github.io/chopin/articles
