### 2024-08-22 acebbc6
test update - TODO: S4 class review -- nonsensical arguments (i.e., "ANY", "NULL")
