+++ tests/testthat/test-check.R
@@ -161,9 +161,9 @@ testthat::test_that(".check_id throws error with nonexistent id", {
 
 
 # .check_vector tests ####
-testthat::test_that(".check_vector with sf", {
-  # testthat::skip_on_ci()
-  # testthat::skip_on_covr()
+# input = sf ####
+testthat::test_that(".check_vector with sf-NULL", {
+
   input_char <- system.file("gpkg/nc.gpkg", package = "sf")
   input_sf <- sf::st_read(input_char)
   checked_sf <-
@@ -173,27 +173,310 @@ testthat::test_that(".check_vector with sf", {
     )
   testthat::expect_equal(chopin:::dep_check(checked_sf), "sf")
 
-})
-
-
-testthat::test_that(".check_vector with terra", {
-  # testthat::skip_on_ci()
-  # testthat::skip_on_covr()
-  input_char <- system.file("gpkg/nc.gpkg", package = "sf")
-  input_vect <- terra::vect(input_char)
-  checked_vect <-
+  checked_terra <-
     chopin:::.check_vector(
-      input_vect, input_id = "FIPS",
+      input_sf, input_id = "FIPS",
       extent = NULL, out_class = "terra"
     )
-  testthat::expect_equal(chopin:::dep_check(checked_vect), "terra")
+  testthat::expect_equal(chopin:::dep_check(checked_terra), "terra")
+
+  testthat::expect_error(chopin:::.check_vector(
+    input = input_sf,
+    input_id = "FIPS",
+    extent = NULL,
+    out_class = "land"
+  ))
+})
+
+# input = sf ####
+testthat::test_that(".check_vector with sf-numeric", {
+  withr::local_package("sf")
+  withr::local_package("terra")
+  withr::local_options(list(sf_use_s2 = FALSE))
+
+  input_char <- system.file("gpkg/nc.gpkg", package = "sf")
+  input_sf <- sf::st_read(input_char)
+  num_ext <- c(-88, -80, 33, 36)
+  num_ext_bbox <- sf::st_as_sfc(sf::st_bbox(terra::ext(num_ext)))
+  num_ext_bbox <- sf::st_set_crs(num_ext_bbox, sf::st_crs(input_sf))
+  input_sf_sub <- input_sf[num_ext_bbox, ]
+
+  checked_sf <-
+    chopin:::.check_vector(
+      input_sf, input_id = "FIPS",
+      extent = num_ext, out_class = "sf"
+    )
+  testthat::expect_equal(chopin:::dep_check(checked_sf), "sf")
+  testthat::expect_true(nrow(input_sf_sub) == nrow(checked_sf))
+
+  checked_terra <-
+    chopin:::.check_vector(
+      input_sf, input_id = "FIPS",
+      extent = num_ext, out_class = "terra"
+    )
+  testthat::expect_equal(chopin:::dep_check(checked_terra), "terra")
+
+  testthat::expect_error(
+    chopin:::.check_vector(
+      input_sf, input_id = "FIPS",
+      extent = num_ext, out_class = "land"
+    )
+
+  )
+})
+
+
+testthat::test_that(".check_vector with sf-SpatVector", {
+  withr::local_package("sf")
+  withr::local_package("terra")
+  withr::local_options(list(sf_use_s2 = FALSE))
+
+  input_char <- system.file("gpkg/nc.gpkg", package = "sf")
+  name_filter <- c("Orange", "Durham", "Wake")
+  input_sf <- sf::st_read(input_char)
+  input_sfu <- sf::st_union(input_sf[input_sf$NAME %in% name_filter, ])
+  input_sfu <- sf::st_transform(input_sfu, "EPSG:5070")
[…]
