+++ man/extract_at.Rd
@@ -97,12 +97,12 @@ extract_at(x, y, ...)
   y = NULL,
   id = NULL,
   func = "mean",
-  extent,
-  radius,
+  extent = NULL,
+  radius = NULL,
   out_class = "sf",
   kernel = NULL,
   kernel_func = stats::weighted.mean,
-  bandwidth,
+  bandwidth = NULL,
   max_cells = 3e+07,
   .standalone = TRUE,
   ...
