### 2024-08-24 #638 @adamhsparks
Hi @sigmafelix, I'm just going through the open issues here as the current EIC to see what the current status may be. Did you have any updates on your comments to the review?
