### 2024-09-03 ac99f3b
0.8.0 - Reviewing review comments, I addressed several comments in this update. - Package documentation now includes caveats section - Bullet points were converted to ordinary paragraphs as much as possible
