### 2024-09-08 bef096c
0.8.1 - Prebuilt mermaid flowcharts: flowchart generation codes are put into another Rmd file - Adjusted logo size - V02 name changed: dropping "in HPC"
