+++ README.Rmd
@@ -14,7 +14,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-# Computation of Spatial Data by Hierarchical and Objective Partitioning of Inputs for Parallel Processing <img src="man/figures/chopin-logo.png" align="right" height="144" alt="overlapping irregular grid polygons filled with orange, green, and teal" /></a>
+# Computation of Spatial Data by Hierarchical and Objective Partitioning of Inputs for Parallel Processing <img src="man/figures/logo.png" align="right" height="210" alt="overlapping irregular grid polygons filled with orange, green, and teal" /></a>
 
 <!-- badges: start -->
 [![cov](https://NIEHS.github.io/chopin/badges/coverage.svg)](https://github.com/NIEHS/chopin/actions)
@@ -69,71 +69,16 @@ In **raster-oriented selection**, we suggest four factors to consider:
 - Raster extent: Using `SpatRaster` in `exactextractr::exact_extract()` is often minimally affected by the raster extent.
 - Memory size: `max_cells_in_memory` argument value of `exactextractr::exact_extract()`, raster resolution, and the number of layers in `SpatRaster` are multiplicatively related to the memory usage.
 
+![](man/figures/README-flowchart-raster.png)
 
-```{r flowchart-mermaid-raster, echo = FALSE, eval = (Sys.getenv("IN_GALLEY") == "")}
-mermaid_chart_raster <-
-'
-graph LR
-	n6695079["Is the spatial resolution finer than 100 meters?"]
-	n11509997["Are there multiple rasters?"]
-	n72001430["exact_extract with suitable max_cells_in_memory value"]
-	n27284812["Do they have the same extent and resolution?"]
-	n83137384["Is a single raster larger than your free memory space?"]
-	n83318893["Do you have memory larger than the total raster file size?"]
-	n14786842["exact_extract with low max_cells_in_memory"]
-	n17102479["exact_extract with high max_cells_in_memory argument value"]
-	n7037868["Stack rasters then process in the single thread"]
-	n58642837["par_multirasters"]
-	n6695079 -->|Yes| n11509997
-	n6695079 -->|No| n72001430
-	n11509997 -->|Yes| n27284812
-	n11509997 -->|No| n83137384
-	n27284812 -->|Yes| n83318893
-	n27284812 -->|No| n58642837
-	n83137384 -->|No| n14786842
-	n83137384 -->|Yes| n17102479
-	n83318893 -->|Yes| n7037868
-	n83318893 -->|No| n58642837
-'
-
-DiagrammeR::mermaid(mermaid_chart_raster, width = 1200, height = 400)
-```
 
 For **vector-oriented selection**, we suggest three factors to consider:
+
 - Number of features: When the number of features is over 100,000, consider using `par_grid` or `par_hierarchy` to split the data into smaller chunks.
 - Hierarchical structure: If the data has a hierarchical structure, consider using `par_hierarchy` to parallelize the operation.
 - Data grouping: If the data needs to be grouped in similar sizes, consider using `par_pad_balanced` or `par_pad_grid` with `mode = "grid_quantile"`.
 
-```{r flowchart-mermaid-vector, echo = FALSE, eval = (Sys.getenv("IN_GALLEY") == "")}
-mermaid_chart_vector <-
-'
-graph LR
-	n21640044["Are there 100K+ features in the input vectors?"]
-	n84295645["Are they hierarchical?"]
-	n82902796["single thread processing"]
-	n34878990["Are the data grouped in similar sizes?"]
[…]
