### 2024-09-08 5843eb9
README fix - logo path
