### 2024-09-09 2b19b0d
0.8.2 - renamed par_map_args to `par_convert_f` - clip_*_ext functions are removed - vect_validate function is removed - Description update for ellipsis arguments in documentation
