+++ tests/testthat/test-check.R
@@ -112,28 +112,9 @@ testthat::test_that("reproject to raster: sf", {
   testthat::expect_s4_class(nctr, "SpatVector")
 })
 
-# vect_validate tests ####
-testthat::test_that("vector validity check is cleared", {
-  # testthat::skip_on_ci()
-  # testthat::skip_on_covr()
-  withr::local_package("sf")
-  withr::local_package("terra")
-  # withr::local_package("chopin")
-  withr::local_options(list(sf_use_s2 = FALSE))
-
-  nc <- system.file(package = "sf", "shape/nc.shp")
-  nc <- sf::read_sf(nc)
-
-  testthat::expect_no_error(chopin:::vect_validate(nc))
-
-  nct <- terra::vect(nc)
-  testthat::expect_no_error(chopin:::vect_validate(nct))
-})
 
 # .check_id tests ####
 testthat::test_that(".check_id throws error with non-character id", {
-  # testthat::skip_on_ci()
-  # testthat::skip_on_covr()
   withr::local_package("sf")
   withr::local_options(list(sf_use_s2 = FALSE))
   input_char <- system.file("gpkg/nc.gpkg", package = "sf")
@@ -146,8 +127,6 @@ testthat::test_that(".check_id throws error with non-character id", {
 
 
 testthat::test_that(".check_id throws error with nonexistent id", {
-  # testthat::skip_on_ci()
-  # testthat::skip_on_covr()
   withr::local_package("sf")
   withr::local_options(list(sf_use_s2 = FALSE))
   input_char <- system.file("gpkg/nc.gpkg", package = "sf")
@@ -789,8 +768,6 @@ testthat::test_that(".intersect_extent returns the intersection extent", {
 
 # .check_character ####
 testthat::test_that(".check_character with non-character inputs", {
-  # testthat::skip_on_ci()
-  # testthat::skip_on_covr()
   # test for non-character input
   testthat::expect_message(
     testthat::expect_error(chopin:::.check_character(3L)),
@@ -808,8 +785,6 @@ testthat::test_that(".check_character with non-character inputs", {
 
 
 testthat::test_that(".check_character with character inputs", {
-  # testthat::skip_on_ci()
-  # testthat::skip_on_covr()
   withr::local_package("sf")
   withr::local_package("terra")
   withr::local_options(list(sf_use_s2 = FALSE))
@@ -833,8 +808,6 @@ testthat::test_that(".check_character with character inputs", {
 
 
 testthat::test_that(".check_character with sf objects", {
-  # testthat::skip_on_ci()
-  # testthat::skip_on_covr()
   withr::local_package("sf")
   withr::local_package("stars")
   withr::local_options(list(sf_use_s2 = FALSE))
@@ -869,8 +842,6 @@ testthat::test_that(".check_character with sf objects", {
 })
 
 testthat::test_that(".check_character with Spat* objects", {
-  # testthat::skip_on_ci()
-  # testthat::skip_on_covr()
   withr::local_package("terra")
 
   ncfile <- system.file(package = "sf", "shape/nc.shp")
@@ -968,8 +939,6 @@ testthat::test_that("`[` methods in chopin -- SpatVector-SpatExtent", {
 
 # .check_vector test duplicates: READY TO REMOVE ####
 testthat::test_that(".check_vector -- SpatVector-SpatExtent", {
-  # testthat::skip_on_ci()
-  # testthat::skip_on_covr()
   withr::local_package("sf")
   withr::local_package("terra")
[…]
