### 2024-09-09 a9e6938
Error fix - deprecated function tests
