+++ tests/testthat/test-scale_process.R
@@ -1118,16 +1118,21 @@ testthat::test_that(
 
 
 # par_map_args tests ####
-testthat::test_that("par_map_args works", {
+testthat::test_that("par_convert_f works", {
   example_fun <- function(x, y, z = 1) {
     return(c(x = x, y = y, z = z))
   }
 
   # Example usage of map_args_xy
   testthat::expect_no_error(
-    result <-
-      par_map_args(fun = example_fun,
-                   name_match = list(a = "x", b = "y"),
-                   a = 10, b = 20, z = 5)
+    resultfoo <-
+      par_convert_f(
+        fun = example_fun,
+        arg_map = c(a = "x", b = "y")
+      )
   )
+
+  testthat::expect_true(is.function(resultfoo))
+  result <- resultfoo(a = 10, b = 20, z = 5)
+  testthat::expect_true(is.vector(result))
 })
