### 2024-09-10 9413efe
0.8.2 update - vignette 04: data download part added
