### 2024-09-11 #638 @sigmafelix
We appreciate @robitalec and @Aariq for the detailed comments. Also, we are grateful for the patience of @beatrizmilz, @ldecicco-USGS, and @adamhsparks. The review comments were extremely helpful for redesigning the key `chopin` internal functions to take input class into account to choose a proper workflow to convert or reproject one input object with respect to the other. Data detection and conversion functions were fundamentally redesigned with S4 method dispatch, and these functions always inform the users about what's happening under the hood. Please find our responses attached to each comment. We look forward to getting responses for the updated version.

----
## Table of contents
1. [Responses to the comments of @robitalec](#response-robitalec-comments)
2. [Responses to the comments of @Aariq](#response-aariq-comments)

----

# Responses to the comments of @robitalec <a name="response-robitalec-comments"></a>

My first main and most important comment is that, at the moment, I find {chopin} hard to get started using.

I think the pitch is currently too narrow (also mentioned [here](https://github.com/ropensci/software-review/issues/630)), which might turn away folks who don't identify with "health" or "climate" research. I think there is potential for many users across disciplines who have large spatial data (everything is spatial!) to use {chopin} to perform analyses in more manageable chunks, allowing folks to improve the efficiency of their analyses and extend the potential of their machines.

> We totally agree on this point. Speaking about some background of the package development, we started this project as a part of large initiative for climate and health research in the NIEHS. Our inclination to climate and health and its reflection on the package name came from our motivation for the development. The package name is changed to _Computation of Spatial Data by Hierarchical and Objective Partitioning of Inputs for Parallel Processing_, which encapsulates the major functions for parallel processing of geospatial data.

When I open the README, one of the first things I see is notes on data restrictions and z dimensions and spherical geometry. I think this should come later in a caveats section or FAQ vignette.

> We moved the caveats section to the very last part of README in the project root.

Next, we get to "Basic design" which reads more like an overview of functions but also gets into which packages underpin {chopin} functionality, package versions, and registering parallel workers. I think this is the section that really needs to give users a concise and clear overview of the functions. This can and should be reused in other places to follow the recommended principle of *multiple points of entry*, where users should be anticipated to first encounter {chopin} across every source of information (README, documentation, vignettes) and given a clear introduction.

> Thank you for letting us know about multiple points of entry. Along with the modification of "Basic design" part to provide users with the list of the major functions, we added introductory scripts with example data in README, the first vignette titled "getting started with chopin," and package documentation (i.e., chopin-package.Rd).

- "We suggest two flowcharts to help which function to use for parallel processing below" - note there is only one flowchart visible below and it isn't linked to in this sentence
> The initial version was actually partitioned into two separate flowcharts, but we realized that it could be confusing as users see only one image there. The revised README includes two flowcharts that are clearly separated into two different images. Each image follows a dedicated description.

- `extract_at_poly` missing description
> We superseded `extract_at_*` functions with `extract_at`, where we simplified the arguments to pass the input object into `terra::buffer()` or `sf::st_buffer()`. The revised README clarified the support for both point and polygon inputs.

It seems there are currently three main families of functions exported and number of helper or internal functions:

- `par_*`
    - `par_grid` and others for running functions in parallel
    - `par_make_*` for making grids
- `summarize_*`
- `extract_at_*`

These are the kinds of questions I'm left as an early user with after reading the overview

Before we proceed to the main part of our response, we want to note that the function names are revised for function organization.

- `par_*`
    - `par_grid` and others for running functions in parallel: unchanged.
    - `par_make_*` for making grids: `par_pad_*` functions make balanced (`par_pad_balanced`), regular (`par_pad_grid(..., mode = "regular")`, or postprocessed grids (`par_pad_grid(..., mode = "advanced")`).
- `summarize_*`: unchanged.
- `extract_at_*`: two functions in this family are integrated into `extract_at()`.

- Should `par_make_*` functions be split into a new `grid_*` family of functions?
> `par_make_*()` functions were designed as the primary function for grid generation to subsequently put its output into `par_grid()` for the main parallel processing. For the simpler interface, the revised `par_pad_*()` functions use these functions internally for grid generation, meaning that `par_make_*()` functions remain as internal functions, including `par_pad_grid()` and `par_pad_balanced()`.

- What is the difference to a user between `extract_at_buffer` with a kernel weight and `summary_aw` an area-weighted summary? What is the difference between `par_make_grid` and `par_make_gridset`?
> `extract_at_buffer()` has clear distinction from `summarize_aw()` where the former is for raster-vector overlay whereas `summarize_aw()` is for areal interpolation with vector datasets. `par_make_gridset()` extended `par_make_grid()` to generate the grid polygons from `par_make_grid()` and their padded datasets. The new submission renamed `par_make_gridset()` to `par_pad_grid()` and `par_make_grid()` is now internal.

- If `par_group_grid` is an extension of par_group_balanced for padded grids, should it be renamed something like `par_group_balanced_padded`? Or should `padded = TRUE` be added as an argument?
> We appreciate on this point on the confusion and discrepancy in interface of two functions `par_group_grid()` and `par_make_gridset()` in the previous version. We consider `par_make_grid()` and `par_group_balanced()` as the functions at the same level that furnish its outputs into `par_make_gridset()` and `par_group_balanced()`. These two functions are moved to internal functions and `par_pad_grid()` and `par_pad_balanced()` are put in place to make the grid generation step into two.

- Are regular R users going to find useful functions here or is {chopin} just for HPC users?
> We agree on your point that the previous version narrowed down its usefulness in HPC. Laptops with 4+ CPU threads with sufficient memory capacity (i.e., 4+GB/thread) can take advantage of parallel processing.

From the Description, current version

```
Package: chopin
Title: CHOPIN: Computation for Climate and Health research On Parallelized INfrastructure
Description: It enables users with basic understanding on geospatial data
    and sf and terra functions to parallelize geospatial operations for
    geospatial exposure assessment modeling and covariate computation.
    Parallelization is done by dividing large datasets into sub-regions
    with regular grids and data's own hierarchy. A computation over the
    large number of raster files can be parallelized with a chopin
    function as well.

```

Suggestions:

- I think there's a lot of useful information in the [flowchart](https://github.com/NIEHS/chopin/blob/main/man/figures/README-flowchart-mermaid-1.png), how can we incorporate more obviously?
> Thank you for the comment. We re-generated the flowchart to improve legibility and to provide clear description on the flowchart.

- No need to repeat the package name, capitalized, in the title
> The acronym was removed.

- I think the acronym is neat, but the name stands on it's own. Broadening the scope would help users understand how they can use {chopin} and this starts at the title
> We agree on the point.

- I don't think it's necessary to state the expectations of users in the description
- See more examples [here](https://cran.r-project.org/web/packages/available_packages_by_name.html)

> We wanted to clarify that this package starts from the solid understanding of spatial data and major spatial data packages such as terra and sf.

```
Package: chopin
Title: Parallelize Geospatial Operations Over Sub-Regions, Regular Grids and Lists of Rasters
Description: It enables users to parallelize geospatial operations over
    sub-regions, regular grids and lists of rasters. Parallelization
    is done by dividing large datasets into sub-regions
    with regular grids and data's own hierarchy. A computation over the
    large number of raster files can be parallelized with a chopin
    function as well. (TODO: maybe mention {future}, {exactextractr}, {terra}, {sf})

```

Another reason that I might be finding {chopin} hard to get started using is that it seems to obscure lots of the regular terra/sf steps I would typically do myself as an R user. For example, the `extract_at_buffer` wraps the `exact_extract` function from {exactextractr} but before calling it, it runs `terra::buffer`, manages projections, crops, calculates extents, etc. This is also the case in functions like `reproject_std` and `vect_valid_repair` which reproject/repair objects using functions terra or sf based on the object type.

This also shows itself where the authors manually switch within functions depending on terra or sf input. For example, using `dep_check` and testing `if == 'terra' else if  == 'sf'`. This feels like an opportunity to take advantage of S3 methods. Read more in [Advanced R: S3](https://adv-r.hadley.nz/s3.html#s3-methods) and [R Packages: 11.9](https://r-pkgs.org/dependencies-in-practice.html#imports-and-exports-related-to-s3). An example applied is in {tmap}'s internal function [`tmapShape`](https://github.com/r-tmap/tmap/blob/master/R/tmapShape.R#L61-L174). Obviously this would require a fair amount of refactoring - I just wanted to highlight an alternative approach that might be more flexible to adding new supported data types in the future.

> We appreciate your comment. Our refactoring focused on using base functions to do most of type/class checking functions that used to rely on if-else statement. The part you pointed out was replaced by `inherits()` with multiple class names.

This is definitely an open ended comment - I just wonder where the point is for users where hiding away managing regular steps in a spatial analysis (like projections, extents, clipping, etc) actually results in more confusion than it helps.

> Agreeing on this comment, we added messaging functions into type conversion and reprojection functions, which are now internal, to provide users with explicit information.

My last general comment is - how do you envision {chopin} fitting into the broader [High Performance Computing](https://cran.r-project.org/web/views/HighPerformanceComputing.html) landscape in R? One thing I'm personally really interested is how it might interface with {targets}. Looks like maybe I'm saved from commenting too much on this, as my co reviewer [@Aariq](https://github.com/Aariq) is one of the authors on the new [{geotargets}](https://github.com/njtierney/geotargets) package along with [@njtierney](https://github.com/njtierney) and [@brownag](https://github.com/brownag) (thank you!). I have been eagerly following their progress as a regular {targets} user that has encountered the challenges of including spatial data in {targets} workflows. I wonder how these tools can be brought together or, alternatively, how the authors might communicate their differences and where a user might choose one over the other.

> We use `targets` in a large-scale spatiotemporal modeling project and have the same thought on integrating `chopin` functions and `targets` pipeline functions with minimal syntactical burden. Our conception is to seamlessly integrate `par_pad_*` function output into generic `tar_target()` for branching out a large geospatial computation target. However, such an integration is beyond the scope of this package, so we briefly introduced how to connect `par_pad_*()` outputs to `tar_target()`. A dedicated vignette for efficient use of `chopin` in HPC is repurposed to the broader contexts with multi-core CPUs.

Below: detailed comments following sections in Documentation and Functionality bullets from the top.

### System requirements

Suggestions:

- Consider listing system requirements (eg. netcdf) using recommendations [here](https://github.com/rstudio/r-system-requirements)
- See example from [{magick}](https://github.com/ropensci/magick/blob/master/DESCRIPTION#L20)

### Vignettes

Issues:

It is unclear to a user which vignette is the introductory one. The rOpenSci Packaging Guide suggests "The general vignette should present a series of examples progressing in complexity from basic to advanced usage.". At the moment, this looks most like the "Extracting Weather/Climate Geospatial Data with `chopin`" vignette but the title is fairly specific and it is listed last on the website.

> We added the first vignette on the general tour of `chopin` functions. Per multiple points of entry principle, the similar introductory materials were placed in the package readme.

Consider splitting information about HPC or hardware into its own vignette. Also consider moving system.time benchmarking to a separate vignette? This could also potentially include tips for saving time and memory, and/or hardware recommendations.

> We put a small amount of computational assets as the example data was small. If we want to use a large dataset to provide the actual results from running the process in HPC, it gives another issue of building vignettes as it might take too long time to run R CMD check.

Consider incorporating a section in the introductory vignette with a caveats section

- why is parallelization slower
- comparing single thread vs multi thread calculations
- why we turn off spherical geometry

> We added the caveats into a separate section in the introductory vignette.

Consider revising the vignettes and README to develop bullet points into paragraphs. Some shorter lists to compare options or list steps can be helpful but, generally, I find a full page of bullets hard to process.

> In the revision, most of the bullet points were removed unless using these would be effective to deliver the key messages, for example, separating one function from another.

Images on the website are giving 404 errors

> We believe all image links are fixed. The current webpage correctly displays the figures.

- Generate computational grids

> We changed the vignettes to use pre-built image files. We hope it works now.

Package issues

- Extracting Weather/Climate Geospatial Data with `chopin`
    - `pkgs <- c("chopin", "terra", "stars", "sf", "future", "doFuture", "dplyr", "parallelly", "tigris", "tictoc") invisible(sapply(pkgs, library, character.only = TRUE, quietly = TRUE))`
        - packages
        - consider loading packages in way that most users will be more familiar with eg. `library(chopin)`
> We changed package loading codes from character vectors and `vapply()` to `library()` in all vignettes and examples.

Using :: or both :: and library in vignettes (users are likely more familiar with `library()`)

- Generate computational grids
- Extracting Weather/Climate Geospatial Data with `chopin`

> These parts are revised to use both `library()` and `::`.

Suggestions:

Missing link to external functions / packages

- Generate computational grids
    - {anticlust}
- Good practice of using `chopin` in HPC
    - {future.apply}
- Extracting Weather/Climate Geospatial Data with `chopin`
    - {tigris}
    - {terra}
    - ...
> We added links to functions in external packages.

### Function documentation

Issues:

Missing top-level documentation

- See [`usethis::use_package_doc`](https://usethis.r-lib.org/reference/use_package_doc.html)
- Consider using this space to translate the flowchart into a written overview of the main functions with inputs and outputs

> Description of the flowcharts in GitHub README is available in the package description.

Suggestions:

Missing links to {chopin} functions, external functions, other packages

- anticlust `par_group_balanced`
- sf (throughout)
- terra (throughout)
- `par_group_balanced` from `par_group_grid`, ...

> External packages such as `anticlust` and cross-referencing in package documentation will be beneficial to users, but we find `sf` and `terra` links may not be as beneficial as the previous one as we assume that users are familiar with `sf` and `terra` already.

- Users may benefit from a definition of an "inexhaustive grid"
> We clarified the "inexhaustive grid" by changing the sentence with this phrase into:
> The results will be irregular grids with or without overlapping parts.

Missing documentation of argument defaults

- `clip_ras_ext` nqsegs = 180L
- `extract_at_poly` func = mean
- ...

> We added default values in the function documentation.

Unclear functions

- Given current documentation, it is unclear why/when/where a user use eg. the following function
    - https://niehs.github.io/chopin/reference/par_fallback.html
> We removed this function and the error handling is done in each `par_*` function.

Use of ...

- Consider explicitly demonstrating and describing this functionality in documentation and vignette
    - Eg. `par_grid`
> The vignette, package documentation, and GitHub README include a quick tour of `par_grid`, `par_hierarchy`, and `par_multirasters`. Each function documentation includes examples in the updated version.

### Examples

Issues:

Missing example

- `extract_at` (reuse example see [roxygen2 vignette](https://roxygen2.r-lib.org/articles/reuse.html))

Suggestions:

Examples not run with `if (FALSE)` or dontrun. Consider a smaller toy example?

- `par_grid`
- `par_hierarchy`
- `par_merge_grid`
- `par_multirasters`
- `vect_valid_repair`

> Examples of `par_grid`, `par_hierarchy`, and `par_multirasters` are not skipped in this update. We keep `par_merge_grid` examples in \dontrun{} as it is an internal function. `vect_valid_repair` was removed.

Using :: or both :: and library in examples

- `clip_vec_ext`
- `clip_ras_ext`
- `datamod`
- `dep_check`
- `dep_switch`
- ...

> `clip_vec_ext` and `clip_ras_ext` were removed. Other functions get both `::` and `library()` in the examples.

### Community guidelines

Issues:

There are no contributing guidelines in the README or CONTRIBUTING

> CONTRIBUTING.md is included in the repository root.

### Automated tests

Issues:

There is good test coverage in the package but I am a bit concerned about the number of times the authors use the `expect_no_error` function.

When I think of effective package testing, I think of both clarity for the user reading the tests and in the authors setting up expectations for how the package may fail (anticipating your own errors, warnings or messages). What does the test `expect_no_error` communicate to the user or set up as expectations?

Here are some ideas for expanding or adjusting the tests:

- inline comments in the test scripts to explain to readers/our future selves as authors ([example](https://github.com/tidyverse/ggplot2/blob/main/tests/testthat/test-coord-flip.R))
- narrow the scope of each `test_that()` call and expand on the description
    - from [https://testthat.r-lib.org/reference/test_that.html], "A test encapsulates a series of expectations about a small, self-contained unit of functionality"
[…]
