### 2024-09-13 #638 @Aariq
Haven't had a chance to try out the refactored version, but read through the responses and the some of the new documentation and it looks like the authors did a great job of responding to comments.  It appears the package is much improved!
