### 2024-09-16 #638 @beatrizmilz
@Aariq and @robitalec, Thank you again for the review. The authors of the package have improved the package using the suggestions. Could you indicate, using the [reviewer approval template](https://devguide.ropensci.org/approval2template.html), what is the reviewer's response and decision? Thank you all so much!
