### 2024-09-17 #638 @Aariq
## Reviewer Response

Looks good!  I'll leave the authors with a few final comments they may want to address in future versions.

- When I install or `load_all()` the package, I get the following message printed: `in method for ‘[’ with signature ‘x="SpatVector",i="bbox",j="missing"’: no definition for class “bbox”`.  Something might be wrong with how this method is implemented.
- In the targets vignette, you can make the targets workflow actually executible in a self-contained way with `targets::tar_dir({ targets::tar_script({ <contents of _targets.R goes here> }) targets::tar_make() })`.  It might be a good idea to do that, and to use a real example, not a hypothetical "calc_something()". I spent about 30 min fixing multiple mistakes in that vignette and still can't get the proposed targets pipeline to run.  It seems like it should work in *concept*, but I would need more time to get it working.  If I end up using `chopin` in any `targets` pipelines, I'll can open an issue or make a PR to improve that vignette.
- Along those lines, I see that the climate examples vignette is now "pre-computed" (contains no executable R code).  That's great for CRAN compliance, but it would be good to keep the original .Rmd with executable chunks and an R script used to output the pre-computed vignette around in the `vignettes/` folder.  You can add these files to .Rbuildignore to avoid notes about non-standard files. Having the original .Rmd around is super important if you ever need to make updates to functions or to that vignette (or, e.g., the `tigris` API changes).

#### Final approval (post-review)

Estimated hours spent reviewing: 6hrs (total, both rounds)
