### 2024-09-19 #638 @robitalec
## Reviewer Response

Thank you @sigmafelix for your detailed response to the reviews.

The documentation has been improved, the README and introductory vignettes are much more approachable now. The flowcharts are clear and legible, providing users a clear path to navigating the different functions available in {chopin}.

The CONTRIBUTING doc is currently organized under the .github/ directory. I'm not familiar with putting it in this location, in my opinion having it in the root of the GitHub repository alongside the README, CODE_OF_CONDUCT, etc would help users find it easier.

There seem to be a lot of verbose messages being printed either from {chopin} functions or package start up messages. In the README under [Extract average SRTM elevations by single and multiple threads](https://github.com/NIEHS/chopin?tab=readme-ov-file#extract-average-srtm-elevations-by-single-and-multiple-threads) we see ~ 100 lines printed. I wonder if these could be optionally printed by the user, by default hidden, using a "verbose" argument?

```
#> Input is a character. Attempt to read it with terra::rast...ℹ Your input function at 37189 is dispatched.
#> Input is a character. Attempt to read it with terra::rast...ℹ Your input function at 37173 is dispatched.
#> Input is a character. Attempt to read it with terra::rast...ℹ Your input function at 37011 is dispatched.
#> Input is a character. Attempt to read it with terra::rast...ℹ Your input function at 37045 is dispatched.
#> Input is a character. Attempt to read it with terra::rast...ℹ Your input function at 37125 is dispatched.
#> Input is a character. Attempt to read it with terra::rast...ℹ Your input function at 37067 is dispatched.
```

```
#> Your input function was successfully run at CGRIDID: 1
#> Your input function was successfully run at CGRIDID: 2
#> Your input function was successfully run at CGRIDID: 3
#> Your input function was successfully run at CGRIDID: 4
#> Your input function was successfully run at CGRIDID: 5
#> Your input function was successfully run at CGRIDID: 6
#> Your input function was successfully run at CGRIDID: 7
#> Your input function was successfully run at CGRIDID: 8"
```

The package start up messages are particularly noticeable in the vignette [Generate computation grids](https://niehs.github.io/chopin/articles/v03_par_pad_grid.html). It could be helpful to have some introductory text at least before the full page of library() calls and accompanying start up messages. Or these messages could be suppressed for the vignette?

The following run successfully on my local machine (as of 2024-09-19):

- README.Rmd
- v01_start.Rmd
- v02_good_practice.Rmd (no code chunks to evaluate)
- v03_par_pad_grid.Rmd
- `devtools::run_examples()` (where I also encountered the method message @Aariq noted)

Code chunks not evaluated, output only a demonstration:

- v04_climate_examples.Rmd (see @Aariq's comment)

Setup code chunks evaluated, {targets} steps not evaluated:

- v05_targets.Rmd

There is another option for demonstrating how {chopin} can interface with {targets}. I used this approach for demonstrated how our package {spatsoc} can be used with {targets} [here](https://github.com/robitalec/targets-spatsoc-networks). This repository essentially takes our "getting started" vignette, and translates it to a {targets} workflow. We didn't make a standalone vignette for it (probably a good idea), but we highlighted it in our [README](https://github.com/ropensci/spatsoc/blob/main/README.Rmd#L71-L76).

This might be an easier way to demonstrate how {chopin} and {targets} can be used together, as a user only has to install packages, clone the template workflow and run `tar_make()`.

#### Final approval (post-review)

Estimated hours spent reviewing: 9.5 hours
