### 2024-09-23 #638 @beatrizmilz
Thank you both @robitalec and @Aariq for the second review and suggestions.

@sigmafelix Congratulations on the package approval! @robitalec and @Aariq gave more suggestions, and I would like to know if you plan to address those, and if yes, if you prefer to do it before or after transferring the repository to the account of rOpenSci.

The following comment from @Aariq seems really important to be addressed:
> When I install or load_all() the package, I get the following message printed: in method for ‘[’ with signature ‘x="SpatVector",i="bbox",j="missing"’: no definition for class “bbox”. Something might be wrong with how this method is implemented.
