### 2024-09-24 #638 @ropensci-review-bot
Approved! Thanks @sigmafelix for submitting and @robitalec, @Aariq for your reviews! :grin:
