### 2024-09-25 #638 @sigmafelix
@beatrizmilz Thank you very much for approving our package! We are excited to see that `chopin` is included in ROpenSci packages. We sincerely appreciate the reviewers @Aariq and @robitalec for extremely helpful comments for improving `chopin`.

For the recent comment, we are aware of the `bbox` class detection error in S4 context as being pointed out by @Aariq. I speculate that this error might come from different mechanisms of setting a class to an object between S3 and S4. I think I will consult `sf` maintainers about this issue further before I make changes.

```
library(sf)
dd <- st_bbox(c(xmin = 16.1, xmax = 16.6, ymax = 48.6, ymin = 47.9), crs = st_crs(4326))
class(dd)
[1] "bbox"

setGeneric("detect_bbox", function(x){})
[1] "detect_bbox"
setMethod("detect_bbox", signature = c(x = "bbox"), function(x){class(x)})
# in method for ‘detect_bbox’ with signature ‘x="bbox"’: no definition for class “bbox”
```
