### 2024-09-28 dc7dc3b
0.8.3 - Per rOpenSci requirements, this push reflected the following:   - Edited DESCRIPTION for mentioning reviewers   - Rebuilt README.md   - Rebuilt vignettes (#103)   - Rebuilt codemeta   - Removed code of conduct file
