+++ README.md
@@ -52,9 +52,14 @@ multiple raster file paths into `par_multirasters()`. **Finally,** users
 run `par_*()` function with the configurations set above to compute
 spatial variables from input data in parallel:
 
-    - `par_grid`: parallelize over artificial grid polygons that are generated from the maximum extent of inputs. `par_pad_grid` is used to generate the grid polygons before running this function.
-    - `par_hierarchy`: parallelize over hierarchy coded in identifier fields (for example, census blocks in each county in the US)
-    - `par_multirasters`: parallelize over multiple raster files
+- `par_grid`: parallelize over artificial grid polygons that are
+  generated from the maximum extent of inputs. `par_pad_grid` is used to
+  generate the grid polygons before running this function.
+
+- `par_hierarchy`: parallelize over hierarchy coded in identifier fields
+  (for example, census blocks in each county in the US)
+
+- `par_multirasters`: parallelize over multiple raster files
 
 For grid partitioning, the entire study area will be divided into partly
 overlapped grids. We suggest two flowcharts to help which function to
@@ -70,9 +75,14 @@ classes for spatial data. Raster-vector overlay is done with
 `exactextractr`. Three helper functions encapsulate multiple geospatial
 data calculation steps over multiple CPU threads.
 
-    - `extract_at`: extract raster values with point buffers or polygons with or without kernel weights
-    - `summarize_sedc`: calculate sums of [exponentially decaying contributions](https://mserre.sph.unc.edu/BMElab_web/SEDCtutorial/index.html)
-    - `summarize_aw`: area-weighted covariates based on target and reference polygons
+- `extract_at`: extract raster values with point buffers or polygons
+  with or without kernel weights
+
+- `summarize_sedc`: calculate sums of [exponentially decaying
+  contributions](https://mserre.sph.unc.edu/BMElab_web/SEDCtutorial/index.html)
+
+- `summarize_aw`: area-weighted covariates based on target and reference
+  polygons
 
 ### Function selection guide
 
@@ -83,33 +93,32 @@ users with large vector data.
 
 In **raster-oriented selection**, we suggest four factors to consider:
 
-  - Number of raster files: for multiple files, `par_multirasters` is
-    recommended. When there are multiple rasters that share the same
-    extent and resolution, consider stacking the rasters into multilayer
-    SpatRaster object by calling `terra::rast(filenames)`.
-  - Raster resolution: We suggest 100 meters as a threshold. Rasters
-    with resolution coarser than 100 meters and a few layers would be
-    better for the direct call of `exactextractr::exact_extract()`.
-  - Raster extent: Using `SpatRaster` in
-    `exactextractr::exact_extract()` is often minimally affected by the
-    raster extent.
-  - Memory size: `max_cells_in_memory` argument value of
-    `exactextractr::exact_extract()`, raster resolution, and the number
-    of layers in `SpatRaster` are multiplicatively related to the memory
[…]
