+++ README.Rmd
@@ -43,7 +43,9 @@ future::plan(future.mirai::mirai_multisession, workers = 4L)
 __Second,__ users choose the proper data parallelization configuration by creating a grid partition of the processing extent, defining the field name with values that are hierarchically coded, or entering multiple raster file paths into `par_multirasters()`. __Finally,__ users run `par_*()` function with the configurations set above to compute spatial variables from input data in parallel:
 
 - `par_grid`: parallelize over artificial grid polygons that are generated from the maximum extent of inputs. `par_pad_grid` is used to generate the grid polygons before running this function.
+
 - `par_hierarchy`: parallelize over hierarchy coded in identifier fields (for example, census blocks in each county in the US)
+
 - `par_multirasters`: parallelize over multiple raster files
 
 
@@ -53,7 +55,9 @@ For grid partitioning, the entire study area will be divided into partly overlap
 Processing functions accept [terra](https://github.com/rspatial/terra)/[sf](https://github.com/r-spatial/sf) classes for spatial data. Raster-vector overlay is done with `exactextractr`. Three helper functions encapsulate multiple geospatial data calculation steps over multiple CPU threads.
 
 - `extract_at`: extract raster values with point buffers or polygons with or without kernel weights
+
 - `summarize_sedc`: calculate sums of [exponentially decaying contributions](https://mserre.sph.unc.edu/BMElab_web/SEDCtutorial/index.html)
+
 - `summarize_aw`: area-weighted covariates based on target and reference polygons
 
 
@@ -87,9 +91,14 @@ For **vector-oriented selection**, we suggest three factors to consider:
 
 ```r
 rlang::check_installed("remotes")
-remotes::install_github("NIEHS/chopin")
+remotes::install_github("ropensci/chopin")
 ```
 
+or you can also set `repos` in `install.packages()` as ROpenSci repository:
+
+```r
+install.packages("chopin", repos = "https://ropensci.r-universe.dev")
+```
 
 ## Examples
 Examples will navigate `par_grid`, `par_hierarchy`, and `par_multirasters` functions in `chopin` to parallelize geospatial operations.
@@ -433,9 +442,6 @@ Parallelization may underperform when the datasets are too small to take advanta
 
 
 ### Notes on data restrictions
-`chopin` works best with **two-dimensional** (**planar**) geometries. Users should disable `s2` spherical geometry mode in `sf` by setting. Running any `chopin` functions at spherical or three-dimensional (e.g., including M/Z dimensions) geometries may produce incorrect or unexpected results.
+`chopin` works best with **two-dimensional** (**planar**) geometries. Users should disable `s2` spherical geometry mode in `sf` by setting `sf::sf_use_s2(FALSE)`. Running any `chopin` functions at spherical or three-dimensional (e.g., including M/Z dimensions) geometries may produce incorrect or unexpected results.
 
-```r
-sf::sf_use_s2(FALSE)
-```
 
