### 2024-10-03 #638 @sigmafelix
@beatrizmilz Could you invite @kyle-messier as the co-author of the package? He has the administrator privilege of the repository to finalize the transfer of the repository to ropensci. Thank you!
