### 2024-10-03 #638 @beatrizmilz
> @beatrizmilz Could you invite @kyle-messier as the co-author of the package? He has the administrator privilege of the repository to finalize the transfer of the repository to ropensci. Thank you!

Hi @sigmafelix ! I saw your message this morning but was not sure how to proceed. I asked other editors and I'll write back soon, ok?
