### 2024-10-05 60fb57c
0.8.3 for ropensci - All URLs are changed to ropensci - Rebuilt README and vignettes - Rebuilt codemeta.json
