+++ tools/litr-source/chopin_rmarkdown_litr.rmd
@@ -42,8 +42,8 @@ usethis::create_package(
     LazyData = "true",
     LazyDataCompression = "xz",
     Depends = "R (>= 4.1)",
-    URL = "https://github.com/NIEHS/chopin",
-    BugReports = "https://github.com/NIEHS/chopin/issues"
+    URL = "https://github.com/ropensci/chopin",
+    BugReports = "https://github.com/ropensci/chopin/issues"
   )
 )
 #usethis::use_build_ignore("rds$", escape = FALSE)
