+++ README.md
@@ -3,8 +3,8 @@
 
 <!-- badges: start -->
 
-[![cov](https://NIEHS.github.io/chopin/badges/coverage.svg)](https://github.com/NIEHS/chopin/actions)
-[![R-CMD-check](https://github.com/NIEHS/chopin/actions/workflows/check-standard.yaml/badge.svg)](https://github.com/NIEHS/chopin/actions/workflows/check-standard.yaml)
+[![cov](https://NIEHS.github.io/chopin/badges/coverage.svg)](https://github.com/ropensci/chopin/actions)
+[![R-CMD-check](https://github.com/ropensci/chopin/actions/workflows/check-standard.yaml/badge.svg)](https://github.com/ropensci/chopin/actions/workflows/check-standard.yaml)
 [![Status at rOpenSci Software Peer
 Review](https://badges.ropensci.org/638_status.svg)](https://github.com/ropensci/software-review/issues/638)
 [![Lifecycle:
@@ -52,14 +52,14 @@ multiple raster file paths into `par_multirasters()`. **Finally,** users
 run `par_*()` function with the configurations set above to compute
 spatial variables from input data in parallel:
 
-- `par_grid`: parallelize over artificial grid polygons that are
-  generated from the maximum extent of inputs. `par_pad_grid` is used to
-  generate the grid polygons before running this function.
+  - `par_grid`: parallelize over artificial grid polygons that are
+    generated from the maximum extent of inputs. `par_pad_grid` is used
+    to generate the grid polygons before running this function.
 
-- `par_hierarchy`: parallelize over hierarchy coded in identifier fields
-  (for example, census blocks in each county in the US)
+  - `par_hierarchy`: parallelize over hierarchy coded in identifier
+    fields (for example, census blocks in each county in the US)
 
-- `par_multirasters`: parallelize over multiple raster files
+  - `par_multirasters`: parallelize over multiple raster files
 
 For grid partitioning, the entire study area will be divided into partly
 overlapped grids. We suggest two flowcharts to help which function to
@@ -75,14 +75,14 @@ classes for spatial data. Raster-vector overlay is done with
 `exactextractr`. Three helper functions encapsulate multiple geospatial
 data calculation steps over multiple CPU threads.
 
-- `extract_at`: extract raster values with point buffers or polygons
-  with or without kernel weights
+  - `extract_at`: extract raster values with point buffers or polygons
+    with or without kernel weights
 
-- `summarize_sedc`: calculate sums of [exponentially decaying
-  contributions](https://mserre.sph.unc.edu/BMElab_web/SEDCtutorial/index.html)
+  - `summarize_sedc`: calculate sums of [exponentially decaying
+    contributions](https://mserre.sph.unc.edu/BMElab_web/SEDCtutorial/index.html)
 
-- `summarize_aw`: area-weighted covariates based on target and reference
-  polygons
+  - `summarize_aw`: area-weighted covariates based on target and
+    reference polygons
 
 ### Function selection guide
 
@@ -93,32 +93,33 @@ users with large vector data.
 
 In **raster-oriented selection**, we suggest four factors to consider:
 
[…]
