### 2024-10-10 #638 @maelle
@sigmafelix I've now invited @kyle-messier.

Sorry for the lack of automation, it's tracked (https://github.com/ropensci/dev_guide/issues/466)
