### 2024-10-10 #638 @sigmafelix
@maelle Thank you for figuring it out for us. I'm glad to hear that the automated invitation will be available in the future.
