### 2024-10-10 #638 @sigmafelix
@maelle Could you check if the invitation was sent or processed? I couldn't find @kyle-messier on the member list of ropensci/chopin team.
