### 2024-10-10 #638 @kyle-messier
@maelle @beatrizmilz I am still not able to transfer the repository. It could be due to setting on our NIEHS GitHub enterprise. Before I ping our enterprise managers, could you confirm whether I should be able to see github.com/ropensci/chopin ? It is currently 404 not found for me.
