### 2024-10-10 #638 @maelle
@kyle-messier you should have received an invitation to https://github.com/orgs/ropensci/teams/chopin, I still see you as pending member. Can you check your email inbox?

On another hand it might be the case that you have to first transfer the repo to your personal account then to ropensci. Repo transfers are sometimes clunky in this way. In any case, official docs https://docs.github.com/en/repositories/creating-and-managing-repositories/transferring-a-repository#transferring-a-repository-owned-by-your-organization
